+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -8,11 +8,16 @@ that this addresses, if any.
 Some details that should be in this section include:
 
 - Why this change was necessary
-- What alternative approaches were considered and why the current approach was chosen
+- What alternative solutions were considered and why the current solution was chosen
 - What tests and documentation have been added/updated
 - What do users and developers need to know about this change
 
-### PR checklist (delete when all critera are met)
+Note that this entire PR description field will be used as the commit message upon
+merge, so please keep it updated along with the PR. Secondary discussions, such as
+intermediate testing and bug statuses that do not affect the final PR, should be in the
+PR comments.
+
+### PR checklist (delete when all criteria are met)
 
 - [ ] I have read the contributing guide `CONTRIBUTING.md`.
 - [ ] I have added the tests to cover my changes.
