### 2023-03-15 6c37277
Fix handling of line feed.
Sphinx parser doesn't support line break during the type hint. Long type hint is not shown properly in the API doc. New logic is added to the Sphinx extension to remove these line feeds.
