+++ qiskit_experiments/library/quantum_volume/qv_analysis.py
@@ -48,7 +48,7 @@ class QuantumVolumeAnalysis(BaseAnalysis):
 
         Analysis Options:
             plot (bool): Set ``True`` to create figure for fit result.
-            ax(AxesSubplot): Optional. A matplotlib axis object to draw.
+            ax (AxesSubplot): Optional. A matplotlib axis object to draw.
         """
         options = super()._default_options()
         options.plot = True
