### 2023-03-15 7d79339
changed 1Q from integer to clifford
the default integer method wasn't working with the later `Clifford()`
invocation.
