+++ qiskit_experiments/library/randomized_benchmarking/clifford_utils.py
@@ -300,6 +300,22 @@ class CliffordUtils:
             num //= k
         return res
 
+    @staticmethod
+    def compute_target_bitstring(circuit: QuantumCircuit) -> str:
+        """For a Clifford circuit C, compute C|0>.
+        Args:
+            circuit: A Clifford QuantumCircuit.
+        Returns:
+            Target bitstring.
+        """
+        # convert circuit to Boolean phase vector of stabilizer table
+        phase_vector = Clifford(circuit).table.phase
+        n = circuit.num_qubits
+
+        # target string has a 1 for each True in the stabilizer half of the phase vector
+        target = "".join(["1" if phase else "0" for phase in phase_vector[n:][::-1]])
+        return target
+
 
 # Constant mapping from 1Q single Clifford gate to 1Q Clifford numerical identifier.
 # This table must be generated using `data.generate_clifford_data.gen_cliff_single_1q_gate_map`, or,
@@ -586,20 +602,3 @@ def _layer_indices_from_num(num: Integral) -> Tuple[Integral, Integral, Integral
     idx1 = num % _NUM_LAYER_1
     idx0 = num // _NUM_LAYER_1
     return idx0, idx1, idx2
-
-
-def compute_target_bitstring(self, circuit: QuantumCircuit) -> str:
-    """For a Clifford circuit C, compute C|0>.
-    Args:
-        circuit: A Clifford QuantumCircuit
-    Returns:
-        Target bit string
-    """
-
-    # convert circuit to Boolean phase vector of stabilizer table
-    phase_vector = Clifford(circuit).table.phase
-    n = circuit.num_qubits
-
-    # target string has a 1 for each True in the stabilizer half of the phase vector
-    target = "".join(["1" if phase else "0" for phase in phase_vector[n:][::-1]])
-    return target
