+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -29,7 +29,7 @@ from qiskit.transpiler.basepasses import TransformationPass
 
 from .rb_experiment import StandardRB
 from .mirror_rb_analysis import MirrorRBAnalysis
-from .clifford_utils import CliffordUtils
+from .clifford_utils import CliffordUtils, _decompose_clifford_ops
 from .sampling_utils import MirrorRBSampler, EdgeGrabSampler
 
 SequenceElementType = Union[Clifford, Integral, QuantumCircuit]
@@ -173,20 +173,11 @@ class MirrorRB(StandardRB):
         """Return a list of Mirror RB circuits.
 
         Returns:
-            A list of :class:`QuantumCircuit`s.
+            A list of :class:`QuantumCircuit`.
         """
         sequences = self._sample_sequences()
         circuits = self._sequences_to_circuits(sequences)
 
-        for circ, seq in zip(circuits, sequences):
-            circ.metadata = {
-                "experiment_type": self._type,
-                "xval": len(seq),
-                "group": "Clifford",
-                "physical_qubits": self.physical_qubits,
-                "target": self._clifford_utils.compute_target_bitstring(circ),
-                "inverting_pauli_layer": self._inverting_pauli_layer,
-            }
         return circuits
 
     def _sample_sequences(self) -> List[Sequence[SequenceElementType]]:
@@ -201,7 +192,8 @@ class MirrorRB(StandardRB):
             QiskitError: If no backend is provided.
 
         Returns:
-            A list of mirror RB sequences. Each element is a full circuit layer.
+            A list of mirror RB sequences. Each element is a list of layers with length
+            matching the corresponding element in ``lengths``.
         """
         if not self._backend:
             raise QiskitError("A backend must be provided.")
@@ -216,22 +208,24 @@ class MirrorRB(StandardRB):
             if edge[0] in self.physical_qubits and edge[1] in self.physical_qubits:
                 experiment_coupling_map.append(edge)
 
+        # adjust the density based on whether the pauli layers are in
+        if self.experiment_options.pauli_randomize:
+            adjusted_2q_density = self.experiment_options.two_qubit_gate_density * 2
+        else:
+            adjusted_2q_density = self.experiment_options.two_qubit_gate_density
+
         sequences = []
 
         for _ in range(self.experiment_options.num_samples):
             for length in self.experiment_options.lengths:
-                print("length ", length)
-                elements = []
+
                 # Sample Clifford layer elements for first half of mirror circuit
-                # TODO need to adjust density if the paulis aren't in
-                elements.append(
-                    self._distribution(
-                        self.num_qubits,
-                        self.experiment_options.two_qubit_gate_density,
-                        self.coupling_map,
-                        length // 2,
-                        seed=self.experiment_options.seed,
-                    )[0]
+                elements = self._distribution(
+                    self.num_qubits,
+                    adjusted_2q_density,
+                    self.coupling_map,
+                    length // 2,
+                    seed=self.experiment_options.seed,
                 )
 
                 # Append inverses of Clifford elements to second half of circuit
@@ -257,29 +251,35 @@ class MirrorRB(StandardRB):
             #     ) // pauli_scale
             #     circuit.metadata["mirror"] = True
             # circuits += mirror_circuits
-            sequences += [elements]
+            sequences.append(elements)
         return sequences
+
+    def _adjoint_clifford(self, op: SequenceElementType) -> SequenceElementType:
+        if isinstance(op, QuantumCircuit):
+            return Clifford.from_circuit(op).adjoint()
+        return op.adjoint()
 
     def _sequences_to_circuits(
         self, sequences: List[Sequence[SequenceElementType]]
[…]
