+++ releasenotes/notes/mirror-rb-ec4d695a9a923971.yaml
@@ -1,7 +1,6 @@
 ---
 features:
   - |
-    A new experiment class :class:`.MirrorRB` is 
-    introduced. This class implements mirror randomized benchmarking, a version 
-    of RB that uses mirror circuits. It is more scalable than other RB protocols and
-    can consequently be used to detect crosstalk errors.
+    A new experiment class :class:`.MirrorRB` is introduced. This class implements
+    mirror randomized benchmarking, a variant of randomized benchmarking that measures
+    the fidelity of user-defined ensembles of randomized mirror circuits.
