### 2023-03-16 e66a58d
fix rng and force default transpile
