### 2023-03-16 735cf21
add more config to `lint/conf.py`
linux build is failing on the minimal conf so I'm adding more lines back
in. hard to debug locally because the mac build has a different set of
warnings.
