### 2023-03-16 3e18953
rename benchmarking experiments to verification
this syncs up the names on the manuals page and the API experiments
page
