+++ docs/conf.py
@@ -94,11 +94,11 @@ exclude_patterns = ["_build", "**.ipynb_checkpoints"]
 # These should ideally be automatically generated using a custom macro to specify
 # chosen cells for thumbnails, like the nbsphinx-gallery tag
 nbsphinx_thumbnails = {
-    "manuals/benchmarking/quantum_volume": "_images/quantum_volume_2_0.png",
+    "manuals/verification/quantum_volume": "_images/quantum_volume_2_0.png",
     "manuals/measurement/readout_mitigation": "_images/readout_mitigation_4_0.png",
-    "manuals/benchmarking/randomized_benchmarking": "_images/randomized_benchmarking_3_1.png",
+    "manuals/verification/randomized_benchmarking": "_images/randomized_benchmarking_3_1.png",
     "manuals/measurement/restless_measurements": "_images/restless_shots.png",
-    "manuals/benchmarking/state_tomography": "_images/state_tomography_3_0.png",
+    "manuals/verification/state_tomography": "_images/state_tomography_3_0.png",
     "manuals/characterization/t1": "_images/t1_0_0.png",
     "manuals/characterization/t2ramsey": "_images/t2ramsey_4_0.png",
     "manuals/characterization/tphi": "_images/tphi_5_1.png",
