### 2023-03-17 4d8ccdd
Change default transpile option to optimization_level=1 for RBs
