+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -26,10 +26,16 @@ from qiskit.quantum_info.operators import Pauli
 from qiskit.providers.backend import Backend
 from qiskit.providers.options import Options
 from qiskit.transpiler.basepasses import TransformationPass
+from qiskit_experiments.warnings import deprecate_arguments
 
 from .rb_experiment import StandardRB
 from .mirror_rb_analysis import MirrorRBAnalysis
-from .clifford_utils import CliffordUtils, _decompose_clifford_ops, _clifford_1q_int_to_instruction
+from .clifford_utils import (
+    CliffordUtils,
+    _decompose_clifford_ops,
+    _clifford_1q_int_to_instruction,
+    compute_target_bitstring,
+)
 from .sampling_utils import MirrorRBSampler, EdgeGrabSampler
 
 SequenceElementType = Union[Clifford, Integral, QuantumCircuit]
@@ -63,6 +69,7 @@ class MirrorRB(StandardRB):
 
     """
 
+    @deprecate_arguments({"qubits": "physical_qubits"}, "0.5")
     def __init__(
         self,
         physical_qubits: Sequence[int],
@@ -105,13 +112,12 @@ class MirrorRB(StandardRB):
         Raises:
             QiskitError: if an odd length or a negative two qubit gate density is provided
         """
-        # All lengths must be even
+
         if not all(length % 2 == 0 for length in lengths):
             raise QiskitError("All lengths must be even")
 
-        # Two-qubit density must be non-negative
-        if two_qubit_gate_density < 0:
-            raise QiskitError("Two-qubit gate density must be non-negative")
+        if two_qubit_gate_density < 0 or two_qubit_gate_density > 1:
+            raise QiskitError("Two-qubit gate density must be between 0 and 1.")
 
         super().__init__(
             physical_qubits,
@@ -122,7 +128,6 @@ class MirrorRB(StandardRB):
             full_sampling=full_sampling,
         )
 
-        self._clifford_utils = CliffordUtils()
         self._distribution = distribution()
 
         self.set_experiment_options(
@@ -132,8 +137,6 @@ class MirrorRB(StandardRB):
             inverting_pauli_layer=inverting_pauli_layer,
             two_qubit_gate_density=two_qubit_gate_density,
         )
-
-        # self.set_transpile_options(optimization_level=1)
 
         self.analysis = MirrorRBAnalysis()
 
@@ -151,12 +154,7 @@ class MirrorRB(StandardRB):
             two_qubit_gate_density (float): Expected proportion of two-qubit gates in
                 the mirror circuit layers (not counting Clifford or Pauli layers at the
                 start and end).
-            lengths (List[int]): A list of RB sequences lengths.
             num_samples (int): Number of samples to generate for each sequence length.
-            seed (None or int or SeedSequence or BitGenerator or Generator): A seed
-                used to initialize ``numpy.random.default_rng`` when generating circuits.
-                The ``default_rng`` will be initialized with this seed value everytime
-                :meth:`circuits` is called.
         """
         options = super()._default_experiment_options()
         options.update_options(
@@ -164,10 +162,7 @@ class MirrorRB(StandardRB):
             pauli_randomize=True,
             two_qubit_gate_density=0.2,
             distribution=None,
-            num_samples=None,
-            seed=None,
             inverting_pauli_layer=False,
-            full_sampling=None,
         )
 
         return options
@@ -204,11 +199,11 @@ class MirrorRB(StandardRB):
 
         # Coupling map is full connectivity by default. If backend has a coupling map,
         # get backend coupling map and create coupling map for physical qubits
-        self.coupling_map = list(permutations(range(max(self.physical_qubits) + 1), 2))
[…]
