+++ test/library/randomized_benchmarking/test_randomized_benchmarking.py
@@ -31,6 +31,9 @@ from qiskit.transpiler import Layout, PassManager, CouplingMap
 from qiskit_experiments.database_service.exceptions import ExperimentEntryNotFound
 from qiskit_experiments.framework.composite import ParallelExperiment
 from qiskit_experiments.library import randomized_benchmarking as rb
+from qiskit_experiments.library.randomized_benchmarking.clifford_utils import (
+    compute_target_bitstring,
+)
 
 
 class RBTestMixin:
@@ -827,7 +830,6 @@ class TestRunInterleavedRB(RBRunTestCase):
         self.assertRoundTripPickle(expdata, check_func=self.experiment_data_equiv)
 
 
-
 class NonlocalCXDepError(TransformationPass):
     """Transpiler pass for simulating nonlocal errors in a quantum device"""
 
@@ -927,36 +929,18 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
             "optimization_level": 1,
         }
 
-    def test_custom_distribution(self):
-        """Test providing a custom distribution."""
-        qubits = (
-            0,
-            1,
-            2,
-        )
-        exp = rb.MirrorRB(
-            physical_qubits=qubits,
-            distribution=rb.RandomEdgeGrabDistribution,
-            two_qubit_gate_density=0.5,
-            lengths=list(range(2, 300, 20)),
-            seed=124,
-            backend=self.backend,
-            num_samples=30,
-        )
-        # test that feeding the circuit from edge grab into discrete yields the same answer(?)
-
     def test_return_same_circuit(self):
         """Test if setting the same seed returns the same circuits."""
         lengths = [10, 20]
         exp1 = rb.MirrorRB(
-            qubits=(0, 1),
+            physical_qubits=(0, 1),
             lengths=lengths,
             seed=123,
             backend=self.backend,
         )
 
         exp2 = rb.MirrorRB(
-            qubits=(0, 1),
+            physical_qubits=(0, 1),
             lengths=lengths,
             seed=123,
             backend=self.backend,
@@ -971,7 +955,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
     def test_full_sampling(self):
         """Test if full sampling generates different circuits."""
         exp1 = rb.MirrorRB(
-            qubits=(0, 1),
+            physical_qubits=(0, 1),
             lengths=[10, 20],
             seed=123,
             backend=self.backend,
@@ -980,7 +964,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         )
 
         exp2 = rb.MirrorRB(
-            qubits=(0, 1),
+            physical_qubits=(0, 1),
             lengths=[10, 20],
             seed=123,
             backend=self.backend,
@@ -997,26 +981,11 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         # top of previous length
         self.assertNotEqual(circs1[1].decompose(), circs2[1].decompose())
 
-    def test_target_bitstring(self):
-        """Test if correct target bitstring is returned."""
-        qc = QuantumCircuit(9)
-        qc.z(0)
-        qc.y(1)
-        qc.y(2)
-        qc.z(3)
-        qc.y(4)
-        qc.x(7)
-        qc.y(8)
-        exp = rb.MirrorRB(qubits=[0], lengths=[2], backend=self.backend)
-        expected_tb = exp._clifford_utils.compute_target_bitstring(qc)
-        actual_tb = "110010110"
-        self.assertEqual(expected_tb, actual_tb)
-
     def test_zero_2q_gate_density(self):
[…]

[…]
