+++ test/library/randomized_benchmarking/test_randomized_benchmarking.py
@@ -24,10 +24,13 @@ from qiskit.exceptions import QiskitError
 from qiskit.providers.fake_provider import FakeManila, FakeManilaV2, FakeWashington, FakeParis
 from qiskit.pulse import Schedule, InstructionScheduleMap
 from qiskit.quantum_info import Operator
-from qiskit_aer import AerSimulator
-from qiskit_aer.noise import NoiseModel, depolarizing_error
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.transpiler import Layout, PassManager, CouplingMap
+from qiskit.transpiler.exceptions import TranspilerError
+
+from qiskit_aer import AerSimulator
+from qiskit_aer.noise import NoiseModel, depolarizing_error
+
 from qiskit_experiments.database_service.exceptions import ExperimentEntryNotFound
 from qiskit_experiments.framework.composite import ParallelExperiment
 from qiskit_experiments.library import randomized_benchmarking as rb
@@ -545,6 +548,7 @@ class TestRunStandardRB(RBRunTestCase):
         epc = expdata.analysis_results("EPC")
 
         epc_expected = 1 - (1 - 1 / 2 * self.p1q) ** 1.0
+        print(epc.value.n)
         self.assertAlmostEqual(epc.value.n, epc_expected, delta=0.1 * epc_expected)
 
     def test_two_qubit(self):
@@ -1099,19 +1103,59 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         self.assertEqual(analysis.config(), loaded.config())
 
 
+@ddt
 class TestRunMirrorRB(RBRunTestCase):
+    """Class for testing execution of mirror RB experiments."""
+
+    def setUp(self):
+        """Setup the tests."""
+        super().setUp()
+
+        # depolarizing error
+        self.p1q = 0.02
+        self.p2q = 0.10
+        self.pvz = 0.0
+
+        # basis gates
+        self.basis_gates = ["sx", "rz", "cx", "id"]
+
+        # setup noise model
+        sx_error = depolarizing_error(self.p1q, 1)
+        rz_error = depolarizing_error(self.pvz, 1)
+        cx_error = depolarizing_error(self.p2q, 2)
+
+        noise_model = NoiseModel()
+        noise_model.add_all_qubit_quantum_error(sx_error, "sx")
+        noise_model.add_all_qubit_quantum_error(rz_error, "rz")
+        noise_model.add_all_qubit_quantum_error(cx_error, "cx")
+
+        self.noise_model = noise_model
+        self.basis_gates = noise_model.basis_gates
+
+        # Need level1 for consecutive gate cancellation for reference EPC value calculation
+        self.transpiler_options = {
+            "basis_gates": self.basis_gates,
+            "optimization_level": 1,
+        }
+
+        # Aer simulator
+        self.backend = AerSimulator(
+            noise_model=noise_model,
+            seed_simulator=123,
+            coupling_map=AerSimulator.from_backend(FakeParis()).configuration().coupling_map,
+        )
+
     def test_single_qubit(self):
         """Test single qubit mirror RB."""
         exp = rb.MirrorRB(
             physical_qubits=(0,),
-            lengths=list(range(2, 300, 20)),
+            lengths=list(range(2, 300, 40)),
             seed=124,
             backend=self.backend,
-            num_samples=30,
+            num_samples=20,
         )
-        # exp.analysis.set_options(gate_error_ratio=None)
+        exp.analysis.set_options(gate_error_ratio=None)
         exp.set_transpile_options(**self.transpiler_options)
-        self.assertAllIdentity(exp.circuits())
 
[…]
