### 2023-03-18 2e62141
update curve analysis tutorial and undo lint
