+++ tox.ini
@@ -72,13 +72,6 @@ setenv =
 commands =
   sphinx-build -T --keep-going -b html {posargs} docs/ docs/_build/html
 
-[testenv:docs-lint]
-passenv = EXPERIMENTS_DEV_DOCS
-setenv = 
-  QISKIT_DOCS_SKIP_EXECUTE = 1
-commands =
-  sphinx-build -c docs/lint -n -T --keep-going -b linkcheck {posargs} docs/ docs/_build/html
-
 [pycodestyle]
 max-line-length = 100
 
