+++ docs/manuals/index.rst
@@ -4,7 +4,7 @@ Experiment Manuals
 These experiment manuals are in-depth dives into individual experiments, their
 operational principles, and how to run them in Qiskit Experiments.
 
-.. _verification:
+.. _verification manuals:
 
 Verification Experiments
 ------------------------
@@ -18,7 +18,7 @@ executed.
 
     verification/*
 
-.. _qubit characterization:
+.. _qubit characterization manuals:
 
 Qubit Characterization Experiments
 ----------------------------------
@@ -30,7 +30,7 @@ These experiment measure specific properties of a qubit.
 
     characterization/*
 
-.. _measurement-related:
+.. _measurement-related manuals:
 
 Measurement-Related Experiments
 -------------------------------
