+++ docs/index.rst
@@ -68,7 +68,7 @@ We've divided up the documentation into four sections with different purposes:
         background, principle, and how to run them in Qiskit Experiments. You'll find in
         these manuals:
         
-        * How to analyze 1- and 2-qubit errors in :doc:`randomized benchmarking </manuals/benchmarking/randomized_benchmarking>`
+        * How to analyze 1- and 2-qubit errors in :doc:`randomized benchmarking </manuals/verification/randomized_benchmarking>`
         * How to calculate the speedup from using :doc:`restless measurements </manuals/measurement/restless_measurements>`
 
         +++
