### 2023-03-18 32c26ac
Fix formatting and add classes to docs
Removed FakeService and FakeJob since they're more for internal testing
and not useful for most users
