+++ qiskit_experiments/test/mock_iq_helpers.py
@@ -26,7 +26,7 @@ IQPoint = Tuple[float, float]
 
 
 class MockIQExperimentHelper:
-    """Abstract class for the MockIQ helper classes
+    """Abstract class for the MockIQ helper classes.
 
     Different tests will use experiment specific helper classes which define the pattern
     of the IQ data that is then analyzed.
@@ -39,13 +39,14 @@ class MockIQExperimentHelper:
     ):
         """Create a MockIQBackend helper object to define how the backend functions.
 
-        `iq_cluster_centers` and `iq_cluster_width` define the base IQ cluster centers and
-        standard-deviations for each qubit in a :class:`MockIQBackend` instance. These are used by
-        :meth:`iq_clusters` by default. Subclasses can override :meth:`iq_clusters` to return a
-        modified version of attr:`iq_cluster_centers` and attr:`iq_cluster_width`.
-        `iq_cluster_centers` is a list of tuples. For a given qubit `i_qbt` and computational state
-        `i_state` (either `0` or `1`), the centers of the IQ clusters are found by indexing
-        `iq_cluster_centers` as follows:
+        :attr:`iq_cluster_centers` and :attr:`iq_cluster_width` define the base IQ
+        cluster centers and standard deviations for each qubit in a
+        :class:`MockIQBackend` instance. These are used by :meth:`iq_clusters` by
+        default. Subclasses can override :meth:`iq_clusters` to return a modified
+        version of :attr:`iq_cluster_centers` and :attr:`iq_cluster_width`.
+        `iq_cluster_centers` is a list of tuples. For a given qubit ``i_qbt`` and
+        computational state ``i_state`` (either `0` or `1`), the centers of the IQ
+        clusters are found by indexing ``iq_cluster_centers`` as follows:
 
         .. code-block:: python
 
@@ -53,8 +54,8 @@ class MockIQExperimentHelper:
             center_inphase = iq_center[0]
             center_quadrature = iq_center[1]
 
-        `iq_cluster_width` is indexed similarly except that there is only one width per qubit: i.e., the
-        standard-deviation of the IQ cluster for qubit `i_qbt` is
+        :attr:`iq_cluster_width` is indexed similarly except that there is only one width
+        per qubit: i.e., the standard deviation of the IQ cluster for qubit ``i_qbt`` is
 
         .. code-block:: python
 
@@ -109,61 +110,68 @@ class MockIQExperimentHelper:
 
         Examples:
 
-            **1 qubit circuit - excited state**
+        **1 qubit circuit - excited state**
 
-            In this experiment, we want to bring a qubit to its excited state and measure it.
-            The circuit:
-                         ┌───┐┌─┐
-                      q: ┤ X ├┤M├
-                         └───┘└╥┘
-                    c: 1/══════╩═
-                               0
+        In this experiment, we want to bring a qubit to its excited state and measure it.
+        The circuit:
 
-            The function that calculates the probability for this circuit, doesn't need any
-            calculation_parameters. It will be as following:
+        .. parsed-literal::
 
-            .. code-block::
+                 ┌───┐┌─┐
+            q:   ┤ X ├┤M├
+                 └───┘└╥┘
+            c: 1/══════╩═
+                        0
 
-                @staticmethod
-                def compute_probabilities(self, circuits: List[QuantumCircuit])
-                    -> List[Dict[str, float]]:
+        The function that calculates the probability for this circuit doesn't need any
+        calculation parameters:
 
-                    output_dict_list = []
-                    for circuit in circuits:
-                        probability_output_dict = {"1": 1.0, "0": 0.0}
-                        output_dict_list.append(probability_output_dict)
-                    return output_dict_list
+        .. code-block::
 
-            **3 qubit circuit**
[…]
