+++ docs/tutorials/visualization.rst
@@ -55,7 +55,7 @@ First, we display the default figure from a :class:`.Rabi` experiment as a start
     backend = SingleTransmonTestBackend()
 
     rabi = Rabi(
-        qubit=0,
+        physical_qubits=(0,),
         backend=backend,
         schedule=sched,
         amplitudes=np.linspace(-0.1, 0.1, 21)
@@ -64,7 +64,7 @@ First, we display the default figure from a :class:`.Rabi` experiment as a start
     rabi_data = rabi.run().block_for_results()
     rabi_data.figure(0)
 
-This is the default figure generated by :class:`OscillationAnalysis`, the data analysis
+This is the default figure generated by :class:`.OscillationAnalysis`, the data analysis
 class for the Rabi experiment. The fitted cosine is shown as a blue line, with the 
 individual measurements from the experiment shown as data points with error bars corresponding
 to their uncertainties. We are also given a small fit report in the caption showing the 
@@ -213,7 +213,6 @@ plotting a discriminator as optional supplementary data, which will show predict
 series over the axis area.
 
 .. jupyter-input::
-
 
     drag_experiment_helper = DragHelper(gate_name="Drag(xp)")
     backend = MockIQBackend(drag_experiment_helper)
