### 2023-03-19 b9acbf2
Removed bulk update usage as it's not robust enough for now
