+++ requirements.txt
@@ -1,7 +1,7 @@
 numpy>=1.17
 scipy>=1.4
 qiskit-terra>=0.22
-qiskit-ibm-experiment>=0.2.5
+qiskit-ibm-experiment>=0.3
 qiskit_dynamics>=0.3.0
 matplotlib>=3.4
 uncertainties
