### 2023-03-19 71e82ea
Added a flag for disabling figure upload
