### 2023-03-19 ab4e90f
Expose HalfAngleCal (#1091)
### Summary

The Half Angle Calibration experiment was not exposed. This PR adds the
necessary imports to correct this.
