+++ qiskit_experiments/library/characterization/drag.py
@@ -190,5 +190,5 @@ class RoughDrag(BaseExperiment, RestlessMixin):
         for run_opt in ["meas_level", "meas_return"]:
             if hasattr(self.run_options, run_opt):
                 metadata[run_opt] = getattr(self.run_options, run_opt)
-        metadata["nreps"] = self.experiment_options.reps
+        metadata["reps"] = self.experiment_options.reps
         return metadata
