+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -22,5 +22,4 @@ PR comments.
 - [ ] I have read the contributing guide `CONTRIBUTING.md`.
 - [ ] I have added the tests to cover my changes.
 - [ ] I have updated the documentation accordingly.
-- [ ] I have added a release note file using `reno` if this change needs to be 
-      documented in the release notes.
+- [ ] I have added a release note file using `reno` if this change needs to be documented in the release notes.
