+++ docs/conf.py
@@ -10,13 +10,6 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-# pylint: disable=invalid-name
-# Configuration file for the Sphinx documentation builder.
-#
-# This file does only contain a selection of the most common options. For a
-# full list see the documentation:
-# http://www.sphinx-doc.org/en/master/config
-
 """
 Sphinx documentation builder.
 """
@@ -53,13 +46,6 @@ author = "Qiskit Development Team"
 
 # -- General configuration ---------------------------------------------------
 
-# If your documentation needs a minimal Sphinx version, state it here.
-#
-# needs_sphinx = '1.0'
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
 extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.autodoc",
@@ -121,11 +107,7 @@ numfig = True
 # strings that are used for format of figure numbers. As a special character,
 # %s will be replaced to figure number.
 numfig_format = {"table": "Table %s"}
-# The language for content autogenerated by Sphinx. Refer to documentation
-# for a list of supported languages.
-#
-# This is also used if you do content translation via gettext catalogs.
-# Usually you set "language" from the command line for these cases.
+
 language = "en"
 
 # List of patterns, relative to source directory, that match files and
@@ -178,6 +160,7 @@ intersphinx_mapping = {
     "matplotlib": ("https://matplotlib.org/stable/", None),
     "qiskit": ("https://qiskit.org/documentation/", None),
     "uncertainties": ("https://pythonhosted.org/uncertainties", None),
+    "qiskit-ibm-provider": ("https://qiskit.org/documentation/partners/qiskit_ibm_provider", None),
 }
 
 
