+++ .github/workflows/main.yml
@@ -72,8 +72,6 @@ jobs:
         run: python -m pip install -U tox
       - name: Run lint
         run: tox -elint
-      - name: Lint docs
-        run: tox -edocs-lint -- -W
   docs:
     name: docs
     runs-on: ubuntu-latest
