+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,7 +1,6 @@
 ### Summary
 
-Please describe what this PR changes as concisely as possible. Link to the issue(s) 
-that this addresses, if any.
+Please describe what this PR changes as concisely as possible. Link to the issue(s) that this addresses, if any.
 
 ### Details and comments
 
@@ -14,7 +13,8 @@ Some details that should be in this section include:
 
 Note that this entire PR description field will be used as the commit message upon
 merge, so please keep it updated along with the PR. Secondary discussions, such as
-intermediate testing and bug statuses that do not affect the final PR, should be in the PR comments.
+intermediate testing and bug statuses that do not affect the final PR, should be in the
+PR comments.
 
 ### PR checklist (delete when all criteria are met)
 
