+++ test/decorators.py
@@ -1,347 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2018, 2019.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-"""Decorators for using with IBMQProvider unit tests.
-
-    Environment variables used by the decorators:
-        * QISKIT_IBM_API_TOKEN: default API token to use.
-        * QISKIT_IBM_API_URL: default API url to use.
-        * QISKIT_IBM_HGP: default hub/group/project to use.
-        * QISKIT_IBM_PRIVATE_HGP: hub/group/project to use for private jobs.
-        * QISKIT_IBM_DEVICE: default device to use.
-        * QISKIT_IBM_USE_STAGING_CREDENTIALS: True if use staging credentials.
-        * QISKIT_IBM_STAGING_API_TOKEN: staging API token to use.
-        * QISKIT_IBM_STAGING_API_URL: staging API url to use.
-        * QISKIT_IBM_STAGING_HGP: staging hub/group/project to use.
-        * QISKIT_IBM_STAGING_DEVICE: staging device to use.
-        * QISKIT_IBM_STAGING_PRIVATE_HGP: staging hub/group/project to use for private jobs.
-"""
-
-import os
-from functools import wraps
-from unittest import SkipTest
-from typing import Optional
-
-from qiskit import QiskitError
-from qiskit.test.testing_options import get_test_options
-
-# pylint can not follow namespace package like qiskit.providers.ibmq
-# pylint: disable=import-error,no-name-in-module
-from qiskit.providers.ibmq import least_busy
-from qiskit.providers.ibmq.ibmqfactory import IBMQFactory
-from qiskit.providers.ibmq.credentials import Credentials, discover_credentials
-from qiskit.providers.ibmq.accountprovider import AccountProvider
-from qiskit.providers.ibmq import IBMQ
-
-# pylint: enable=import-error,no-name-in-module
-
-
-def requires_qe_access(func):
-    """Decorator that signals that the test uses the online API.
-
-    It involves:
-        * determines if the test should be skipped by checking environment
-            variables.
-        * if the `QISKIT_IBM_USE_STAGING_CREDENTIALS` environment variable is
-          set, it reads the credentials from an alternative set of environment
-          variables.
-        * if the test is not skipped, it reads `qe_token` and `qe_url` from
-            environment variables or qiskitrc.
-        * if the test is not skipped, it appends `qe_token` and `qe_url` as
-            arguments to the test function.
-
-    Args:
-        func (callable): test function to be decorated.
-
-    Returns:
-        callable: the decorated function.
-    """
-
-    @wraps(func)
-    def _wrapper(obj, *args, **kwargs):
-        if get_test_options()["skip_online"]:
-            raise SkipTest("Skipping online tests")
-
-        credentials = _get_credentials()
-        kwargs.update({"qe_token": credentials.token, "qe_url": credentials.url})
-
-        return func(obj, *args, **kwargs)
-
-    return _wrapper
-
-
-def requires_providers(func):
-    """Decorator that signals the test uses the online API, via a public and premium provider.
-
-    This decorator delegates into the `requires_qe_access` decorator, but
-    instead of the credentials it appends a dictionary, containing the open access project
-    `public_provider` and a `premium_provider`, to the decorated function.
-
-    Args:
[…]
