+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -11,10 +11,7 @@ Some details that should be in this section include:
 - What tests and documentation have been added/updated
 - What do users and developers need to know about this change
 
-Note that this entire PR description field will be used as the commit message upon
-merge, so please keep it updated along with the PR. Secondary discussions, such as
-intermediate testing and bug statuses that do not affect the final PR, should be in the
-PR comments.
+Note that this entire PR description field will be used as the commit message upon merge, so please keep it updated along with the PR. Secondary discussions, such as intermediate testing and bug statuses that do not affect the final PR, should be in the PR comments.
 
 ### PR checklist (delete when all criteria are met)
 
