### 2023-03-21 577ff71
Write own docstring parser for default options and add automatic see also injection that references superclasses.
