+++ docs/_ext/custom_styles/styles.py
@@ -21,7 +21,6 @@ from typing import Union, List, Dict
 
 from qiskit_experiments.framework.base_analysis import BaseAnalysis
 from qiskit_experiments.framework.base_experiment import BaseExperiment
-from qiskit_experiments.visualization import BaseDrawer, BasePlotter
 from sphinx.config import Config as SphinxConfig
 
 from .formatter import (
@@ -32,19 +31,21 @@ from .formatter import (
 )
 from .section_parsers import load_standard_section, load_fit_parameters
 from .utils import (
-    _generate_options_documentation,
     _generate_analysis_ref,
-    _format_default_options,
+    _get_superclass,
 )
 
-section_regex = re.compile(r"# section: (?P<section_name>\S+)")
+
+_section_regex = re.compile(r"\s*# section:\s*(?P<section_key>\S+)")
 
 
 class QiskitExperimentDocstring(ABC):
     """Qiskit Experiment style docstring parser base class."""
 
     # mapping of sections supported by this style to parsing method or function
-    __sections__ = {}
+    __sections__ = {
+        "header": load_standard_section,
+    }
 
     # section formatter
     __formatter__ = DocstringSectionFormatter
@@ -55,6 +56,7 @@ class QiskitExperimentDocstring(ABC):
         docstring_lines: Union[str, List[str]],
         config: SphinxConfig,
         indent: str = "",
+        **extra_sections: List[str],
     ):
         """Create new parser and parse formatted docstring."""
 
@@ -67,54 +69,58 @@ class QiskitExperimentDocstring(ABC):
         self._indent = indent
         self._config = config
 
-        self._parsed_lines = self._classify(lines)
+        self._parsed_lines = self._classify(lines, **extra_sections)
 
-    def _classify(self, docstrings: List[str]) -> Dict[str, List[str]]:
+    def _classify(
+        self,
+        docstring_lines: List[str],
+        **extra_sections: List[str],
+    ) -> Dict[str, List[str]]:
         """Classify formatted docstring into sections."""
         sectioned_docstrings = dict()
 
-        def add_new_section(section: str, lines: List[str]):
-            if lines:
-                parser = self.__sections__[section]
-                if not parser:
+        for sec_key, parsed_lines in extra_sections.items():
+            if sec_key not in self.__sections__:
+                raise KeyError(
+                    f"Section key {sec_key} is not a valid Qiskit Experiments extension "
+                    f"section keys. Use one of {','.join(self.__sections__.keys())}."
+                )
+            sectioned_docstrings[sec_key] = parsed_lines
+
+        current_section = "header"
+        min_indent = sys.maxsize
+        tmp_lines = []
+        for line in docstring_lines:
+            matched = _section_regex.match(line)
+            if matched:
+                # Process previous section
+                if min_indent < sys.maxsize:
+                    tmp_lines = [_line[min_indent:] for _line in tmp_lines]
+                parser = self.__sections__[current_section]
+                sectioned_docstrings[current_section] = parser(tmp_lines)
+                # Start new line
+                sec_key = matched["section_key"]
+                if sec_key not in self.__sections__:
                     raise KeyError(
-                        f"Section {section} is automatically generated section. "
-                        "This section cannot be overridden by class docstring."
+                        f"Section key {sec_key} is not a valid Qiskit Experiments extension "
+                        f"section keys. Use one of {','.join(self.__sections__.keys())}."
                     )
-                sectioned_docstrings[section] = parser(temp_lines)
-
-        current_section = list(self.__sections__.keys())[0]
-        temp_lines = list()
-        margin = sys.maxsize
-        for docstring_line in docstrings:
-            match = re.match(section_regex, docstring_line.strip())
-            if match:
[…]
