+++ docs/_ext/custom_styles/utils.py
@@ -15,13 +15,18 @@ A collection of utilities to generate documentation.
 
 import inspect
 import re
-from typing import List, Set, Tuple, Dict, Any, Callable, Type
+import collections
+from typing import List, Callable, Type, Iterator
 
 from sphinx.config import Config as SphinxConfig
 from sphinx.ext.napoleon.docstring import GoogleDocstring
 from sphinx.util.docstrings import prepare_docstring
 
 from qiskit_experiments.framework import BaseExperiment
+
+
+_parameter_regex = re.compile(r'(.+?)\(\s*(.*[^\s]+)\s*\):(.*[^\s]+)')
+_rest_role_regex = re.compile(r':(.+?) (.+?):\s*(.*[^\s]+)')
 
 
 def _trim_empty_lines(docstring_lines: List[str]) -> List[str]:
@@ -37,124 +42,6 @@ def _trim_empty_lines(docstring_lines: List[str]) -> List[str]:
         i_end -= 1
 
     return docstring_lines[i_start:i_end]
-
-
-def _parse_option_field(
-    docstring: str,
-    config: SphinxConfig,
-    target_args: Set[str],
-    indent: str = "",
-) -> Tuple[List[str], Set[str]]:
-    """A helper function to extract descriptions of target arguments."""
-    # Convert str into list of reST lines.
-    rest_doc = prepare_docstring(docstring, tabsize=len(indent))
-
-    # Extra formatting to remove line feed and non-target documentation.
-    line_ind = 0
-    while line_ind < len(rest_doc):
-        if "Options:" in rest_doc[line_ind]:
-            break
-        line_ind += 1
-    else:
-        return [], target_args
-    line0 = rest_doc[line_ind + 1]
-    section_indent = len(line0) - len(line0.lstrip())
-
-    to_parse = []
-    described_params = set()
-    tmp = ""
-    for line in rest_doc[line_ind + 1:]:
-        if line[section_indent:].startswith(indent):
-            # Remove line-feed for safety.
-            # GoogleDocstring parser doesn't support line break during type annotation.
-            #
-            # e.g.
-            #     param1 (some_very_long_type1 or some_very_long_type2
-            #         some_very_long_type2): here is documentation.
-            #
-            # Parsing above documentation always fails, and type is not recognized.
-            tmp += " " + line.lstrip()
-        else:
-            if tmp:
-                argname = tmp.split(maxsplit=1)[0]
-                if argname in target_args:
-                    # Add only description for target option. Otherwise, just ignore.
-                    described_params.add(argname)
-                    to_parse.append(tmp)
-            tmp = line
-    # find missing parameters
-    missing = target_args - described_params
-
-    if not to_parse:
-        return [],  missing
-
-    # readd section header, e.g. XXX Options:
-    to_parse.insert(0, rest_doc[line_ind])
-
-    # use GoogleDocstring parameter parser
-    experiment_option_parser = GoogleDocstring(docstring=to_parse, config=config)
-    parsed_lines = experiment_option_parser.lines()
-
-    return parsed_lines, missing
-
-
-def _generate_options_documentation(
-    current_class: Type,
-    method_name: str,
-    target_args: List[str] = None,
-    config: SphinxConfig = None,
-    indent: str = "",
-) -> List[str]:
-    """Automatically generate documentation from the default options method."""
-    options_docstring_lines = []
-
-    if not target_args:
-        default_opts_clsmethod = getattr(current_class, method_name, None)
-        if not default_opts_clsmethod:
-            # getter option is not defined
-            return []
-        target_args = set(default_opts_clsmethod().__dict__.keys())
-
-    mro_classes = inspect.getmro(current_class)
-    for i, mro_cls in enumerate(mro_classes):
[…]
