+++ qiskit_experiments/library/tomography/qpt_analysis.py
@@ -59,9 +59,6 @@ class ProcessTomographyAnalysis(TomographyAnalysis):
     # section: reference
         .. ref_arxiv:: 1 1106.5458
 
-    # section: see_also
-        :class:`.TomographyAnalysis`
-
     """
 
     @classmethod
