### 2023-03-21 caa06a4
suppress zero div warning
