+++ releasenotes/notes/suppress-runtime-warning-a741dc96f6a0ce7a.yaml
@@ -1,0 +1,8 @@
+---
+other:
+  - |
+    NumPy runtime warning for zero division is suppressed in :class:`.CurveAnalysis`.
+    This warning could occur in the edge case where the experiment data
+    may contain data point with zero uncertainty.
+    Such data point is safely ignored by LMFIT, since it may apply infinite fit weight.
+    This runtime warning suppression makes standard output cleaner.
