### 2023-03-21 7651798
rewrote circuit generation logic and sampler
the sampler now outputs tuples of indices and elements instead of
circuit layers. 1q cliffords are kept as integers through the entire
sampling process. performance has improved a lot.
