+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -15,46 +15,88 @@ Utilities for sampling layers in randomized benchmarking experiments
 
 import warnings
 from abc import ABC, abstractmethod
-from typing import Optional, Union, Sequence, List, Tuple
+from typing import Optional, Union, Sequence, List, Tuple, TypeVar
 from numbers import Integral
 
 import numpy as np
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
+from functools import lru_cache
 
-from qiskit import QuantumCircuit, QuantumRegister
+from qiskit import QuantumCircuit, QuantumRegister, Operator
 from qiskit.exceptions import QiskitError
 from qiskit.circuit import Gate
-from qiskit.circuit.library import (
-    IGate,
-    XGate,
-    YGate,
-    ZGate,
-    CXGate,
-    CYGate,
-    SGate,
-    SdgGate,
-    SXGate,
-    SXdgGate,
-)
-from qiskit.quantum_info import random_unitary, random_clifford, Pauli
+from qiskit.quantum_info.operators.base_operator import BaseOperator
+from qiskit.quantum_info import random_unitary, Pauli, random_pauli
 from qiskit.extensions import UnitaryGate
 from qiskit.converters import circuit_to_dag
-from .clifford_utils import CliffordUtils
+from .clifford_utils import CliffordUtils, _clifford_1q_int_to_instruction
+from .rb_experiment import SequenceElementType
+
+from time import time
+
+Idx = TypeVar("Idx", int, Tuple[int, ...])
+Gate = TypeVar("Gate", Instruction, Operator, QuantumCircuit, BaseOperator)
 
 
-class MirrorRBSampler(ABC):
-    """Sampling distribution for the mirror randomized benchmarking experiment."""
+def timer_func(func):
+    # This function shows the execution time of
+    # the function object passed
+    def wrap_func(*args, **kwargs):
+        t1 = time()
+        result = func(*args, **kwargs)
+        t2 = time()
+        print(f"Function {func.__name__!r} executed in {(t2-t1):.4f}s")
+        return result
+
+    return wrap_func
+
+
+class RBSampler(ABC):
+    """Sampling distribution for randomized benchmarking experiments.
+    Subclasses must implement the ``__call__()`` method."""
 
     def __init__(self):
         pass
 
     @abstractmethod
-    def __call__(self, num_qubits, seed=None, **params):
-        """Samplers should define this method so that it returns a single sampled layer."""
+    def __call__(self, num_qubits, seed=None, **params) -> Tuple[Tuple[Idx, Gate], ...]:
+        """Samplers should define this method such that it returns sampled layers
+        given the input parameters. Each layer is represented by a list of size-2 tuples
+        where the first element is a qubit index or a tuple of indices, and the second
+        element represents the gate that should be applied."""
         pass
 
 
-class EdgeGrabSampler(MirrorRBSampler):
+class SingleQubitSampler(RBSampler):
+    def __call__(
+        self,
+        num_qubits,
+        length,
+        gate_set: Optional[Union[str, List]] = "clifford",
+        seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
+    ) -> Tuple[Tuple[Idx, Gate], ...]:
+        """Samples a layer of random single-qubit gates from a specified gate set.
+
+        Returns:
+        For two qubits and length 3, a possible output would be
+        (((0,20),(1,14)),((0,11),(1,4)),((0,2),(1,9)))
+        """
+        rng = default_rng(seed=seed)
+        if gate_set == "clifford":
+            layers = []
+            for i in rng.integers(CliffordUtils.NUM_CLIFFORD_1_QUBIT, size=(length, num_qubits)):
+                layers.append(tuple(zip((range(num_qubits)), i)))
+        elif gate_set == "pauli":
[…]
