+++ qiskit_experiments/library/randomized_benchmarking/__init__.py
@@ -48,12 +48,13 @@ Utilities
     RBUtils
     MirrorRBSampler
     EdgeGrabSampler
+    SingleQubitSampler
 
 """
 from .rb_experiment import StandardRB
 from .interleaved_rb_experiment import InterleavedRB
 from .mirror_rb_experiment import MirrorRB
-from .sampling_utils import MirrorRBSampler, EdgeGrabSampler
+from .sampling_utils import RBSampler, EdgeGrabSampler, SingleQubitSampler
 from .rb_analysis import RBAnalysis
 from .interleaved_rb_analysis import InterleavedRBAnalysis
 from .mirror_rb_analysis import MirrorRBAnalysis
