+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -12,10 +12,10 @@
 """
 Mirror RB Experiment class.
 """
-from typing import Union, Iterable, Optional, List, Sequence
+from typing import Union, Iterable, Optional, List, Sequence, Tuple
 from numbers import Integral
-from itertools import permutations
-
+import itertools
+from time import time
 
 from numpy.random import Generator, BitGenerator, SeedSequence, default_rng
 
@@ -30,8 +30,21 @@ from qiskit_experiments.warnings import deprecate_arguments
 
 from .rb_experiment import StandardRB, SequenceElementType
 from .mirror_rb_analysis import MirrorRBAnalysis
-from .clifford_utils import compute_target_bitstring
-from .sampling_utils import MirrorRBSampler, EdgeGrabSampler
+from .clifford_utils import compute_target_bitstring, CliffordUtils, inverse_1q
+from .sampling_utils import RBSampler, EdgeGrabSampler, SingleQubitSampler
+
+
+def timer_func(func):
+    # This function shows the execution time of
+    # the function object passed
+    def wrap_func(*args, **kwargs):
+        t1 = time()
+        result = func(*args, **kwargs)
+        t2 = time()
+        print(f"Function {func.__name__!r} executed in {(t2-t1):.4f}s")
+        return result
+
+    return wrap_func
 
 
 class MirrorRB(StandardRB):
@@ -67,7 +80,7 @@ class MirrorRB(StandardRB):
         self,
         physical_qubits: Sequence[int],
         lengths: Iterable[int],
-        distribution: MirrorRBSampler = EdgeGrabSampler,
+        distribution: RBSampler = EdgeGrabSampler,
         local_clifford: bool = True,
         pauli_randomize: bool = True,
         two_qubit_gate_density: float = 0.2,
@@ -171,6 +184,25 @@ class MirrorRB(StandardRB):
 
         return circuits
 
+    def _inverse_layer(self, layer) -> List[Tuple[Union[int, Tuple], int]]:
+        """Generates the inverse layer of a Clifford mirror RB layer by inverting the
+        single-qubit Cliffords and keeping the CXs identical. See
+        :class:`.RBSampler` for the format of the layer.
+
+        Raises:
+            QiskitError: If the layer has invalid format.
+        """
+        inverse_layer = []
+        for elem in layer:
+            if isinstance(elem[0], int):
+                inverse_layer.append((elem[0], inverse_1q(elem[1])))
+            elif isinstance(elem[0], tuple):
+                inverse_layer.append(elem)
+            else:
+                raise QiskitError("Invalid layer from sampler.")
+        return tuple(inverse_layer)
+
+    @timer_func
     def _sample_sequences(self) -> List[Sequence[SequenceElementType]]:
         """Sample layers of mirror RB using the provided distribution and user options.
         First, layers are sampled using the distribution, then Pauli-dressed if
@@ -186,13 +218,12 @@ class MirrorRB(StandardRB):
             A list of mirror RB sequences. Each element is a list of layers with length
             matching the corresponding element in ``lengths``.
         """
-        rng = default_rng(seed=self.experiment_options.seed)
         if not self._backend:
-            raise QiskitError("A backend must be provided.")
+            raise QiskitError("A backend must be provided for circuit generation.")
 
         # Coupling map is full connectivity by default. If backend has a coupling map,
         # get backend coupling map and create coupling map for physical qubits
-        coupling_map = list(permutations(range(max(self.physical_qubits) + 1), 2))
+        coupling_map = list(itertools.permutations(range(max(self.physical_qubits) + 1), 2))
         if self._backend_data.coupling_map:
             coupling_map = self._backend_data.coupling_map
         experiment_coupling_map = []
[…]
