+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -27,24 +27,17 @@ from qiskit.providers.backend import Backend
 from qiskit.providers.options import Options
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit_experiments.warnings import deprecate_arguments
+from qiskit.circuit.library import CXGate
 
 from .rb_experiment import StandardRB, SequenceElementType
 from .mirror_rb_analysis import MirrorRBAnalysis
-from .clifford_utils import compute_target_bitstring, CliffordUtils, inverse_1q
-from .sampling_utils import RBSampler, EdgeGrabSampler, SingleQubitSampler
-
-
-def timer_func(func):
-    # This function shows the execution time of
-    # the function object passed
-    def wrap_func(*args, **kwargs):
-        t1 = time()
-        result = func(*args, **kwargs)
-        t2 = time()
-        print(f"Function {func.__name__!r} executed in {(t2-t1):.4f}s")
-        return result
-
-    return wrap_func
+from .clifford_utils import (
+    compute_target_bitstring,
+    CliffordUtils,
+    inverse_1q,
+    _clifford_1q_int_to_instruction,
+)
+from .sampling_utils import RBSampler, EdgeGrabSampler, SingleQubitSampler, GateType
 
 
 class MirrorRB(StandardRB):
@@ -184,39 +177,49 @@ class MirrorRB(StandardRB):
 
         return circuits
 
-    def _inverse_layer(self, layer) -> List[Tuple[Union[int, Tuple], int]]:
+    def _inverse_layer(
+        self, layer: List[Tuple[Tuple[int, ...], GateType]]
+    ) -> List[Tuple[Tuple[int, ...], GateType]]:
         """Generates the inverse layer of a Clifford mirror RB layer by inverting the
         single-qubit Cliffords and keeping the CXs identical. See
         :class:`.RBSampler` for the format of the layer.
+
+        Args:
+            layer: The input layer.
+
+        Returns:
+            The layer that performs the inverse operation to the input layer.
 
         Raises:
             QiskitError: If the layer has invalid format.
         """
         inverse_layer = []
         for elem in layer:
-            if isinstance(elem[0], int):
+            if len(elem[0]) == 1:
                 inverse_layer.append((elem[0], inverse_1q(elem[1])))
-            elif isinstance(elem[0], tuple):
+            elif len(elem[0]) == 2:
                 inverse_layer.append(elem)
             else:
                 raise QiskitError("Invalid layer from sampler.")
         return tuple(inverse_layer)
 
-    @timer_func
     def _sample_sequences(self) -> List[Sequence[SequenceElementType]]:
         """Sample layers of mirror RB using the provided distribution and user options.
+
         First, layers are sampled using the distribution, then Pauli-dressed if
         ``pauli_randomize`` is ``True``. The inverse of the resulting circuit is
         appended to the end. If ``local_clifford`` is ``True``, then cliffords are added
         to the beginning and end. If ``inverting_pauli_layer`` is ``True``, a Pauli
         layer will be appended at the end to set the output bitstring to all zeros.
 
-        Raises:
-            QiskitError: If no backend is provided.
-
         Returns:
             A list of mirror RB sequences. Each element is a list of layers with length
-            matching the corresponding element in ``lengths``.
+            matching the corresponding element in ``lengths``. The layers are made up
+            of tuples in the format ((one or more qubit indices), gate). Single-qubit
+            Cliffords are represented by integers for speed.
+
+        Raises:
+            QiskitError: If no backend is provided.
         """
         if not self._backend:
             raise QiskitError("A backend must be provided for circuit generation.")
@@ -240,11 +243,9 @@ class MirrorRB(StandardRB):
[…]
