+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -20,35 +20,19 @@ from numbers import Integral
 
 import numpy as np
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
-from functools import lru_cache
 
-from qiskit import QuantumCircuit, QuantumRegister, Operator
+from qiskit import QuantumCircuit, QuantumRegister
 from qiskit.exceptions import QiskitError
-from qiskit.circuit import Gate
+from qiskit.circuit import Instruction
+from qiskit.circuit.library import CXGate
 from qiskit.quantum_info.operators.base_operator import BaseOperator
-from qiskit.quantum_info import random_unitary, Pauli, random_pauli
+from qiskit.quantum_info import random_unitary, Pauli, random_pauli, Operator
 from qiskit.extensions import UnitaryGate
 from qiskit.converters import circuit_to_dag
 from .clifford_utils import CliffordUtils, _clifford_1q_int_to_instruction
 from .rb_experiment import SequenceElementType
 
-from time import time
-
-Idx = TypeVar("Idx", int, Tuple[int, ...])
-Gate = TypeVar("Gate", Instruction, Operator, QuantumCircuit, BaseOperator)
-
-
-def timer_func(func):
-    # This function shows the execution time of
-    # the function object passed
-    def wrap_func(*args, **kwargs):
-        t1 = time()
-        result = func(*args, **kwargs)
-        t2 = time()
-        print(f"Function {func.__name__!r} executed in {(t2-t1):.4f}s")
-        return result
-
-    return wrap_func
+GateType = TypeVar("Gate", Instruction, Operator, QuantumCircuit, BaseOperator)
 
 
 class RBSampler(ABC):
@@ -59,11 +43,11 @@ class RBSampler(ABC):
         pass
 
     @abstractmethod
-    def __call__(self, num_qubits, seed=None, **params) -> Tuple[Tuple[Idx, Gate], ...]:
+    def __call__(self, num_qubits, seed=None, **params) -> List[Tuple[Tuple[int, ...], GateType]]:
         """Samplers should define this method such that it returns sampled layers
         given the input parameters. Each layer is represented by a list of size-2 tuples
-        where the first element is a qubit index or a tuple of indices, and the second
-        element represents the gate that should be applied."""
+        where the first element is a tuple of qubit indices, and the second
+        element represents the gate that should be applied to the indices."""
         pass
 
 
@@ -74,25 +58,43 @@ class SingleQubitSampler(RBSampler):
         length,
         gate_set: Optional[Union[str, List]] = "clifford",
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
-    ) -> Tuple[Tuple[Idx, Gate], ...]:
+    ) -> List[Tuple[Tuple[int, ...], GateType]]:
         """Samples a layer of random single-qubit gates from a specified gate set.
 
+        Args:
+            num_qubits: The number of qubits to generate layers for.
+            length: The length of the sequence to output.
+            gate_set: The one qubit gate set to sample from. Can be either a
+                list of gates, "clifford", or "pauli". Default is "clifford".
+            seed: Seed for random generation.
+
         Returns:
-        For two qubits and length 3, a possible output would be
-        (((0,20),(1,14)),((0,11),(1,4)),((0,2),(1,9)))
+            A ``length``-long list of :class:`qiskit.circuit.QuantumCircuit` layers over
+            ``num_qubits`` qubits. Each layer is represented by a list of tuples which
+            are in the format ((one or more qubit indices), gate). Single-qubit
+            Cliffords are represented by integers for speed. For two qubits, length
+            3, and the default Clifford gate set, an example output would be
+
+                .. code-block::
[…]
