### 2023-03-22 97a746d
fixed wrong indexing for general coupling maps
also changed samplers to qubit list input instead of num_qubits.
