+++ qiskit_experiments/library/randomized_benchmarking/__init__.py
@@ -46,7 +46,7 @@ Utilities
     :toctree: ../stubs/
 
     RBUtils
-    MirrorRBSampler
+    RBSampler
     EdgeGrabSampler
     SingleQubitSampler
 
