+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -16,21 +16,16 @@ Utilities for sampling layers in randomized benchmarking experiments
 import warnings
 from abc import ABC, abstractmethod
 from typing import Optional, Union, Sequence, List, Tuple, TypeVar
-from numbers import Integral
 
-import numpy as np
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
 
-from qiskit import QuantumCircuit, QuantumRegister
+from qiskit import QuantumCircuit
 from qiskit.exceptions import QiskitError
 from qiskit.circuit import Instruction
 from qiskit.circuit.library import CXGate
 from qiskit.quantum_info.operators.base_operator import BaseOperator
-from qiskit.quantum_info import random_unitary, Pauli, random_pauli, Operator
-from qiskit.extensions import UnitaryGate
-from qiskit.converters import circuit_to_dag
-from .clifford_utils import CliffordUtils, _clifford_1q_int_to_instruction
-from .rb_experiment import SequenceElementType
+from qiskit.quantum_info import random_pauli, Operator
+from .clifford_utils import CliffordUtils
 
 GateType = TypeVar("Gate", Instruction, Operator, QuantumCircuit, BaseOperator)
 
@@ -43,26 +38,34 @@ class RBSampler(ABC):
         pass
 
     @abstractmethod
-    def __call__(self, num_qubits, seed=None, **params) -> List[Tuple[Tuple[int, ...], GateType]]:
+    def __call__(self, qubits, seed=None, **params) -> List[Tuple[Tuple[int, ...], GateType]]:
         """Samplers should define this method such that it returns sampled layers
         given the input parameters. Each layer is represented by a list of size-2 tuples
         where the first element is a tuple of qubit indices, and the second
-        element represents the gate that should be applied to the indices."""
+        element represents the gate that should be applied to the indices.
+
+        Args:
+            qubits: A sequence of qubits to generate layers for.
+            seed: Seed for random generation.
+
+        """
         pass
 
 
 class SingleQubitSampler(RBSampler):
+    """A sampler that samples layers of random single-qubit gates from a specified gate set."""
+
     def __call__(
         self,
-        num_qubits,
+        qubits,
         length,
         gate_set: Optional[Union[str, List]] = "clifford",
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
     ) -> List[Tuple[Tuple[int, ...], GateType]]:
-        """Samples a layer of random single-qubit gates from a specified gate set.
+        """Samples random single-qubit gates from a specified gate set.
 
         Args:
-            num_qubits: The number of qubits to generate layers for.
+            qubits: A sequence of qubits to generate layers for.
             length: The length of the sequence to output.
             gate_set: The one qubit gate set to sample from. Can be either a
                 list of gates, "clifford", or "pauli". Default is "clifford".
@@ -75,31 +78,32 @@ class SingleQubitSampler(RBSampler):
             Cliffords are represented by integers for speed. For two qubits, length
             3, and the default Clifford gate set, an example output would be
 
-                .. code-block::
+                .. parsed-literal::
+
                     [(((0,), 2), ((1,), 2)), (((0,), 11), ((1,), 4)), (((0,), 17), ((1,), 17))]
         """
         rng = default_rng(seed=seed)
+        num_qubits = len(qubits)
         if gate_set == "clifford":
             layers = []
             for i in rng.integers(CliffordUtils.NUM_CLIFFORD_1_QUBIT, size=(length, num_qubits)):
-                layers.append(tuple(zip(((j,) for j in range(num_qubits)), i)))
+                layers.append(tuple(zip(((j,) for j in qubits), i)))
[…]

[…]
