+++ releasenotes/notes/suppress-runtime-warning-a741dc96f6a0ce7a.yaml
@@ -5,4 +5,4 @@ other:
     This warning could occur in the edge case where the experiment data
     may contain data point with zero uncertainty.
     Such data point is safely ignored by LMFIT, since it may apply infinite fit weight.
-    This runtime warning suppression makes standard output cleaner.
+    This runtime warning suppression makes standard error cleaner.
