### 2023-03-22 eca6ce2
Documentation fixes and formatting (#1095)
### Summary

This PR is essentially the documentation fixes from #1080 since the lint
CI pipeline needs more work.

### Details and comments

- The docs are no longer incompatible with the newest
sphinx-autodoc-typehints so the version pin from #1017 has been removed.
This closes #1018.
- Updates curve analysis tutorial to the newest version (#983 had
accidentally reverted to an older version)
- Changes arxiv autolinks from pdfs to abstract page to be more user
friendly
- Renames the new ExperimentData how-to to rerunning analysis and adds
additional 
[…]
