+++ qiskit_experiments/curve_analysis/data_processing.py
@@ -1,293 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2021.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-"""
-Data processing utility functions for curve fitting experiments
-"""
-# pylint: disable = invalid-name
-
-from typing import List, Dict, Tuple, Optional, Callable
-import numpy as np
-from qiskit.exceptions import QiskitError
-
-
-def filter_data(data: List[Dict[str, any]], **filters) -> List[Dict[str, any]]:
-    """Return the list of filtered data
-
-    Args:
-        data: list of data dicts.
-        filters: kwargs for filtering based on metadata
-                 values.
-
-    Returns:
-        The list of filtered data. If no filters are provided this will be the
-        input list.
-    """
-    if not filters:
-        return data
-    filtered_data = []
-    for datum in data:
-        include = True
-        metadata = datum["metadata"]
-        for key, val in filters.items():
-            if key not in metadata or metadata[key] != val:
-                include = False
-                break
-        if include:
-            filtered_data.append(datum)
-    return filtered_data
-
-
-def mean_xy_data(
-    xdata: np.ndarray,
-    ydata: np.ndarray,
-    sigma: Optional[np.ndarray] = None,
-    shots: Optional[np.ndarray] = None,
-    method: str = "sample",
-) -> Tuple[np.ndarray, ...]:
-    r"""Return (x, y_mean, sigma) data.
-
-    The mean is taken over all ydata values with the same xdata value using
-    the specified method. For each x the mean :math:`\overline{y}` and variance
-    :math:`\sigma^2` are computed as
-
-    * ``"sample"`` (default) *Sample mean and variance*
-      :math:`\overline{y} = \sum_{i=1}^N y_i / N`,
-      :math:`\sigma^2 = \sum_{i=1}^N ((\overline{y} - y_i)^2) / N`
-    * ``"iwv"`` *Inverse-weighted variance*
-      :math:`\overline{y} = (\sum_{i=1}^N y_i / \sigma_i^2 ) \sigma^2`
-      :math:`\sigma^2 = 1 / (\sum_{i=1}^N 1 / \sigma_i^2)`
-    * ``"shots_weighted_variance"`` *Sample mean and variance with weights from shots*
-      :math:`\overline{y} = \sum_{i=1}^N n_i y_i / M`,
-      :math:`\sigma^2 = \sum_{i=1}^N (n_i \sigma_i / M)^2`,
-      where :math:`n_i` is the number of shots per data point and :math:`M = \sum_{i=1}^N n_i`
-      is a total number of shots from different circuit execution at the same x value.
-      If ``shots`` is not provided, this applies uniform weights to all values.
-
-    Args
-        xdata: 1D or 2D array of xdata from curve_fit_data or
-               multi_curve_fit_data
-        ydata: array of ydata returned from curve_fit_data or
-               multi_curve_fit_data
-        sigma: Optional, array of standard deviations in ydata.
-        shots: Optional, array of shots used to get a data point.
-        method: The method to use for computing y means and
-                standard deviations sigma (default: "sample").
-
-    Returns:
-        tuple: ``(x, y_mean, sigma, shots)``, where
-               ``x`` is an arrays of unique x-values, ``y`` is an array of
-               sample mean y-values, ``sigma`` is an array of sample standard
[…]
