### 2023-03-22 95d0e52
Update manual and API docs
