+++ tools/deploy_documentation_dev.sh
@@ -29,4 +29,5 @@ pwd
 # Push to qiskit.org website
 openssl aes-256-cbc -K $encrypted_rclone_key -iv $encrypted_rclone_iv -in tools/rclone.conf.enc -out $RCLONE_CONFIG_PATH -d
 echo "Pushing built docs to dev website"
+rclone sync --progress ./docs/_build/html IBMCOS:qiskit-org-web-resources/documentation/ecosystem/dev
 rclone sync --progress ./docs/_build/html IBMCOS:qiskit-org-web-resources/documentation/experiments/dev
