### 2023-03-22 4344c0e
Revert "Ignore outdated pylint warnings"
This reverts commit 83615b613c2e536739cef43baf9c240d3bbf6703.

Since the pylint PR can likely be merged soon, we can just wait for it
and merge this one without workarounds.
