### 2023-03-22 bccac09
Update pylint to 2.16 (#1042)
### Summary

This change set updates pylint from 2.7 to 2.16. The code was modified
to satsify new checks added to pylint between those versions. Obsolete
settings were removed from the pylintrc file.

### Details and comments

The old version of pylint produces false positives that were fixed in a
later version than 2.7.
