+++ qiskit_experiments/curve_analysis/curve_data.py
@@ -464,7 +464,7 @@ class OptionsDict(dict):
         """
         for key, value in kwargs.items():
             if self.get(key) is None:
-                self.__setitem__(key, value)
+                self[key] = value
 
     @staticmethod
     def format(value: Any) -> Any:
