+++ test/library/calibration/test_fine_frequency.py
@@ -16,8 +16,8 @@ from test.base import QiskitExperimentsTestCase
 import numpy as np
 from ddt import ddt, data
 
+from qiskit import pulse
 from qiskit.providers.fake_provider import FakeArmonkV2
-import qiskit.pulse as pulse
 
 from qiskit_experiments.library import (
     FineFrequency,
