### 2023-03-22 #5257 @eggerdj
Dear Daniel, after discussing with the team here are some reviewer suggestions taken from the list and other names that you may wish to consider:
```
- Michael Goerz (on the JOSS list)
- Tejas Shetty (on the JOSS list)
- Someone from Prof. Stefan Filipp's group at WMI
- Lauren Capelluto https://www.linkedin.com/in/lcapelluto
- Benjamin Rosand https://www.linkedin.com/in/benjamin-rosand
- Zachary Schoenfeld https://www.linkedin.com/in/zachary-schoenfeld-ab4678155
```
