### 2023-03-22 a6c8ff8
refactored sampling utils and added tests
Improved the sampler interface by making several input parameters part
of base class attributes instead of set at __call__. Now the
distribution can be more easily customized with defined gate ratios.
