+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -14,30 +14,129 @@ Utilities for sampling layers in randomized benchmarking experiments
 """
 
 import warnings
+import math
 from abc import ABC, abstractmethod
-from typing import Optional, Union, List, Tuple, TypeVar
+from typing import Optional, Union, List, Tuple, Sequence, TypeVar
+from collections import defaultdict
 
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
 
 from qiskit import QuantumCircuit
 from qiskit.circuit import Instruction
-from qiskit.circuit.library import CXGate
 from qiskit.quantum_info.operators.base_operator import BaseOperator
-from qiskit.quantum_info import random_pauli, Operator
-from .clifford_utils import CliffordUtils
+from qiskit.quantum_info import Operator
+from qiskit.exceptions import QiskitError
 
-GateType = TypeVar("Gate", Instruction, Operator, QuantumCircuit, BaseOperator)
+from .clifford_utils import CliffordUtils, _CLIFF_SINGLE_GATE_MAP_1Q
+
+GateType = TypeVar("GateType", str, int, Instruction, Operator, QuantumCircuit, BaseOperator)
 
 
 class RBSampler(ABC):
     """Sampling distribution for randomized benchmarking experiments.
     Subclasses must implement the ``__call__()`` method."""
 
-    def __init__(self):
-        pass
+    def __init__(
+        self,
+        gate_distribution=None,
+        coupling_map: Optional[List[List[int]]] = None,
+        seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
+    ):
+        """Instantiate the sampler. Note that this is different from invoking the
+        instance.
+
+        Args:
+            seed: Seed for random generation.
+            gate_distribution: The gate distribution for sampling.
+        """
+        self._gate_distribution = gate_distribution
+        self._coupling_map = coupling_map
+        self.seed = default_rng(seed)
+
+    @property
+    def coupling_map(self):
+        """The coupling map of the system to sample over."""
+        return self._coupling_map
+
+    @coupling_map.setter
+    def coupling_map(self, coupling_map):
+        self._coupling_map = coupling_map
+
+    @property
+    def seed(self):
+        """The seed for random generation."""
+        return self._rng
+
+    @seed.setter
+    def seed(self, seed):
+        self._rng = default_rng(seed)
+
+    @property
+    def gate_distribution(self):
+        """The gate distribution for sampling. The distribution is a list of tuples
+        of the form ``(probability, width of gate, gate)``. Gates can be actual
+        operators, circuits, or the special keywords "clifford", "pauli", "idle". An
+        example distribution for the edge grab sampler might be
+
+        .. parsed-literal::
+            [(0.8, 1, "clifford"), (0.2, 2, CXGate)]
+
+        The probabilities must sum to 1."""
+        return self._gate_distribution
+
+    @gate_distribution.setter
+    def gate_distribution(self, dist: List[Tuple[float, int, GateType]]):
+        """Set the distribution of gates used in the sampler.
+
+        Args:
+            dist: A list of tuples with format ``(probability, width of gate, gate)``.
+        """
+        if not isinstance(dist, List):
+            raise TypeError("The gate distribution should be a list.")
+        if sum(list(zip(*dist))[0]) != 1:
+            raise QiskitError("Gate distribution probabilities must sum to 1.")
+
+        self._gate_distribution = dist
+
+    def _probs_by_gate_size(self):
+        """Return a list of gates and their probabilities indexed by the size of the gate. The
+        probability distributions are normalized to 1 within each gate size category.
+        """
+        if not self._gate_distribution:
+            raise QiskitError("Gate distribution must be set before sampling.")
[…]
