+++ qiskit_experiments/library/randomized_benchmarking/rb_experiment.py
@@ -57,17 +57,23 @@ class StandardRB(BaseExperiment, RestlessMixin):
     """An experiment to characterize the error rate of a gate set on a device.
 
     # section: overview
-        Randomized Benchmarking (RB) is an efficient and robust method
-        for estimating the average error rate of a set of quantum gate operations.
-        See `Qiskit Textbook
-        <https://qiskit.org/textbook/ch-quantum-hardware/randomized-benchmarking.html>`_
-        for an explanation on the RB method.
 
-        A standard RB experiment generates sequences of random Cliffords
-        such that the unitary computed by the sequences is the identity.
-        After running the sequences on a backend, it calculates the probabilities to get back to
-        the ground state, fits an exponentially decaying curve, and estimates
-        the Error Per Clifford (EPC), as described in Refs. [1, 2].
+    Randomized Benchmarking (RB) is an efficient and robust method
+    for estimating the average error rate of a set of quantum gate operations.
+    See `Qiskit Textbook
+    <https://qiskit.org/textbook/ch-quantum-hardware/randomized-benchmarking.html>`_
+    for an explanation on the RB method.
+
+    A standard RB experiment generates sequences of random Cliffords
+    such that the unitary computed by the sequences is the identity.
+    After running the sequences on a backend, it calculates the probabilities to get back to
+    the ground state, fits an exponentially decaying curve, and estimates
+    the Error Per Clifford (EPC), as described in Refs. [1, 2].
+
+    .. note::
+        In 0.5.0, the default value of ``optimization_level`` in ``transpile_options`` changed
+        from ``0`` to ``1`` for RB experiments. That may result in shorter RB circuits
+        hence slower decay curves than before.
 
     # section: analysis_ref
         :class:`RBAnalysis`
@@ -75,11 +81,6 @@ class StandardRB(BaseExperiment, RestlessMixin):
     # section: reference
         .. ref_arxiv:: 1 1009.3639
         .. ref_arxiv:: 2 1109.6887
-
-    .. note::
-        In 0.5.0, the default value of ``optimization_level`` in ``transpile_options`` changed
-        from ``0`` to ``1`` for RB experiments. That may result in shorter RB circuits
-        hence slower decay curves than before.
     """
 
     @deprecate_arguments({"qubits": "physical_qubits"}, "0.5")
