### 2023-03-23 999bd6f
added distribution property to `MirrorRB`
now users can fully customize the distribution after instantiation. also changed `local_clifford`
to `start_end_clifford` which is clearer.
