+++ test/library/randomized_benchmarking/test_randomized_benchmarking.py
@@ -19,7 +19,7 @@ import numpy as np
 from ddt import ddt, data, unpack
 
 from qiskit.circuit import Delay, QuantumCircuit, Parameter, Gate
-from qiskit.circuit.library import SXGate, CXGate, TGate, CZGate
+from qiskit.circuit.library import SXGate, CXGate, TGate, CZGate, ECRGate
 from qiskit.exceptions import QiskitError
 from qiskit.providers.fake_provider import FakeManila, FakeManilaV2, FakeWashington, FakeParis
 from qiskit.pulse import Schedule, InstructionScheduleMap
@@ -36,7 +36,9 @@ from qiskit_experiments.framework.composite import ParallelExperiment
 from qiskit_experiments.library import randomized_benchmarking as rb
 from qiskit_experiments.library.randomized_benchmarking.clifford_utils import (
     compute_target_bitstring,
+    CliffordUtils,
 )
+from qiskit_experiments.library.randomized_benchmarking.sampling_utils import EdgeGrabSampler
 
 
 class RBTestMixin:
@@ -928,6 +930,8 @@ class NoiseSimulator(AerSimulator):
 class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
     """Test for mirror RB."""
 
+    seed = 123
+
     def setUp(self):
         """Setup the tests."""
         super().setUp()
@@ -946,14 +950,14 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         exp1 = rb.MirrorRB(
             physical_qubits=(0, 1),
             lengths=lengths,
-            seed=123,
+            seed=self.seed,
             backend=self.backend,
         )
 
         exp2 = rb.MirrorRB(
             physical_qubits=(0, 1),
             lengths=lengths,
-            seed=123,
+            seed=self.seed,
             backend=self.backend,
         )
 
@@ -968,7 +972,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         exp1 = rb.MirrorRB(
             physical_qubits=(0, 1),
             lengths=[10, 20],
-            seed=123,
+            seed=self.seed,
             backend=self.backend,
             num_samples=1,
             full_sampling=True,
@@ -977,7 +981,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         exp2 = rb.MirrorRB(
             physical_qubits=(0, 1),
             lengths=[10, 20],
-            seed=123,
+            seed=self.seed,
             backend=self.backend,
             num_samples=1,
             full_sampling=False,
@@ -998,7 +1002,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         exp = rb.MirrorRB(
             physical_qubits=(0, 1),
             lengths=[40],
-            seed=124,
+            seed=self.seed,
             backend=self.backend,
             num_samples=1,
             two_qubit_gate_density=0,
@@ -1017,7 +1021,7 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
         exp = rb.MirrorRB(
             physical_qubits=(0, 1, 2, 3),
             lengths=[40],
-            seed=125,
+            seed=self.seed,
             backend=backend,
             num_samples=1,
             two_qubit_gate_density=0.5,
@@ -1029,16 +1033,16 @@ class TestMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
                 num_cxs += 1
         self.assertEqual(80, num_cxs)
 
-    def test_local_clifford(self):
+    def test_start_end_clifford(self):
         """Test that the number of layers is correct depending on whether
-        local_clifford is set to True or False by counting the number of barriers."""
[…]
