+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -12,9 +12,11 @@
 """
 Mirror RB Experiment class.
 """
+import warnings
 from typing import Union, Iterable, Optional, List, Sequence, Tuple
 from numbers import Integral
 import itertools
+import numpy as np
 from numpy.random import Generator, BitGenerator, SeedSequence, default_rng
 
 from qiskit.circuit import QuantumCircuit, Instruction, Barrier
@@ -40,18 +42,18 @@ class MirrorRB(StandardRB):
     and Pauli gates.
 
     # section: overview
-        Mirror randomized benchmarking (mirror RB) is a method to estimate the average
-        error-rate of quantum gates that is more scalable than the standard RB methods.
+        Mirror randomized benchmarking (mirror RB) estimates the average error rate of
+        quantum gates using gate layers sampled from a distribution that are then
+        inverted in the second half of the circuit.
 
-        A mirror RB experiment generates circuits of layers of Cliffords interleaved
-        with layers of Pauli gates and capped at the start and end by a layer of
-        single-qubit Cliffords. The second half of the Clifford layers are the
-        inverses of the first half of Clifford layers. After running the circuits on
-        a backend, various quantities (success probability, adjusted success
-        probability, and effective polarization) are computed and used to fit an
-        exponential decay curve and calculate the EPC (error per Clifford, also
-        referred to as the average gate infidelity) and entanglement infidelity (see
-        references for more info).
+        The default mirror RB experiment generates circuits of layers of Cliffords
+        interleaved with layers of Pauli gates and capped at the start and end by a
+        layer of single-qubit Cliffords. The second half of the Clifford layers are the
+        inverses of the first half of Clifford layers. After running the circuits on a
+        backend, various quantities (success probability, adjusted success probability,
+        and effective polarization) are computed and used to fit an exponential decay
+        curve and calculate the EPC (error per Clifford, also referred to as the average
+        gate infidelity) and entanglement infidelity (see references for more info).
 
     # section: analysis_ref
         :class:`MirrorRBAnalysis`
@@ -69,7 +71,7 @@ class MirrorRB(StandardRB):
         physical_qubits: Sequence[int],
         lengths: Iterable[int],
         distribution: RBSampler = EdgeGrabSampler,
-        local_clifford: bool = True,
+        start_end_clifford: bool = True,
         pauli_randomize: bool = True,
         two_qubit_gate_density: float = 0.2,
         two_qubit_gate: GateType = CXGate,
@@ -85,12 +87,12 @@ class MirrorRB(StandardRB):
             physical_qubits: A list of physical qubits for the experiment.
             lengths: A list of RB sequences lengths.
             distribution: The probability distribution over the layer set to sample.
-                Defaults to the :class:`.EdgeGrabSampler`. Distribution options can be set by
+                Defaults to :class:`.EdgeGrabSampler`. Distribution options can be set by
                 accessing the distribution object of this class.
-            local_clifford: If True, begin the circuit with uniformly random 1-qubit
+            start_end_clifford: If True, begin the circuit with uniformly random 1-qubit
                 Cliffords and end the circuit with their inverses.
-            pauli_randomize: If True, surround each inner Clifford layer with
-                uniformly random Paulis.
+            pauli_randomize: If True, surround each inner Clifford layer with layers of
+                uniformly random 1-qubit Paulis.
             two_qubit_gate_density: Expected proportion of qubit sites with two-qubit
                 gates over all circuit layers (not counting optional layers at the start
                 and end).
[…]

[…]
