### 2023-03-23 377eaa2
Update test/library/calibration/test_drag.py
Co-authored-by: Will Shanks <wshaos@posteo.net>
