### 2023-03-23 842b9d1
Update test/library/calibration/test_drag.py
Co-authored-by: Will Shanks <wshaos@posteo.net>
