### 2023-03-23 cee42e6
Update docs/_ext/custom_styles/utils.py
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
