### 2023-03-23 887b5a7
Update docs/_ext/custom_styles/utils.py
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
