### 2023-03-23 20ef26e
Fix a bug of redoing DragCalAnalysis (#1094)
### Summary

This PR updates `DRAGCalAnalysis` to fix a bug of redoing analysis.

Fix #1086 

### Details and comments

- Move analysis option `reps` to experiment metadata.
- Completely replace model with one generated on the fly with `reps`
value, rather than appending.

---------

Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
Co-authored-by: Will Shanks <wshaos@posteo.net>
