### 2023-03-23 #5257 @danielskatz
👋 @lcapelluto & @brosand, would one of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html   Overall, the JOSS review process is open and iterative, and aims at ensuring the work can be published by improving it.
