### 2023-03-23 #5257 @danielskatz
👋 @eggerdj - you can change the draft if you want, then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.
