### 2023-03-23 #5257 @goerz
@eggerdj Is this submission a prelude to removing the note in the documentation that “This package is still under active development and it is very likely that there will be breaking API changes in future releases”? Generally, I’d consider a stable API a prerequisite for a publication about a package
