### 2023-03-23 3c49a98
Address comments and remove support for None arguments
