### 2023-03-23 fb23e99
Update tests
Changed tolerances to be based on std_dev.
Optimization level is removed since it's now 1 by default.
