+++ releasenotes/notes/uarray-warning-d4c38566a510e58f.yaml
@@ -1,6 +1,6 @@
 ---
 fixes:
   - |
-    A ``RuntimeWarning`` will no longer be generated by numpy when running a
+    A ``RuntimeWarning`` will no longer be generated by ``numpy`` when running a
     data processor on level one data. See `#1071
-    <https://github.com/Qiskit/qiskit-experiments/issues/1071>_`.
+    <https://github.com/Qiskit/qiskit-experiments/issues/1071>`_.
