+++ releasenotes/notes/docs-refactoring-9f46f6539f57e8bd.yaml
@@ -4,3 +4,6 @@ other:
     The package documentation has been updated with introductory tutorials and how-tos
     for solving specific problems. It is now refactored into four sections: learning
     tutorials, how-to guides, experiment manuals, and the API references.
+  - |
+    The internal curve analysis helper functions in ``curve_analysis.data_processing``
+    has been moved to ``curve_analysis.utils``.
