+++ releasenotes/notes/fix-drag-reanalysis-46f4c6679555242d.yaml
@@ -1,11 +1,10 @@
 ---
-upgrade:
+other:
   - |
-    An analysis option `reps` in :class:`.DragCalAnalysis` was dropped.
-    Now analysis is bootstrapped with circuit metadata and setting this value 
-    no longer impacts the analysis result. 
-    This upgrade doesn't introduce any breaking API change for existing experiments.
+    An analysis option ``reps`` in :class:`.DragCalAnalysis` was dropped. Now analysis
+    is bootstrapped with circuit metadata and setting this value no longer impacts the
+    analysis result. This upgrade doesn't introduce any breaking API change for existing
+    experiments.
 fixes:
   - |
-    A bug of redoing :meth:`.DragCalAnalysis.run` generates wrong fit models
-    was fixed.
+    Fixed a bug where redoing :meth:`.DragCalAnalysis.run` generated wrong fit models.
