+++ releasenotes/notes/ecr_lib-381cb18885e81abd.yaml
@@ -2,7 +2,7 @@
 features:
   - |
     A new basis gate library called :class:`.EchoedCrossResonance` has been added.
-upgrade:
+other:
   - |
     The :class:`.Calibrations` class has been updated to use the reference
     mechanism in Qiskit Pulse in which a schedule can refer to another
