### 2023-03-23 9923e92
Better validation of meas_freqs in resonator_spectrosopy
