### 2023-03-23 babf0ec
Minimize usage of backend.defaults()
