+++ test/library/characterization/test_qubit_spectroscopy.py
@@ -18,6 +18,7 @@ from qiskit.qobj.utils import MeasLevel
 from qiskit.circuit.library import XGate
 from qiskit_experiments.framework import ParallelExperiment
 
+from qiskit_experiments.framework import BackendData
 from qiskit_experiments.library import QubitSpectroscopy, EFSpectroscopy
 from qiskit_experiments.test.mock_iq_backend import MockIQBackend, MockIQParallelBackend
 from qiskit_experiments.test.mock_iq_helpers import MockIQSpectroscopyHelper as SpectroscopyHelper
@@ -43,7 +44,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         backend.target.add_instruction(XGate(), properties={(0,): None})
 
         qubit = 1
-        freq01 = backend.defaults().qubit_freq_est[qubit]
+        freq01 = BackendData(backend).drive_freqs[qubit]
         frequencies = np.linspace(freq01 - 10.0e6, freq01 + 10.0e6, 21)
 
         spec = QubitSpectroscopy([qubit], frequencies)
@@ -53,7 +54,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         result = expdata.analysis_results(1)
         self.assertRoundTripSerializable(result.value, check_func=self.ufloat_equiv)
 
-        self.assertTrue(4.999e9 < result.value.n < 5.001e9)
+        self.assertAlmostEqual(result.value.n, freq01, delta=1e6)
         self.assertEqual(result.quality, "good")
         self.assertEqual(str(result.device_components[0]), f"Q{qubit}")
 
@@ -66,7 +67,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         result = expdata.analysis_results(1)
         self.assertRoundTripSerializable(result.value, check_func=self.ufloat_equiv)
 
-        self.assertTrue(5.0049e9 < result.value.n < 5.0051e9)
+        self.assertAlmostEqual(result.value.n, freq01 + 5e6, delta=1e6)
         self.assertEqual(result.quality, "good")
 
     def test_spectroscopy_end2end_kerneled(self):
@@ -83,7 +84,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         backend.target.add_instruction(XGate(), properties={(0,): None})
 
         qubit = 0
-        freq01 = backend.defaults().qubit_freq_est[qubit]
+        freq01 = BackendData(backend).drive_freqs[qubit]
         frequencies = np.linspace(freq01 - 10.0e6, freq01 + 10.0e6, 21)
 
         spec = QubitSpectroscopy([qubit], frequencies)
@@ -128,7 +129,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         )
         backend.target.add_instruction(XGate(), properties={(0,): None})
         qubit = 0
-        freq01 = backend.defaults().qubit_freq_est[qubit]
+        freq01 = BackendData(backend).drive_freqs[qubit]
         frequencies = np.linspace(freq01 - 10.0e6, freq01 + 10.0e6, 21)
 
         # Note that the backend is not sophisticated enough to simulate an e-f
@@ -175,7 +176,7 @@ class TestQubitSpectroscopy(QiskitExperimentsTestCase):
         backend.target.add_instruction(XGate(), properties={(0,): None})
 
         qubit = 1
-        freq01 = backend.defaults().qubit_freq_est[qubit]
+        freq01 = BackendData(backend).drive_freqs[qubit]
         frequencies = np.linspace(freq01 - 10.0e6, freq01 + 10.0e6, 21)
[…]
