+++ test/calibration/test_calibrations.py
@@ -36,6 +36,7 @@ from qiskit import QuantumCircuit, pulse, transpile
 from qiskit.circuit.library import CXGate, XGate
 from qiskit.pulse.transforms import inline_subroutines, block_to_schedule
 from qiskit.providers import BackendV2, Options
+from qiskit.providers.fake_provider import FakeArmonkV2, FakeBelemV2
 from qiskit.transpiler import Target
 
 from qiskit_experiments.framework import BackendData
@@ -45,7 +46,6 @@ from qiskit_experiments.calibration_management.basis_gate_library import (
     FixedFrequencyTransmon,
 )
 from qiskit_experiments.exceptions import CalibrationError
-from qiskit_experiments.test.fake_pulse_backends import FakeArmonkV2Pulse, FakeBelemV2Pulse
 
 
 class MinimalBackend(BackendV2):
@@ -307,7 +307,7 @@ class TestCalibrationsBasic(QiskitExperimentsTestCase):
     def test_from_backend(self):
         """Test that when generating calibrations from backend
         the data is passed correctly"""
-        backend = FakeBelemV2Pulse()
+        backend = FakeBelemV2()
         cals = Calibrations.from_backend(backend)
         config_args = cals.config()["kwargs"]
         control_channel_map_size = len(config_args["control_channel_map"].chan_map)
@@ -1495,7 +1495,7 @@ class TestSavingAndLoading(CrossResonanceTest):
         """
 
         library = FixedFrequencyTransmon()
-        backend = FakeArmonkV2Pulse()
+        backend = FakeArmonkV2()
         cals = Calibrations.from_backend(backend, libraries=[library])
 
         cals.parameters_table()
@@ -1508,7 +1508,7 @@ class TestSavingAndLoading(CrossResonanceTest):
         self.assertEqual(cals.get_parameter_value("amp", (0,), "x"), 0.5)
         self.assertEqual(
             cals.get_parameter_value("drive_freq", (0,)),
-            backend.defaults().qubit_freq_est[0],
+            BackendData(backend).drive_freqs[0],
         )
 
 
@@ -1519,7 +1519,7 @@ class TestInstructionScheduleMap(QiskitExperimentsTestCase):
         """Test that we can setup with a library."""
 
         cals = Calibrations.from_backend(
-            FakeArmonkV2Pulse(),
+            FakeArmonkV2(),
             libraries=[
                 FixedFrequencyTransmon(basis_gates=["x", "sx"], default_values={"duration": 320})
             ],
@@ -1540,7 +1540,7 @@ class TestInstructionScheduleMap(QiskitExperimentsTestCase):
     def test_instruction_schedule_map_export(self):
         """Test that exporting the inst map works as planned."""
 
-        backend = FakeBelemV2Pulse()
+        backend = FakeBelemV2()
 
         cals = Calibrations.from_backend(
             backend,
@@ -1569,7 +1569,7 @@ class TestInstructionScheduleMap(QiskitExperimentsTestCase):
         """Test that we can use the inst_map to inject the cals into the circuit."""
 
         cals = Calibrations.from_backend(
-            FakeArmonkV2Pulse(),
+            FakeArmonkV2(),
             libraries=[FixedFrequencyTransmon(basis_gates=["x"])],
         )
 
@@ -1619,7 +1619,7 @@ class TestInstructionScheduleMap(QiskitExperimentsTestCase):
         """Test that updating a parameter will force an inst map update."""
 
         cals = Calibrations.from_backend(
-            FakeBelemV2Pulse(),
+            FakeBelemV2(),
             libraries=[FixedFrequencyTransmon(basis_gates=["sx", "x"])],
         )
 
[…]
