+++ test/library/calibration/test_ramsey_xy.py
@@ -94,9 +94,7 @@ class TestRamseyXY(QiskitExperimentsTestCase):
         # Check that qubit frequency after running the cal is shifted by freq_shift, i.e. 4 MHz.
         f01 = self.cals.get_parameter_value(freq_name, 0)
         self.assertTrue(len(self.cals.parameters_table(parameters=[freq_name])["data"]), 2)
-        self.assertLess(
-            abs(f01 - (freq_shift + BackendData(FakeArmonkV2()).drive_freqs[0])), tol
-        )
+        self.assertLess(abs(f01 - (freq_shift + BackendData(FakeArmonkV2()).drive_freqs[0])), tol)
 
     def test_update_with_failed_analysis(self):
         """Test that calibration update handles analysis producing no results
