### 2023-03-23 7224f59
fix nonlocal test
there's an issue with `NoiseSimulator` throwing a coupling map related
error on 2 qubits that seems unrelated to mirror since standard
RB also fails. Changed the coupling map to 2 qubits to avoid this. Also
updated the gate type.
