+++ qiskit_experiments/curve_analysis/base_curve_analysis.py
@@ -160,7 +160,7 @@ class BaseCurveAnalysis(BaseAnalysis, ABC):
             data_processor (Callable): A callback function to format experiment data.
                 This can be a :class:`.DataProcessor`
                 instance that defines the `self.__call__` method.
-            normalization (bool) : Set ``True`` to normalize y values within range [-1, 1].
+            normalization (bool): Set ``True`` to normalize y values within range [-1, 1].
                 Default to ``False``.
             average_method (str): Method to average the y values when the same x values
                 appear multiple times. One of "sample", "iwv" (i.e. inverse weighted variance),
