+++ test/library/randomized_benchmarking/test_rb_experiment.py
@@ -21,6 +21,9 @@ from qiskit.circuit.library import SXGate
 from qiskit.exceptions import QiskitError
 from qiskit.providers.fake_provider import FakeManilaV2
 from qiskit.pulse import Schedule, InstructionScheduleMap
+from qiskit_aer import AerSimulator
+from qiskit_aer.noise import NoiseModel, depolarizing_error
+from qiskit_experiments.framework.composite import ParallelExperiment
 from qiskit_experiments.library import randomized_benchmarking as rb
 
 
@@ -193,3 +196,268 @@ class TestStandardRB(QiskitExperimentsTestCase, RBTestMixin):
             for inst in qc:
                 if inst.operation.name == "cx":
                     self.assertEqual(inst.qubits, expected_qubits)
+
+
+class TestRunStandardRB(QiskitExperimentsTestCase, RBTestMixin):
+    """Test for running StandardRB."""
+
+    def setUp(self):
+        """Setup the tests."""
+        super().setUp()
+
+        # depolarizing error
+        self.p1q = 0.004
+        self.p2q = 0.04
+        self.pvz = 0.0
+        self.pcz = 0.06
+
+        # basis gates
+        self.basis_gates = ["rz", "sx", "cx"]
+
+        # setup noise model
+        sx_error = depolarizing_error(self.p1q, 1)
+        rz_error = depolarizing_error(self.pvz, 1)
+        cx_error = depolarizing_error(self.p2q, 2)
+        cz_error = depolarizing_error(self.pcz, 2)
+
+        noise_model = NoiseModel()
+        noise_model.add_all_qubit_quantum_error(sx_error, "sx")
+        noise_model.add_all_qubit_quantum_error(rz_error, "rz")
+        noise_model.add_all_qubit_quantum_error(cx_error, "cx")
+        noise_model.add_all_qubit_quantum_error(cz_error, "cz")
+
+        self.noise_model = noise_model
+
+        # Need level1 for consecutive gate cancellation for reference EPC value calculation
+        self.transpiler_options = {
+            "basis_gates": self.basis_gates,
+        }
+
+        # Aer simulator
+        self.backend = AerSimulator(noise_model=noise_model, seed_simulator=123)
+
+    def test_single_qubit(self):
+        """Test single qubit RB."""
+        exp = rb.StandardRB(
+            physical_qubits=(0,),
+            lengths=list(range(1, 300, 30)),
+            seed=123,
+            backend=self.backend,
+        )
+        exp.analysis.set_options(gate_error_ratio=None)
+        exp.set_transpile_options(**self.transpiler_options)
+        self.assertAllIdentity(exp.circuits())
+
+        expdata = exp.run()
+        self.assertExperimentDone(expdata)
+
+        # Given we have gate number per Clifford n_gpc, we can compute EPC as
+        # EPC = 1 - (1 - r)^n_gpc
+        # where r is gate error of SX gate, i.e. dep-parameter divided by 2.
+        # We let transpiler use SX and RZ.
+        # The number of physical gate per Clifford will distribute
+        # from 0 to 2, i.e. arbitrary U gate can be decomposed into up to 2 SX with RZs.
+        # We may want to expect the average number of SX is (0 + 1 + 2) / 3 = 1.0.
+        epc = expdata.analysis_results("EPC")
+
+        epc_expected = 1 - (1 - 1 / 2 * self.p1q) ** 1.0
+        self.assertAlmostEqual(epc.value.n, epc_expected, delta=3 * epc.value.std_dev)
+
+    def test_two_qubit(self):
+        """Test two qubit RB. Use default basis gates."""
+        exp = rb.StandardRB(
+            physical_qubits=(0, 1),
+            lengths=list(range(1, 30, 3)),
+            seed=123,
+            backend=self.backend,
+        )
[…]

[…]
