### 2023-03-24 58ec44c
remove banner and update deprecation link
