+++ qiskit_experiments/__init__.py
@@ -17,13 +17,6 @@ Qiskit Experiments (:mod:`qiskit_experiments`)
 
 .. currentmodule:: qiskit_experiments
 
-.. warning::
-
-    This package is still under active development and it is very likely
-    that there will be breaking API changes in future releases.
-    If you encounter any bugs, please open an issue on
-    `GitHub <https://github.com/Qiskit/qiskit-experiments/issues>`_.
-
 Qiskit Experiments provides both a general
 :mod:`~qiskit_experiments.framework` for creating and
 running experiments through Qiskit and optionally storing results in an
