### 2023-03-24 09fe450
address review comments
- removed deprecation warning
- changed two_qubit_gate type to `Instruction`. this required changing
the input to `rng.choice` to a numpy array with `dype=Instruction`
- renamed `RBSampler` to `BaseSampler`
- removed `distribution` experiment option since it's set separately
(may need to refactor this)
