+++ pyproject.toml
@@ -1,6 +1,5 @@
 [build-system]
 requires = ["setuptools", "wheel"]
-build-backend = "setuptools.build_meta"
 
 [tool.black]
 line-length = 100
