+++ tox.ini
@@ -58,19 +58,21 @@ commands = black {posargs} qiskit_experiments test tools setup.py
 [testenv:docs]
 passenv = EXPERIMENTS_DEV_DOCS
 commands =
-  sphinx-build -T --keep-going -b html {posargs} docs/ docs/_build/html
+  sphinx-build -T -W --keep-going -b html {posargs} docs/ docs/_build/html
 
 [testenv:docs-parallel]
+usedevelop = False
 passenv = EXPERIMENTS_DEV_DOCS
 commands =
-  sphinx-build -j auto -T --keep-going -b html {posargs} docs/ docs/_build/html
+  sphinx-build -j auto -T -W --keep-going -b html {posargs} docs/ docs/_build/html
 
 [testenv:docs-minimal]
+usedevelop = False
 passenv = EXPERIMENTS_DEV_DOCS
 setenv = 
   QISKIT_DOCS_SKIP_EXECUTE = 1
 commands =
-  sphinx-build -T --keep-going -b html {posargs} docs/ docs/_build/html
+  sphinx-build -T -W --keep-going -b html {posargs} docs/ docs/_build/html
 
 [pycodestyle]
 max-line-length = 100
