+++ tox.ini
@@ -56,6 +56,7 @@ envdir = .tox/lint
 commands = black {posargs} qiskit_experiments test tools setup.py
 
 [testenv:docs]
+usedevelop = False
 passenv = EXPERIMENTS_DEV_DOCS
 commands =
   sphinx-build -T -W --keep-going -b html {posargs} docs/ docs/_build/html
