### 2023-03-26 4951cb1
change sampler output to list of named tuple
