+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -16,7 +16,7 @@ Utilities for sampling layers in randomized benchmarking experiments
 import warnings
 import math
 from abc import ABC, abstractmethod
-from typing import Optional, Union, List, Tuple, Sequence, TypeVar
+from typing import Optional, Union, List, Tuple, Sequence, TypeVar, NamedTuple
 from collections import defaultdict
 
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
@@ -30,9 +30,19 @@ from .clifford_utils import CliffordUtils, _CLIFF_SINGLE_GATE_MAP_1Q
 GateTypeT = TypeVar("GateTypeT", str, int, Instruction)
 
 
+class GateInstruction(NamedTuple):
+    """Named tuple class for sampler output."""
+
+    # the list of qubits to apply the operation on
+    qargs: tuple
+    # the operation to apply
+    op: GateTypeT
+
+
 class BaseSampler(ABC):
-    """Base class for the sampling distribution for randomized benchmarking experiments.
-    Subclasses must implement the ``__call__()`` method."""
+    """Base class for samplers that generate circuit layers based on a defined
+    algorithm and gate set. Subclasses must implement the ``__call__()`` method
+    which outputs a number of circuit layers."""
 
     def __init__(
         self,
@@ -40,7 +50,8 @@ class BaseSampler(ABC):
         coupling_map: Optional[List[List[int]]] = None,
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
     ):
-        """
+        """Initializes the sampler.
+
         Args:
             seed: Seed for random generation.
             gate_distribution: The gate distribution for sampling.
@@ -94,7 +105,8 @@ class BaseSampler(ABC):
         for gate in dist:
             if gate[2] not in ["clifford", "pauli"] and not isinstance(gate[2], Instruction):
                 raise TypeError(
-                    "The only allowed gates in the distribution are 'clifford', 'pauli', and Instruction instances."
+                    "The only allowed gates in the distribution are 'clifford', 'pauli', "
+                    "and Instruction instances."
                 )
 
         self._gate_distribution = dist
@@ -135,17 +147,19 @@ class BaseSampler(ABC):
         return gate_probs
 
     @abstractmethod
-    def __call__(
-        self, qubits: Sequence, length: int = 1
-    ) -> List[Tuple[Tuple[int, ...], GateTypeT]]:
+    def __call__(self, qubits: Sequence, length: int = 1) -> List[Tuple[GateInstruction]]:
         """Samplers should define this method such that it returns sampled layers
-        given the input parameters. Each layer is represented by a list of size-2 tuples
-        where the first element is a tuple of qubit indices, and the second
-        element represents the gate that should be applied to the indices.
+        given the input parameters. Each layer is represented by a list of
+        ``GateInstruction`` namedtuples, where ``GateInstruction.op`` is the gate to be
+        applied and ``GateInstruction.qargs`` is the tuple of qubit indices to
+        apply the gate to.
 
         Args:
             qubits: A sequence of qubits to generate layers for.
             length: The number of layers to generate. Defaults to 1.
+
+        Returns:
+            A list of layers consisting of GateInstruction objects.
         """
         return None
 
@@ -157,8 +171,9 @@ class SingleQubitSampler(BaseSampler):
         self,
         qubits: Sequence,
         length: int = 1,
-    ) -> List[List[Tuple[Tuple[int, ...], GateTypeT]]]:
-        """Samples random single-qubit gates from the specified gate set.
+    ) -> List[Tuple[GateInstruction]]:
+        """Samples random single-qubit gates from the specified gate set. The
+        input gate distribution must consist solely of single qubit gates.
 
         Args:
             qubits: A sequence of qubits to generate layers for.
[…]
