### 2023-03-26 a12b8b2
Replace target class variables with properties
