### 2023-03-27 9fe27bf
Refactor RB unit tests (#1089)
### Summary

Separate one big `test_randomized_benchmarking.py` file into three files
(`test_standard_rb.py`, `test_interleaved_rb.py` and
`test_rb_analysis.py`) for the faster individual tests for `StandarRB`,
`InterleavedRB` or `RBAnalysis` in local env. And rename
`rb_experiment.py` to `standard_rb.py` to follow naming convention.

### Details and comments
- Rename `rb_experiment.py` to `standard_rb.py`
- Split `test_randomized_benchmarking.py` into `test_standard_rb.py`,
`test_interleaved_rb.py` and `test_rb_analysis.py`

New test files `test_standard_rb.py` 
[…]
