### 2023-03-27 585c053
change gate distribution to named tuple
