+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -16,7 +16,7 @@ Utilities for sampling layers in randomized benchmarking experiments
 import warnings
 import math
 from abc import ABC, abstractmethod
-from typing import Optional, Union, List, Tuple, Sequence, TypeVar, NamedTuple
+from typing import Optional, Union, List, Tuple, Sequence, TypeVar, NamedTuple, Dict
 from collections import defaultdict
 
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
@@ -39,6 +39,17 @@ class GateInstruction(NamedTuple):
     op: GateTypeT
 
 
+class GateDistribution(NamedTuple):
+    """Named tuple class for sampler input distribution."""
+
+    # probability of the instruction to be sampled
+    prob: float
+    # the number of qubits the instruction applies to
+    nq: int
+    # the instruction, can be Instructions or the special keywords "clifford", "pauli"
+    op: Union[Instruction, str]
+
+
 class BaseSampler(ABC):
     """Base class for samplers that generate circuit layers based on a defined
     algorithm and gate set. Subclasses must implement the ``__call__()`` method
@@ -49,7 +60,7 @@ class BaseSampler(ABC):
         gate_distribution=None,
         coupling_map: Optional[List[List[int]]] = None,
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
-    ):
+    ) -> None:
         """Initializes the sampler.
 
         Args:
@@ -61,38 +72,39 @@ class BaseSampler(ABC):
         self.seed = default_rng(seed)
 
     @property
-    def coupling_map(self):
+    def coupling_map(self) -> Optional[List[List[int]]]:
         """The coupling map of the system to sample over."""
         return self._coupling_map
 
     @coupling_map.setter
-    def coupling_map(self, coupling_map):
+    def coupling_map(self, coupling_map) -> None:
         self._coupling_map = coupling_map
 
     @property
-    def seed(self):
+    def seed(self) -> Union[int, SeedSequence, BitGenerator, Generator]:
         """The seed for random generation."""
         return self._rng
 
     @seed.setter
-    def seed(self, seed):
+    def seed(self, seed) -> None:
         self._rng = default_rng(seed)
 
     @property
-    def gate_distribution(self):
-        """The gate distribution for sampling. The distribution is a list of tuples
-        of the form ``(probability, width of gate, gate)``. Gates can be actual
-        operators, circuits, or the special keywords "clifford", "pauli", "idle". An
-        example distribution for the edge grab sampler might be
+    def gate_distribution(self) -> List[GateDistribution]:
+        """The gate distribution for sampling. The distribution is a list of
+        ``GateDistribution`` named tuples with field names ``(prob, nq, op)``,
+        where the probabilites must sum to 1 and ``nq`` is the number of qubits
+        the ``op`` applies to. Gates can be ``Instruction`` or the special
+        keywords "clifford", "pauli". An example distribution for the edge grab
+        sampler might be
 
         .. parsed-literal::
             [(0.8, 1, "clifford"), (0.2, 2, CXGate())]
-
-        The probabilities must sum to 1."""
+        """
         return self._gate_distribution
 
     @gate_distribution.setter
-    def gate_distribution(self, dist: List[Tuple[float, int, GateTypeT]]):
+    def gate_distribution(self, dist: List[GateDistribution]) -> None:
         """Set the distribution of gates used in the sampler.
 
         Args:
@@ -100,10 +112,12 @@ class BaseSampler(ABC):
         """
[…]
