### 2023-03-27 95b9274
Update FakeBackend to use BackendV2 (#1102)
### Summary

This change set updates the `FakeBackend` used by some tests to derive
from `BackendV2` rather than `BackendV1`

### Details and comments

When combined with
[#900](https://github.com/Qiskit/qiskit-experiments/pull/900) which
updates some calibration tests and
[#1040](https://github.com/Qiskit/qiskit-experiments/pull/1040) which
updates `T2HahnBackend`, this change results in almost all the tests in
the test suite using `BackendV2`. The remaining tests using a `Backend`
that is not `BackendV2` use the `AerSimulator` which will be updated
[…]
