### 2023-03-27 44cd94d
Add docstrings about the limited use of defaults()
