### 2023-03-27 479f89e
Update calibration and mixin tests to work with BackendV2 (#900)
### Summary

This change continues the work started in #843 to make
qiskit-experiments compatible with BackendV2. It updates the
spectrscopy, fine frequency and amplitude, and T1 tests to use
`BackendV2` rather than `BackendV1`. Additionally it adds support for
querying `BackendV2` for measurement frequencies.

### Details and comments

The main changes are:

* Use `backend.defaults()` to look up `measure_freq_est` for `BackendV2`
as well as `BackendV1`.
* Improve the error message when `measure_freq_est` can not be accessed
for 
[…]
