### 2023-03-27 3812049
refactored experiment options
now `sampling_algorithm` specifies the algorithm in str form and
`sampler_opts` is passed at sampler instantiation. two-qubit gate and
density are only passed to the default edge grab sampler,
otherwise users have to implement their own methods.
also addressed some review comments.
