+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -23,11 +23,25 @@ from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
 import numpy as np
 
 from qiskit.circuit import Instruction
+from qiskit.circuit.gate import Gate
 from qiskit.exceptions import QiskitError
+from qiskit.transpiler import CouplingMap
 
 from .clifford_utils import CliffordUtils, _CLIFF_SINGLE_GATE_MAP_1Q
 
-GateTypeT = TypeVar("GateTypeT", str, int, Instruction)
+
+class GenericClifford(Gate):
+    """Representation of a generic multi-qubit Clifford gate for sampling."""
+
+    def __init__(self, n_qubits):
+        super().__init__("generic_clifford", n_qubits, [])
+
+
+class GenericPauli(Gate):
+    """Representation of a generic multi-qubit Pauli gate for sampling."""
+
+    def __init__(self, n_qubits):
+        super().__init__("generic_pauli", n_qubits, [])
 
 
 class GateInstruction(NamedTuple):
@@ -36,18 +50,16 @@ class GateInstruction(NamedTuple):
     # the list of qubits to apply the operation on
     qargs: tuple
     # the operation to apply
-    op: GateTypeT
+    op: Instruction
 
 
 class GateDistribution(NamedTuple):
     """Named tuple class for sampler input distribution."""
 
-    # probability of the instruction to be sampled
+    # probability with which to sample the instruction
     prob: float
-    # the number of qubits the instruction applies to
-    nq: int
-    # the instruction, can be Instructions or the special keywords "clifford", "pauli"
-    op: Union[Instruction, str]
+    # the instruction to include in sampling
+    op: Instruction
 
 
 class BaseSampler(ABC):
@@ -57,8 +69,6 @@ class BaseSampler(ABC):
 
     def __init__(
         self,
-        gate_distribution=None,
-        coupling_map: Optional[List[List[int]]] = None,
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
     ) -> None:
         """Initializes the sampler.
@@ -67,18 +77,7 @@ class BaseSampler(ABC):
             seed: Seed for random generation.
             gate_distribution: The gate distribution for sampling.
         """
-        self._gate_distribution = gate_distribution
-        self._coupling_map = coupling_map
-        self.seed = default_rng(seed)
-
-    @property
-    def coupling_map(self) -> Optional[List[List[int]]]:
-        """The coupling map of the system to sample over."""
-        return self._coupling_map
-
-    @coupling_map.setter
-    def coupling_map(self, coupling_map) -> None:
-        self._coupling_map = coupling_map
+        self.seed = seed
 
     @property
     def seed(self) -> Union[int, SeedSequence, BitGenerator, Generator]:
@@ -92,14 +91,12 @@ class BaseSampler(ABC):
     @property
     def gate_distribution(self) -> List[GateDistribution]:
         """The gate distribution for sampling. The distribution is a list of
-        ``GateDistribution`` named tuples with field names ``(prob, nq, op)``,
-        where the probabilites must sum to 1 and ``nq`` is the number of qubits
-        the ``op`` applies to. Gates can be ``Instruction`` or the special
-        keywords "clifford", "pauli". An example distribution for the edge grab
-        sampler might be
+        ``GateDistribution`` named tuples with field names ``(prob, op)``, where
+        the probabilites must sum to 1 and ``op`` is the Instruction instance to
+        be sampled. An example distribution for the edge grab sampler is
 
         .. parsed-literal::
-            [(0.8, 1, "clifford"), (0.2, 2, CXGate())]
[…]
