+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_experiment.py
@@ -34,12 +34,12 @@ from .clifford_utils import (
     _clifford_1q_int_to_instruction,
 )
 from .sampling_utils import (
-    BaseSampler,
     EdgeGrabSampler,
     SingleQubitSampler,
-    GateTypeT,
     GateInstruction,
     GateDistribution,
+    GenericClifford,
+    GenericPauli,
 )
 
 # two qubit gates that are their own inverse
@@ -47,8 +47,8 @@ _self_adjoint_gates = [CXGate, CYGate, CZGate, ECRGate, SwapGate, iSwapGate]
 
 
 class MirrorRB(StandardRB):
-    """An experiment to measure gate infidelity using random mirrored layers of Clifford
-    and Pauli gates.
+    """An experiment to measure gate infidelity using mirrored circuit
+    layers sampled from a defined distribution.
 
     # section: overview
         Mirror randomized benchmarking (mirror RB) estimates the average error rate of
@@ -78,15 +78,18 @@ class MirrorRB(StandardRB):
 
     """
 
+    sampler_map = {"edge_grab": EdgeGrabSampler, "single_qubit": SingleQubitSampler}
+
     def __init__(
         self,
         physical_qubits: Sequence[int],
         lengths: Iterable[int],
-        distribution: BaseSampler = EdgeGrabSampler,
         start_end_clifford: bool = True,
         pauli_randomize: bool = True,
+        sampling_algorithm: str = "edge_grab",
         two_qubit_gate_density: float = 0.2,
         two_qubit_gate: Instruction = CXGate(),
+        sampler_opts: dict = {},
         backend: Optional[Backend] = None,
         num_samples: int = 3,
         seed: Optional[Union[int, SeedSequence, BitGenerator, Generator]] = None,
@@ -98,18 +101,20 @@ class MirrorRB(StandardRB):
         Args:
             physical_qubits: A list of physical qubits for the experiment.
             lengths: A list of RB sequences lengths.
-            distribution: The probability distribution over the layer set to sample.
-                Defaults to :class:`.EdgeGrabSampler`. Distribution options can be set by
-                accessing the distribution object of this class.
+            sampling_algorithm: The sampling algorithm to use for generating
+                circuit layers. Defaults to "edge_grab" which uses :class:`.EdgeGrabSampler`.
             start_end_clifford: If True, begin the circuit with uniformly random 1-qubit
                 Cliffords and end the circuit with their inverses.
-            pauli_randomize: If True, surround each inner Clifford layer with layers of
+            pauli_randomize: If True, surround each sampled circuit layer with layers of
                 uniformly random 1-qubit Paulis.
             two_qubit_gate_density: Expected proportion of qubit sites with two-qubit
                 gates over all circuit layers (not counting optional layers at the start
-                and end).
+                and end). Only has effect if the default sampler
+                :class:`.EdgeGrabSampler` is used.
             two_qubit_gate: The two-qubit gate to use. Defaults to
-                :class:`~qiskit.circuit.library.CXGate`.
+                :class:`~qiskit.circuit.library.CXGate`. Only has effect if the
+                default sampler :class:`.EdgeGrabSampler` is used.
+            sampler_opts: Dictionary of keyword arguments to pass to the sampler.
             backend: The backend to run the experiment on.
             num_samples: Number of samples to generate for each sequence length.
             seed: Optional, seed used to initialize ``numpy.random.default_rng``.
@@ -142,15 +147,16 @@ class MirrorRB(StandardRB):
         )
 
         self.set_experiment_options(
+            sampling_algorithm=sampling_algorithm,
+            sampler_opts=sampler_opts,
             start_end_clifford=start_end_clifford,
             pauli_randomize=pauli_randomize,
-            inverting_pauli_layer=inverting_pauli_layer,
[…]

[…]
