### 2023-03-27 dd08c0c
Fix doc build on Macs (#1105)
### Summary

This PR follows https://github.com/Qiskit/qiskit-terra/pull/9765 and
turns off development mode for tox doc builds because of a known issue
on Macs where some pages output build warnings and don't build properly.
Now Mac doc builds should work fully.
