+++ .github/workflows/main.yml
@@ -97,7 +97,7 @@ jobs:
           python -m pip install -U tox
           sudo apt-get install -y pandoc graphviz
       - name: Build Docs
-        run: tox -edocs -- -W
+        run: tox -edocs
       - name: Compress Artifacts
         run: |
           mkdir artifacts
