### 2023-03-27 8135fa9
fixes and removing unnecessary notes
