+++ releasenotes/notes/0.5/update-number-to-2q-clifford-mapping-c28f1f29b0205d57.yaml
@@ -1,13 +1,4 @@
 ---
-fixes:
-  - |
-    Fixed a bug where :meth:`.CliffordUtils.random_cliffords` always returned a single Clifford
-    ignoring the ``size`` arguments. Now it returns a list of Cliffords with ``size``.
-  - |
-    Fixed a bug where both :meth:`.CliffordUtils.random_cliffords` and
-    :meth:`.CliffordUtils.random_clifford_circuits` with integer ``seed``
-    returned a list of the same circuits (sampled with a common seed).
-    Now it returns a list of circuits sampled with different seeds.
 deprecations:
   - |
     Two helper methods :meth:`.CliffordUtils.random_cliffords` and
