+++ releasenotes/notes/0.5/update-cr-hamtomo-with-duration-380da3452045cd0c.yaml
@@ -29,6 +29,5 @@ deprecations:
     have been deprecated and will be removed in Qiskit Experiments 0.6.
 fixes:
   - |
-    There has been a bug in the :class:`.EchoedCrossResonanceHamiltonian` experiment 
-    in which the Hamiltonian coefficients were overestimated by a factor of 2.
-    This bug has been fixed.
+    Fixe a bug where the :class:`.EchoedCrossResonanceHamiltonian` experiment 
+    overestimated Hamiltonian coefficients by a factor of 2.
