+++ releasenotes/notes/0.5/suppress-runtime-warning-a741dc96f6a0ce7a.yaml
@@ -1,7 +1,7 @@
 ---
 other:
   - |
-    NumPy runtime warning for zero division is suppressed in :class:`.CurveAnalysis`.
+    NumPy runtime warning for zero division has been suppressed in :class:`.CurveAnalysis`.
     This warning could occur in the edge case where the experiment data
     may contain data point with zero uncertainty.
     Such data point is safely ignored by LMFIT, since it may apply infinite fit weight.
