+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_analysis.py
@@ -12,8 +12,7 @@
 """
 Mirror RB analysis class.
 """
-from collections import defaultdict
-from typing import Dict, List, Sequence, Tuple, Union, Optional, TYPE_CHECKING
+from typing import Dict, List, Tuple, Union, TYPE_CHECKING
 
 import lmfit
 from qiskit.exceptions import QiskitError
@@ -22,11 +21,11 @@ import numpy as np
 from scipy.spatial.distance import hamming
 
 import qiskit_experiments.curve_analysis as curve
-from qiskit_experiments.exceptions import AnalysisError
-from qiskit_experiments.framework import AnalysisResultData, ExperimentData
-from qiskit_experiments.framework.analysis_result import AnalysisResult
+from qiskit_experiments.framework import AnalysisResultData
 from qiskit_experiments.data_processing.exceptions import DataProcessorError
 from uncertainties import unumpy as unp  # pylint: disable=wrong-import-order
+
+from .rb_analysis import RBAnalysis
 
 if TYPE_CHECKING:
     from uncertainties import UFloat
@@ -35,7 +34,7 @@ if TYPE_CHECKING:
 QubitGateTuple = Tuple[Tuple[int, ...], str]
 
 
-class MirrorRBAnalysis(curve.CurveAnalysis):
+class MirrorRBAnalysis(RBAnalysis):
     r"""A class to analyze mirror randomized benchmarking experiment.
 
     # section: overview
@@ -103,14 +102,6 @@ class MirrorRBAnalysis(curve.CurveAnalysis):
         .. ref_arxiv:: 1 2112.09853
 
     """
-
-    def __init__(self):
-        super().__init__(
-            models=[lmfit.models.ExpressionModel(expr="a * alpha ** x + b", name="mirror")]
-        )
-        self._gate_counts_per_clifford = None
-        self._physical_qubits = None
-        self._num_qubits = None
 
     @classmethod
     def _default_options(cls):
@@ -183,14 +174,16 @@ class MirrorRBAnalysis(curve.CurveAnalysis):
         Returns:
             List of fit options that are passed to the fitter function.
         """
+
         user_opt.bounds.set_if_empty(a=(0, 1), alpha=(0, 1), b=(0, 1))
+        num_qubits = len(self._physical_qubits)
 
         # Initialize guess for baseline and amplitude based on infidelity type
-        b_guess = 1 / 4**self._num_qubits
+        b_guess = 1 / 4**num_qubits
         if self.options.y_axis == "Success Probability":
-            b_guess = 1 / 2**self._num_qubits
+            b_guess = 1 / 2**num_qubits
 
-        mirror_curve = curve_data.get_subset_of("mirror")
+        mirror_curve = curve_data.get_subset_of("rb_decay")
         alpha_mirror = curve.guess.rb_decay(mirror_curve.x, mirror_curve.y, b=b_guess)
         a_guess = (curve_data.y[0] - b_guess) / (alpha_mirror ** curve_data.x[0])
 
@@ -246,7 +239,8 @@ class MirrorRBAnalysis(curve.CurveAnalysis):
         quality: str,
         **metadata,
     ) -> List[AnalysisResultData]:
-        """Create analysis results for important fit parameters.
+        """Create analysis results for important fit parameters. Besides the
+        default standard RB parameters, Entanglement Infidelity (EI) is also calculated.
 
         Args:
             fit_data: Fit outcome.
@@ -260,65 +254,15 @@ class MirrorRBAnalysis(curve.CurveAnalysis):
         num_qubits = len(self._physical_qubits)
 
         # nrb is calculated for both EPC and EI per the equations in the docstring
[…]
