+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -16,7 +16,7 @@ Utilities for sampling layers in randomized benchmarking experiments
 import warnings
 import math
 from abc import ABC, abstractmethod
-from typing import Optional, Union, List, Tuple, Sequence, TypeVar, NamedTuple, Dict
+from typing import Optional, Union, List, Tuple, Sequence, NamedTuple, Dict
 from collections import defaultdict
 
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence

[…]
