### 2023-03-28 b71d098
change y-axis option to `analyzed_quantity`
