### 2023-03-28 0510478
changed to `CouplingMap` type and added validation
Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>
