+++ test/library/randomized_benchmarking/test_mirror_rb.py
@@ -635,7 +635,7 @@ class TestRunMirrorRB(QiskitExperimentsTestCase, RBTestMixin):
     )
     def test_invalid_configuration(self, configs):
         """Test raise error when creating experiment with invalid configs."""
-        self.assertRaises(QiskitError, rb.MirrorRB, **configs)
+        self.assertRaises((QiskitError, ValueError), rb.MirrorRB, **configs)
 
     @data(
         {
