### 2023-03-28 1c98049
added data processor node for analysis
and added validation for analysis option
