+++ qiskit_experiments/library/randomized_benchmarking/mirror_rb_analysis.py
@@ -1,6 +1,6 @@
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2021.
+# (C) Copyright IBM 2023.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -12,26 +12,17 @@
 """
 Mirror RB analysis class.
 """
-from typing import Dict, List, Tuple, Union, TYPE_CHECKING
-
-import lmfit
-from qiskit.exceptions import QiskitError
-
+from typing import List, Union
 import numpy as np
+from uncertainties import unumpy as unp
 from scipy.spatial.distance import hamming
 
 import qiskit_experiments.curve_analysis as curve
-from qiskit_experiments.framework import AnalysisResultData
-from qiskit_experiments.data_processing.exceptions import DataProcessorError
-from uncertainties import unumpy as unp  # pylint: disable=wrong-import-order
+from qiskit_experiments.framework import AnalysisResultData, ExperimentData
+from qiskit_experiments.data_processing import DataProcessor
+from qiskit_experiments.data_processing.data_action import DataAction
 
 from .rb_analysis import RBAnalysis
-
-if TYPE_CHECKING:
-    from uncertainties import UFloat
-
-# A dictionary key of qubit aware quantum instruction; type alias for better readability
-QubitGateTuple = Tuple[Tuple[int, ...], str]
 
 
 class MirrorRBAnalysis(RBAnalysis):
@@ -145,20 +136,16 @@ class MirrorRBAnalysis(RBAnalysis):
         # Do this by setting options to "Success Probability" or "Adjusted Success Probability"
         default_options.analyzed_quantity = "Effective Polarization"
 
-        return default_options
-
-    def set_options(self, **fields):
-        if "analyzed_quantity" in fields:
-            if fields["analyzed_quantity"] not in [
+        default_options.set_validator(
+            field="analyzed_quantity",
+            validator_value=[
                 "Success Probability",
                 "Adjusted Success Probability",
                 "Effective Polarization",
-            ]:
-                raise QiskitError(
-                    'analyzed_quantity must be one of "Success Probability", "Adjusted Success Probability", '
-                    'or "Effective Polarization"'
-                )
-        super().set_options(**fields)
+            ],
+        )
+
+        return default_options
 
     def _generate_fit_guesses(
         self,
@@ -190,48 +177,6 @@ class MirrorRBAnalysis(RBAnalysis):
         user_opt.p0.set_if_empty(b=b_guess, a=a_guess, alpha=alpha_mirror)
 
         return user_opt
-
-    def _format_data(
-        self,
-        curve_data: curve.CurveData,
-    ) -> curve.CurveData:
-        """Postprocessing for the processed dataset.
-
-        Args:
-            curve_data: Processed dataset created from experiment results.
-
-        Returns:
-            Formatted data.
-        """
-        # TODO Eventually move this to data processor, then create RB data processor.
-
-        # take average over the same x value by keeping sigma
-        data_allocation, xdata, ydata, sigma, shots = curve.utils.multi_mean_xy_data(
-            series=curve_data.data_allocation,
-            xdata=curve_data.x,
-            ydata=curve_data.y,
-            sigma=curve_data.y_err,
-            shots=curve_data.shots,
-            method="sample",
-        )
-
-        # sort by x value in ascending order
-        data_allocation, xdata, ydata, sigma, shots = curve.utils.data_sort(
-            series=data_allocation,
-            xdata=xdata,
-            ydata=ydata,
-            sigma=sigma,
-            shots=shots,
-        )
-
-        return curve.CurveData(
[…]
