### 2023-03-28 fb41d6d
change sampler to return a generator
