+++ qiskit_experiments/library/randomized_benchmarking/sampling_utils.py
@@ -18,6 +18,7 @@ import math
 from abc import ABC, abstractmethod
 from typing import Optional, Union, List, Tuple, Sequence, NamedTuple, Dict
 from collections import defaultdict
+from typing import Iterator
 from numpy.random import Generator, default_rng, BitGenerator, SeedSequence
 import numpy as np
 
@@ -120,7 +121,6 @@ class BaseSampler(ABC):
                 raise TypeError(
                     "The only allowed gates in the distribution are Instruction instances."
                 )
-
         self._gate_distribution = dist
 
     def _probs_by_gate_size(self, distribution: Sequence[GateDistribution]) -> Dict:
@@ -168,7 +168,7 @@ class BaseSampler(ABC):
         return gate_probs
 
     @abstractmethod
-    def __call__(self, qubits: Sequence, length: int = 1) -> List[Tuple[GateInstruction, ...]]:
+    def __call__(self, qubits: Sequence, length: int = 1) -> Iterator[Tuple[GateInstruction, ...]]:
         """Samplers should define this method such that it returns sampled layers
         given the input parameters. Each layer is represented by a list of
         ``GateInstruction`` namedtuples, where ``GateInstruction.op`` is the gate to be
@@ -180,7 +180,7 @@ class BaseSampler(ABC):
             length: The number of layers to generate. Defaults to 1.
 
         Returns:
-            A list of layers consisting of GateInstruction objects.
+            A generator of layers consisting of GateInstruction objects.
         """
         raise NotImplementedError
 
@@ -188,11 +188,26 @@ class BaseSampler(ABC):
 class SingleQubitSampler(BaseSampler):
     """A sampler that samples layers of random single-qubit gates from a specified gate set."""
 
+    @BaseSampler.gate_distribution.setter
+    def gate_distribution(self, dist: List[GateDistribution]) -> None:
+        """Set the distribution of gates used in the sampler.
+
+        Args:
+            dist: A list of tuples with format ``(probability, gate)``.
+        """
+        super(SingleQubitSampler, type(self)).gate_distribution.fset(self, dist)
+
+        gateset = self._probs_by_gate_size(dist)
+        if not math.isclose(sum(gateset[1][1]), 1):
+            raise QiskitError(
+                "The distribution for SingleQubitSampler should be all single qubit gates."
+            )
+
     def __call__(
         self,
         qubits: Sequence,
         length: int = 1,
-    ) -> List[Tuple[GateInstruction]]:
+    ) -> Iterator[Tuple[GateInstruction]]:
         """Samples random single-qubit gates from the specified gate set. The
         input gate distribution must consist solely of single qubit gates.
 
@@ -201,7 +216,7 @@ class SingleQubitSampler(BaseSampler):
             length: The length of the sequence to output.
 
         Returns:
-            A ``length``-long list of :class:`qiskit.circuit.QuantumCircuit`
+            A ``length``-long iterator of :class:`qiskit.circuit.QuantumCircuit`
             layers over ``qubits``. Each layer is represented by a list of
             ``GateInstruction`` tuples where ``GateInstruction.op`` is the gate
             to be applied and ``GateInstruction.qargs`` is the tuple of qubit
@@ -210,10 +225,6 @@ class SingleQubitSampler(BaseSampler):
         """
 
         gateset = self._probs_by_gate_size(self._gate_distribution)
-        if not math.isclose(sum(gateset[1][1]), 1):
-            raise QiskitError(
-                "The distribution for SingleQubitSampler should be all single qubit gates."
-            )
 
         samples = self._rng.choice(
             np.array(gateset[1][0], dtype=object),
[…]
