### 2023-03-28 014aca1
Different handling of provider to enable better data retrievel on load
