### 2023-03-28 884cee0
setting `end_datetime` when a job terminates.
