### 2023-03-28 5c9fc17
`_add_result_data` can now handle missing `job_id`
