+++ docs/tutorials/curve_analysis.rst
@@ -6,11 +6,11 @@ Curve Analysis: Fitting your data
 For most experiments, we are interested in fitting our results to a pre-defined 
 mathematical model.
 The Curve Analysis module provides the analysis base class for a variety of experiments with
-a single experimental parameter sweep. This analysis subclasses can override
+a single experimental parameter sweep. Analysis subclasses can override
 several class attributes to customize the behavior from data processing to post-processing,
-including providing systematic initial guess for parameters tailored to the experiment.
+including providing systematic initial guesses for parameters tailored to the experiment.
 Here we describe how the Curve Analysis module works and how you can create new 
-analyses that inherits from the base class.
+analyses that inherit from the base class.
 
 
 .. _curve_analysis_overview:
@@ -23,7 +23,7 @@ different sets of experiment results. A single experiment can define sub-experim
 consisting of multiple circuits which are tagged with common metadata,
 and curve analysis sorts the experiment results based on the circuit metadata.
 
-This is an example of showing the abstract data structure of typical curve analysis experiment:
+This is an example of showing the abstract data structure of a typical curve analysis experiment:
 
 .. jupyter-input::
     :emphasize-lines: 1,10,19
@@ -51,7 +51,7 @@ This is an example of showing the abstract data structure of typical curve analy
         - "series_B": y_B = f_B(x_B; p0, p1, p2)
         - fixed parameters {p1: v}
 
-Here the experiment runs two subset of experiments, namely, series A and series B.
+Here the experiment runs two subset ofs experiments, namely, series A and series B.
 The analysis defines corresponding fit models :math:`f_A(x_A)` and :math:`f_B(x_B)`.
 Data extraction function in the analysis creates two datasets, :math:`(x_A, y_A)`
 for the series A and :math:`(x_B, y_B)` for the series B, from the experiment data.
@@ -74,7 +74,7 @@ This objective function can be described by two fit functions as follows.
 
 The solver conducts the least square curve fitting against this objective function
 and returns the estimated parameters :math:`\Theta_{\mbox{opt}}`
-that minimizes the reduced chi-squared value.
+that minimize the reduced chi-squared value.
 The parameters to be evaluated are :math:`\Theta = \Theta_{\rm fit} \cup \Theta_{\rm fix}`,
 where :math:`\Theta_{\rm fit} = \theta_A \cup \theta_B`.
 Since series A and B share the parameters in this example, :math:`\Theta_{\rm fit} = \{p_0, p_2\}`,
@@ -82,14 +82,14 @@ and the fixed parameters are :math:`\Theta_{\rm fix} = \{ p_1 \}` as mentioned.
 Thus, :math:`\Theta = \{ p_0, p_1, p_2 \}`.
 
 Experiment for each series can perform individual parameter sweep for :math:`x_A` and
-:math:`x_B`, and experiment data yield outcomes :math:`y_A` and :math:`y_B`, which might
-be different size. Data processing function may also compute :math:`\sigma_A` and
-:math:`\sigma_B` which are the uncertainty of outcomes arising from the sampling error
+:math:`x_B`, and experiment data yields outcomes :math:`y_A` and :math:`y_B`, which might
+be of different size. Data processing functions may also compute :math:`\sigma_A` and
+:math:`\sigma_B`, which are the uncertainty of outcomes arising from the sampling error
 or measurement error.
 
-More specifically, the curve analysis defines following data model.
+More specifically, the curve analysis defines the following data model.
 
-- Model: Definition of a single curve that is a function of reserved parameter "x".
+- Model: Definition of a single curve that is a function of a reserved parameter "x".
 
 - Group: List of models. Fit functions defined under the same group must share the
[…]
