### 2023-03-28 f01b2d6
Update deprecation policy (#1101)
### Summary

Now that the API is relatively stable, this PR removes the banner and.
Going forward, we'll abide by the [Qiskit deprecation
policy](https://qiskit.org/documentation/deprecation_policy.html) when
making public-facing changes.
