+++ docs/apidocs/index.rst
@@ -1,13 +1,6 @@
 API Reference
 =============
 
-.. warning::
-
-    This package is still under active development and it is very likely
-    that there will be breaking API changes in future releases.
-    If you encounter any bugs, please open an issue on
-    `GitHub <https://github.com/Qiskit/qiskit-experiments/issues>`_.
-
 The API documentation is organized into two sections below. The package modules include the framework, the
 experiment library, experiment modules, and test utilities. Experiment modules are 
 the main categories of the experiment library itself, such as qubit characterization
