### 2023-03-28 164a765
Apply suggestions from code review
Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>
