+++ releasenotes/notes/0.5/update-number-to-2q-clifford-mapping-c28f1f29b0205d57.yaml
@@ -8,7 +8,7 @@ other:
   - |
     :meth:`.CliffordUtils.clifford_2_qubit` (and :meth:`.CliffordUtils.clifford_2_qubit_circuit`)
     changed its mapping between integers and 2Q Cliffords.
-    That means 2Q RB experiments may sample different set of circuits from before
+    As a consequence, circuits sampled by 2Q RB experiments may have been changed,
     even if exactly the same arguments are used for their construction.
   - |
     Removed unnecessary ``Barrier`` instructions in front of circuits generated by
