+++ releasenotes/notes/0.5/readout-error-c95b99ae5a6ba7ac.yaml
@@ -8,7 +8,7 @@ features:
     qubits for the backend.
 fixes:
   - |
-    Fixed bug in the :class:`~.LocalReadoutError` experiment where analysis
+    Fixed a bug in the :class:`~.LocalReadoutError` experiment where analysis
     would fail when run on an ideal simulator with no readout error. See 
     `Issue #992 <https://github.com/Qiskit/qiskit-experiments/issues/992>`_
     for additional details.
