+++ releasenotes/notes/0.5/fix-missing_calibration_updator_call-a255b28dd1449ea4.yaml
@@ -2,5 +2,5 @@
 
 fixes:
   - |
-    Fixed a bug where :class:`.Calibrations` was not updated when you ran calibration 
-    experiments through the composite experiment framework.
+    Fixed a bug where :class:`.Calibrations` was not updated when calibration 
+    experiments were run through the composite experiment framework.
