### 2023-03-28 f8b7bed
Update docs/tutorials/curve_analysis.rst
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
