### 2023-03-28 18eae36
Update docs/tutorials/curve_analysis.rst
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
