### 2023-03-28 8348865
Update docs/tutorials/curve_analysis.rst
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
