### 2023-03-28 c81683e
Update docs/tutorials/curve_analysis.rst
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
