+++ test/library/randomized_benchmarking/test_sampling_utils.py
@@ -15,8 +15,10 @@ Tests for RB sampling utils.
 """
 
 from test.base import QiskitExperimentsTestCase
+from ddt import ddt, data
 
-from qiskit.circuit.library import XGate, CXGate
+from qiskit.circuit.library import XGate, CXGate, CCXGate
+from qiskit.exceptions import QiskitError
 
 from qiskit_experiments.library.randomized_benchmarking.sampling_utils import (
     SingleQubitSampler,
@@ -26,6 +28,7 @@ from qiskit_experiments.library.randomized_benchmarking.sampling_utils import (
 from qiskit_experiments.library.randomized_benchmarking.clifford_utils import CliffordUtils
 
 
+@ddt
 class TestSamplingUtils(QiskitExperimentsTestCase):
     """Tests for the Sampler classes."""
 
@@ -34,18 +37,17 @@ class TestSamplingUtils(QiskitExperimentsTestCase):
     def test_gate_distribution(self):
         """Test the gate distribution is calculated correctly."""
         sampler = SingleQubitSampler(seed=self.seed)
-        sampler.gate_distribution = [(0.8, GenericClifford(1)), (0.2, CXGate())]
+        sampler.gate_distribution = [(0.8, GenericClifford(1)), (0.2, XGate())]
         dist = sampler._probs_by_gate_size(sampler.gate_distribution)
-        self.assertEqual(len(dist[1][0]), 24)
-        self.assertAlmostEqual(sum(dist[1][1]), 0.8)
-        self.assertEqual(dist[2][0], [CXGate()])
-        self.assertAlmostEqual(sum(dist[2][1]), 0.2)
+        self.assertEqual(len(dist[1][0]), 25)
+        for i in dist[1][0]:
+            self.assertTrue(i == XGate() or 0 <= i < CliffordUtils.NUM_CLIFFORD_1_QUBIT)
 
     def test_1q_custom_gate(self):
         """Test that the single qubit sampler works with custom gates."""
         sampler = SingleQubitSampler(seed=self.seed)
         sampler.gate_distribution = [(1, XGate())]
-        layer = sampler((0,), 3)
+        layer = list(sampler((0,), 3))
         self.assertEqual(
             layer,
             [(((0,), XGate()),), (((0,), XGate()),), (((0,), XGate()),)],
@@ -66,8 +68,7 @@ class TestSamplingUtils(QiskitExperimentsTestCase):
         layer = sampler((0,), 3)
         for i in layer:
             self.assertTrue(
-                (i[0].op < CliffordUtils.NUM_CLIFFORD_1_QUBIT and i[0].op >= 0)
-                or i[0].op == CXGate()
+                0 <= i[0].op < CliffordUtils.NUM_CLIFFORD_1_QUBIT or i[0].op == CXGate()
             )
 
     def test_edgegrab_all_2q(self):
@@ -79,3 +80,22 @@ class TestSamplingUtils(QiskitExperimentsTestCase):
         layer = sampler(range(10), 3)
         for i in layer:
             self.assertTrue(i[0].op == CXGate())
+
+    def test_invalid_1q_distribution(self):
+        """Test that the single qubit sampler rejects incorrect
+        distributions."""
+        sampler = SingleQubitSampler(seed=self.seed)
+        with self.assertRaises(QiskitError):
+            sampler.gate_distribution = [(0.8, GenericClifford(1)), (0.2, CXGate())]
+
+    @data(
+        [(1, GenericClifford(2))],
+        [(0.5, GenericClifford(2)), (0.5, GenericClifford(3))],
+        [(0.2, GenericClifford(2)), (0.3, XGate()), (0.5, CCXGate())],
+    )
+    def test_invalid_edgegrab_distribution(self, distribution):
[…]
