### 2023-03-28 d04d99e
Release Notes 0.5 (#1057)
### Summary
- Aggregated and updated release notes for 0.5
- Added the test pulse backends and `CliffordUtils` to docs
- Regularized release notes to past tense ("Fixed error..." instead of
"Fixes error...")

---------

Co-authored-by: Yael Ben-Haim <yaelbh@il.ibm.com>
Co-authored-by: Will Shanks <willshanks@us.ibm.com>
