+++ releasenotes/notes/0.3/fix-composite-copy-a869e9773f6a4d48.yaml
@@ -1,10 +1,10 @@
 ---
 fixes:
   - |
-    Fix :meth:`.ParallelExperiment.copy` and :meth:`.BatchExperiment.copy`
+    Fixed :meth:`.ParallelExperiment.copy` and :meth:`.BatchExperiment.copy`
     so that the copies preserves any references between the original
     component experiments analysis classes and the :class:`.CompositeAnalysis`
     classes component analysis classes.
   - |
-    Fixes :meth:`.CompositeAnalysis.copy` to recursively make a copy of the
+    Fixed :meth:`.CompositeAnalysis.copy` to recursively make a copy of the
     component analysis classes.
