+++ releasenotes/notes/0.5/tomography-b091ce13d6983bc1.yaml
@@ -1,12 +1,12 @@
 ---
 features:
   - |
-    Adds ``backend``, ``analysis``, and ``target`` init kwargs to the
+    Added ``backend``, ``analysis``, and ``target`` init kwargs to the
     :class:`~.StateTomography` and :class:`~.ProcessTomography` experiments.
-    These allow specifying intended backend, a custom analysis class, or a
-    custom target for fidelity calculations, when initializing the experiments.
+    These allow specifying the intended backend, a custom analysis class, or a
+    custom target for fidelity calculations when initializing the experiments.
   - |
-    Improves :class:`.LocalMeasurementBasis` and `.LocalPreparationBasis`
+    Improved :class:`.LocalMeasurementBasis` and :class:`.LocalPreparationBasis`
     tomography basis classes support for initializing a noisy basis for
     performing state preparation and measurement error mitigated
     :class:`.StateTomography` and :class:`.ProcessTomography` experiments.
@@ -16,11 +16,11 @@ features:
     noisy prepared states are generated by applying the ideal instructions
     to the noisy initial state.
 
-    For measurement basis a noisy POVM or quantum channel can be supplied for
+    For measurement bases, a noisy POVM or quantum channel can be supplied for
     the 0-index basis (typically the Z-basis), and other bases index POVMs will
     be generated by applying the ideal inverse instructions to the noisy POVMS.
   - |
-    Adds an optional ``mitigator`` kwarg to :class:`.PauliMeasurementBasis`
+    Added an optional ``mitigator`` kwarg to :class:`.PauliMeasurementBasis`
     which can be used to initialize the basis with a
     :class:`~qiskit.result.LocalReadoutMitigator` to construct a readout error mitigated
     basis for use with :class:`.StateTomography` and
@@ -29,32 +29,32 @@ features:
     The :class:`.LocalReadoutError` experiment can be run to obtain the 
     :class:`~qiskit.result.LocalReadoutMitigator` from its analysis results.
   - |
-    Adds readout error mitigated tomography experiments
+    Added readout error mitigated tomography experiments
     :class:`.MitigatedStateTomography` and :class:`.MitigatedProcessTomography`.
     These are both implemented as a :class:`.BatchExperiment` consisting of a
-    :class:`~.LocalReadoutError` characterization experiment followed by either
-    a :class:`~.StateTomography` or class:`~.ProcessTomography` experiment.
+    :class:`.LocalReadoutError` characterization experiment followed by either
+    a :class:`.StateTomography` or :class:`.ProcessTomography` experiment.
   - |
-    Adds support for conditional tomographic reconstruction to the
+    Added support for conditional tomographic reconstruction to the
     :class:`.StateTomography` and :class:`.ProcessTomography` experiments.
 
     There are three types of conditioning that can be used independently
     or together in any combination for reconstruction of a state or
     channel as a list of components conditional on these values.
     
-    The `"conditional_circuit_clbits"` init option can be used to
+    The ``conditional_circuit_clbits`` init option can be used to
     specify any subset of clbits in an tomography circuit containing
     clbits to be conditioned on when peforming the tomographic
     reconstruction. The conditioning outcome value of the clbits is
     stored in the analysis results ``extra`` field.
 
-    The `"conditional_measurement_indices"` analysis option can be used
+    The ``conditional_measurement_indices`` analysis option can be used
     to condition on the measurement basis index and outcome value of a
     specific subset of tomographic basis measurements. The conditioning
[…]
