+++ releasenotes/notes/ecr_lib-381cb18885e81abd.yaml
@@ -1,9 +1,0 @@
----
-features:
-  - |
-    A new basis gate library called :class:`.EchoedCrossResonance` is been added.
-upgrade:
-  - |
-    The :class:`.Calibrations` class has been updated to use the reference
-    mechansim in Qiskit Pulse in which a schedule can refer to another
-    schedule only by name.
