+++ releasenotes/notes/T1_experiment_level_1_mesurment_analysis-078db79e3b0c16b9.yaml
@@ -1,6 +1,0 @@
----
-features:
-  - |
-    Added new class :class:`.T1KerneledAnalysis`. This class is used for T1
-    Experiment with the option `meas_level=MeasLevel.KERNELED`. The analysis
-    normalizes the data and fixes its orientation.
