+++ releasenotes/notes/fix-missing_calibration_updator_call-a255b28dd1449ea4.yaml
@@ -1,6 +1,0 @@
----
-
-fixes:
-  - |
-    Fixes a bug where :class:`.Calibrations` is not updated when you run any calibration 
-    experiment through composite experiment.
