+++ releasenotes/notes/0.5/calibration-backendv2-e564f466eb1c9999.yaml
@@ -4,7 +4,7 @@ fixes:
     Updated querying of :class:`~qiskit.providers.Backend` metadata to support
     the look up of qubit T1 and measurement drive frequency, in order to
     support :class:`~qiskit.providers.BackendV2` backends. The look up of the
-    latter is ``qiskit-ibm-provider`` specific. This change fixes errors
+    latter is ``qiskit-ibm-provider`` specific. This change fixed errors
     failing to find these properties when using :class:`.ResonatorSpectroscopy`
-    (issue `#1099 <https://github.com/Qiskit/qiskit-experiments/issues/1099>_`)
+    (issue `#1099 <https://github.com/Qiskit/qiskit-experiments/issues/1099>`_)
     and when using restless measurements with ``BackendV2`` backends.
