+++ releasenotes/notes/0.5/curve-analysis-4bcc10cf3a39a85d.yaml
@@ -1,23 +1,23 @@
 ---
 features:
   - |
-    New Curve Analysis options `average_method` has been added. This option
+    New :class:`.BaseCurveAnalysis` option ``average_method`` has been added. This option
     modifies an averaging technique for y values over the same x values.
-    Defaults to `sample` for the RB experiments and `shots_weighted` for the rest of analysis.
+    It defaults to ``sample`` for the RB experiments and ``shots_weighted`` for the rest of analysis.
     Previously the setup was hardcoded in the ``_format_data`` method of the analysis class,
     and no statistical difference has been introduced with introduction of this option.
 deprecations:
   - |
-    Providing data_sort_key directly to the LMFIT model to instantiate :class:`.CurveAnalysis` 
+    Providing ``data_sort_key`` directly to the LMFIT model to instantiate :class:`.CurveAnalysis` 
     has been deprecated. This option is not officially supported by the LMFIT,
     and thus curve analysis cannot guarantee this option is properly managed 
     in all LMFIT model subclasses.
 developer:
   - |
-    To map experiment result data to a particular LMFIT model in the :class:`.CurveAnalysis`,
-    an author must provide the data_subfit_map analysis option rather than directly binding 
-    data_sort_key with the target LMFIT model. 
-    The data_subfit_map option is a dictionary keyed on the model name. For example,
+    To map experiment result data to a particular LMFIT model in :class:`.CurveAnalysis`,
+    an author must provide the ``data_subfit_map`` analysis option rather than directly binding 
+    ``data_sort_key`` with the target LMFIT model. 
+    The ``data_subfit_map`` option is a dictionary keyed on the model name. For example,
     
     .. code-block:: python3
     
@@ -37,5 +37,5 @@ developer:
           options.data_subfit_map = {"expr1": {"tag": "1"}, "expr2": {"tag": "2"}}
           return options
       
-    As shown in above, the dictionary had been attached to each LMFIT model 
-    is now moved to the data_subfit_map option.
+    As shown in above, the dictionary that had been attached to each LMFIT model 
+    is now moved to the ``data_subfit_map`` option.
