+++ releasenotes/notes/0.5/backend-timing-bc05fd3cc7b41a45.yaml
@@ -1,6 +1,6 @@
 ---
 features:
   - |
-    Introduced a new class ``qiskit_experiments.framework.BackendTiming`` which
+    Introduced a new class :class:`qiskit_experiments.framework.BackendTiming`, which
     provides helper methods for rounding pulse and delay times to values
     compatible with a backend's timing constraints.
