+++ releasenotes/notes/0.5/tomo-barriers-aae4aafedaca5c3d.yaml
@@ -1,10 +1,10 @@
 ---
 fixes:
   - |
-    Fix qpy serialization and deserialization of tomography experiments. The
+    Fixed qpy serialization and deserialization of tomography experiments. The
     barrier instructions in tomography experiments were created with the wrong
     Python type which qpy did not support. This issue was most acute when using
-    `qiskit-ibm-provider` which submits circuits to the provider using qpy.
+    ``qiskit-ibm-provider`` which submits circuits to the provider using qpy.
     There could have been subtler issues with circuit timing using a different
     provider if the barriers were not separating important circuit
     instructions. See `#1060 <https://github.com/Qiskit/qiskit-experiments/issues/1060>`_.
