### 2023-03-28 be34082
Update docs/manuals/measurement/restless_measurements.rst
Co-authored-by: Will Shanks <wshaos@posteo.net>
