### 2023-03-28 d4d7419
Changed fake backends to V2 and added seeds for docs (#1108)
### Summary

Because visualization doc builds are continuing to throw warnings (see
#1107), this PR fixes the simulator seed until we can replicate this
error with a known seed and debug the problem. In addition, V1 simulator
backends in docs have been replaced with ``FakePerth``, which mocks a
currently available free backend.

### Details and comments

- Fake V2 backends are missing the `default_rep_delay` attribute, so the
code that uses this in `restless_measurement.rst` has been replaced with
a hardcoded value from the real back
[…]
