### 2023-03-28 0558984
drop 3.7 support and upgrade numpy to 1.22
also started using Literal types and updated tomography analysis
