+++ releasenotes/notes/drop_python_3_7_support-0529a7122e94b004.yaml
@@ -1,0 +1,4 @@
+---
+upgrade:
+  - |
+    Dropped support for Python 3.7 and upgraded NumPy to 1.22.
