+++ tools/deploy_documentation.sh
@@ -26,6 +26,14 @@ tox -edocs
 echo "show current dir: "
 pwd
 
+<<<<<<< HEAD
+=======
+CURRENT_TAG=`git describe --abbrev=0`
+IFS=. read -ra VERSION <<< "$CURRENT_TAG"
+STABLE_VERSION=${VERSION[0]}.${VERSION[1]}
+echo "Building for stable version $STABLE_VERSION"
+
+>>>>>>> 8fe22c9 (Fix IFS setting in doc deploy script (#1110))
 # Push to qiskit.org website
 openssl aes-256-cbc -K $encrypted_rclone_key -iv $encrypted_rclone_iv -in tools/rclone.conf.enc -out $RCLONE_CONFIG_PATH -d
 echo "Pushing built docs to website"
