### 2023-03-29 fdd16f7
remove pins and update test, rb
