+++ test/visualization/test_utils.py
@@ -42,14 +42,14 @@ class TestDataExtentCalculator(QiskitExperimentsTestCase):
         ]
 
         # Iterate over pairs of adjacent bin edges, which define the maximum and minimum for the region.
-        # This is done by zipping over shifted subarrays of bin_edges as follows:
+        # This is done by generating sliding windows of bin_edges as follows:
         #      [[a], [b], [c], [d], [e], [f]], g]
         #  [a, [[b], [c], [d], [e], [f], [g]]
         # The result is a list of pairs representing a moving window of size 2.
-        # TODO: Replace this with numpy.[...].sliding_window_view when Qiskit requires numpy>=0.20.0
+
         dummy_data = []
         for (x_min, x_max), (y_min, y_max) in it.product(
-            *tuple(list(zip(b[0:-1], b[1:])) for b in bin_edges)
+            *np.lib.stride_tricks.sliding_window_view(bin_edges, 2, 1)
         ):
             _dummy_data = np.asarray(
                 [
