+++ .mergify.yml
@@ -6,4 +6,4 @@ pull_request_rules:
     actions:
       backport:
         branches:
-          - stable/0.4
+          - stable/0.5
