### 2023-03-29 ed372c4
Bugfix in `_retrieve_data()` and test fix in `test_save()`
