### 2023-03-29 4cadedc
matplotlib 3.7 fails - MarkerStyle
this single change allows qiskit_experiments to work for matplotlib 3.7

```
venvs/s/581f024a/venv/lib/python3.9/site-packages/qiskit_experiments/__init__.py:68: in <module>
    from . import library
venvs/s/581f024a/venv/lib/python3.9/site-packages/qiskit_experiments/library/__init__.py:116: in <module>
    from .calibration import (
venvs/s/581f024a/venv/lib/python3.9/site-packages/qiskit_experiments/library/calibration/__init__.py:62: in <module>
    from .rough_frequency import RoughFrequencyCal
venvs/s/581f024a/venv/lib/python3.9/site-packages/qiskit_exp
[…]
