+++ qiskit_experiments/library/tomography/tomography_analysis.py
@@ -26,7 +26,7 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 from qiskit.quantum_info.operators.channel.quantum_channel import QuantumChannel
 
 from qiskit_experiments.exceptions import AnalysisError
-from qiskit_experiments.framework import BaseAnalysis, AnalysisResultData, Options
+from qiskit_experiments.framework import BaseAnalysis, AnalysisResultData, Options, numpy_version
 from .fitters import (
     tomography_fitter_data,
     postprocess_fitter,
@@ -326,7 +326,14 @@ class TomographyAnalysis(BaseAnalysis):
         prob_data = outcome_data / shot_data[None, :, None]
         bs_fidelities = []
         for _ in range(self.options.target_bootstrap_samples):
-            sampled_data = rng.multinomial(shot_data, prob_data)
+            # TODO: remove conditional once numpy is pinned at 1.22 and above
+            if numpy_version() >= (1, 22):
+                sampled_data = rng.multinomial(shot_data, prob_data)
+            else:
+                sampled_data = np.zeros_like(outcome_data)
+                for i in range(prob_data.shape[0]):
+                    for j in range(prob_data.shape[1]):
+                        sampled_data[i, j] = rng.multinomial(shot_data[j], prob_data[i, j])
 
             try:
                 state_results = self._fit_state_results(
