+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.22
+numpy>=1.17
 scipy>=1.4
 qiskit-terra>=0.22
 qiskit-ibm-experiment>=0.2.5
