### 2023-03-30 6d88ef3
remove 3.7 from cron tests
