### 2023-03-30 8b12c7c
Added a missing backend parameter to RoughFrequencyCal (#1115)
Fixes #1020

(cherry picked from commit 36003028929874f15d5338e84d44ddbe7a194fc7)
