+++ releasenotes/notes/backend-in-rough-frequency-cal-8582610249e2327b.yaml
@@ -1,0 +1,5 @@
+---
+fixes:
+  - |
+    Added a missing `backend` parameter to :class:`~.RoughFrequencyCal`.
+
