+++ qiskit_experiments/calibration_management/save_utils.py
@@ -1,0 +1,244 @@
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2021.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+"""A helper module to save calibration data to local storage.
+
+These model version must be upgraded when :class:`.Calibrations` data structure changes
+and current data model no longer supports the upgraded data structure.
+"""
+
+from dataclasses import dataclass, field, asdict
+from datetime import datetime
+from typing import List, Dict, Any
+
+from qiskit.pulse import ScheduleBlock
+
+from .calibrations import Calibrations
+from .control_channel_map import ControlChannelMap
+from .parameter_value import ParameterValue
+
+
+@dataclass(frozen=True)
+class ParameterModelV1:
+    """A data schema of a single calibrated parameter.
+
+    .. note::
+        This is intentionally agnostic to the data structure of
+        Qiskit Experiments Calibrations for portability.
+
+    """
+
+    param_name: str
+    qubits: List[int]
+    schedule: str = ""
+    value: float = 0.0
+    datetime: datetime = datetime.now()
+    valid: bool = True
+    exp_id: str = ""
+    group: str = ""
+
+    @classmethod
+    def apply_schema(cls, data: Dict[str, Any]):
+        """Consume dictionary and returns canonical data model."""
+        return ParameterModelV1(**data)
+
+
+@dataclass(frozen=True)
+class ScheduleModelV1:
+    """A data schema of a single templated schedule.
+
+    .. note::
+        This is intentionally agnostic to the data structure of
+        Qiskit Experiments Calibrations for portability.
+
+    """
+
+    data: ScheduleBlock
+    qubits: List[int] = field(default_factory=list)
+    num_qubits: int = 0
+
+    @classmethod
+    def apply_schema(cls, data: Dict[str, Any]):
+        """Consume dictionary and returns canonical data model."""
+        return ScheduleModelV1(**data)
+
+
+@dataclass
+class CalibrationModelV1:
+    """A data schema of whole device representation.
+
+    .. note::
+        This is intentionally agnostic to the data structure of
+        Qiskit Experiments Calibrations for portability.
+
+    """
+
+    backend_name: str
+    backend_version: str
+    device_coupling_graph: List[List[int]]
+    control_channel_map: ControlChannelMap
+    schedules: List[ScheduleModelV1] = field(default_factory=list)
+    parameters: List[ParameterModelV1] = field(default_factory=list)
+    schema_version: str = "1.0"
+
+    @classmethod
+    def apply_schema(cls, data: Dict[str, Any]):
+        """Consume dictionary and returns canonical data model."""
+        in_data = {}
+        for key, value in data.items():
+            if key == "schedules":
+                value = list(map(ScheduleModelV1.apply_schema, value))
+            if key == "parameters":
+                value = list(map(ParameterModelV1.apply_schema, value))
+            in_data[key] = value
+        return CalibrationModelV1(**in_data)
+
+
+def to_dict(
+    cals: Calibrations,
+    most_recent_only: bool = True,
+) -> Dict[str, Any]:
+    """A helper function to convert calibration data into dictionary.
+
+    Args:
+        cals: A calibration instance to save.
+        most_recent_only: Set True to save calibration parameters with most recent time stamps.
+
+    Returns:
+        Canonicalized calibration data in dictionary format.
+    """
+    # This can dispatch canonicalize function as version evolves.
[…]
