+++ test/calibration/test_calibrations.py
@@ -350,6 +350,65 @@ class TestCalibrationsBasic(QiskitExperimentsTestCase):
             backend.target.add_instruction(gate, properties=properties)
         Calibrations.from_backend(backend)
 
+    def test_equality(self):
+        """Test the equal method on calibrations."""
+        backend = FakeBelemV2()
+        library = FixedFrequencyTransmon(basis_gates=["sx", "x"])
+
+        cals1 = Calibrations.from_backend(
+            backend, libraries=[library], add_parameter_defaults=False
+        )
+        cals2 = Calibrations.from_backend(
+            backend, libraries=[library], add_parameter_defaults=False
+        )
+        self.assertTrue(cals1 == cals2)
+
+        date_time = datetime.now(timezone.utc).astimezone()
+        param_val = ParameterValue(0.12345, date_time=date_time)
+        cals1.add_parameter_value(param_val, "amp", 3, "x")
+
+        # The two objects are different due to missing parameter value
+        self.assertFalse(cals1 == cals2)
+
+        # The two objects are different due to time stamps
+        param_val2 = ParameterValue(0.12345, date_time=date_time - timedelta(seconds=1))
+        cals2.add_parameter_value(param_val2, "amp", 3, "x")
+        self.assertFalse(cals1 == cals2)
+
+        # The two objects are different due to missing parameter value
+        cals3 = Calibrations.from_backend(
+            backend, libraries=[library], add_parameter_defaults=False
+        )
+        self.assertFalse(cals1 == cals3)
+
+        # The two objects are identical due to time stamps
+        cals2.add_parameter_value(param_val, "amp", 3, "x")
+        self.assertFalse(cals1 == cals3)
+
+        # The schedules contained in the cals are different.
+        library2 = FixedFrequencyTransmon(basis_gates=["sx", "x", "y"])
+        cals1 = Calibrations.from_backend(backend, libraries=[library])
+        cals2 = Calibrations.from_backend(backend, libraries=[library2])
+        self.assertFalse(cals1 == cals2)
+
+        # Ensure that the equality is not sensitive to parameter adding order.
+        cals1 = Calibrations.from_backend(
+            backend, libraries=[library], add_parameter_defaults=False
+        )
+        cals2 = Calibrations.from_backend(
+            backend, libraries=[library], add_parameter_defaults=False
+        )
+        param_val1 = ParameterValue(0.54321, date_time=date_time)
+        param_val2 = ParameterValue(0.12345, date_time=date_time - timedelta(seconds=1))
+
+        cals1.add_parameter_value(param_val2, "amp", 3, "x")
+        cals1.add_parameter_value(param_val1, "amp", 3, "x")
+
+        cals2.add_parameter_value(param_val1, "amp", 3, "x")
+        cals2.add_parameter_value(param_val2, "amp", 3, "x")
+
+        self.assertTrue(cals1 == cals2)
+
 
 class TestOverrideDefaults(QiskitExperimentsTestCase):
     """
@@ -1442,14 +1501,15 @@ class TestSavingAndLoading(CrossResonanceTest):
         """Clean-up after the test."""
         super().tearDown()
 
-        for file in ["parameter_values.csv", "parameter_config.csv", "schedules.csv"]:
+        for file in ["parameter_values.csv", "parameter_config.csv", "schedules.csv", ".json"]:
             if os.path.exists(self._prefix + file):
                 os.remove(self._prefix + file)
 
     def test_save_load_parameter_values(self):
         """Test that we can save and load parameter values."""
 
-        self.cals.save("csv", overwrite=True, file_prefix=self._prefix)
+        with self.assertWarns(DeprecationWarning):
[…]
