### 2023-03-31 2dfb2e4
Update module docstring
