### 2023-03-31 a646a98
fix experiment thumbnail and added manual links
