+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 3.3.0
-envlist = py311,py310,py39,py38,py37,lint
+envlist = py311,py310,py39,py38,lint
 isolated_build = true
 
 [testenv]
