### 2023-04-01 9da7676
tox to nox
this commit switches current usages of tox to nox. the documentation isn't
updated pending results of CI runs.
