+++ .github/workflows/cron-staging.yml
@@ -35,12 +35,12 @@ jobs:
             ${{ runner.os }}-${{ matrix.python-version }}-pip-
             ${{ runner.os }}-${{ matrix.python-version }}
       - name: Install Deps
-        run: python -m pip install -U tox setuptools virtualenv wheel
+        run: python -m pip install -U nox setuptools virtualenv wheel
       - name: Install and Run Tests
-        run: tox -e terra-main
+        run: nox -e test-terra-main
         if: runner.os != 'macOS'
       - name: Install and Run Tests
-        run: tox -e terra-main
+        run: nox -e test-terra-main
         if: runner.os == 'macOS'
         env:
           OMP_NUM_THREADS: 1
