+++ .github/workflows/main.yml
@@ -40,12 +40,12 @@ jobs:
             ${{ runner.os }}-${{ matrix.python-version }}-pip-
             ${{ runner.os }}-${{ matrix.python-version }}
       - name: Install Deps
-        run: python -m pip install -U tox setuptools virtualenv wheel
+        run: python -m pip install -U nox setuptools virtualenv wheel
       - name: Install and Run Tests
-        run: tox -e py
+        run: nox -e test
         if: runner.os != 'macOS'
       - name: Install and Run Tests
-        run: tox -e py
+        run: nox -e test
         if: runner.os == 'macOS'
         env:
           OMP_NUM_THREADS: 1
@@ -69,9 +69,9 @@ jobs:
             ${{ runner.os }}-${{ matrix.python-version }}-pip-
             ${{ runner.os }}-${{ matrix.python-version }}-
       - name: Install Deps
-        run: python -m pip install -U tox
+        run: python -m pip install -U nox
       - name: Run lint
-        run: tox -elint
+        run: nox -e lint
   docs:
     name: docs
     runs-on: ubuntu-latest
@@ -94,10 +94,10 @@ jobs:
             ${{ runner.os }}-
       - name: Install Deps
         run: |
-          python -m pip install -U tox
+          python -m pip install -U nox
           sudo apt-get install -y pandoc graphviz
       - name: Build Docs
-        run: tox -edocs
+        run: nox -e docs
       - name: Compress Artifacts
         run: |
           mkdir artifacts
