+++ tox.ini
@@ -1,81 +1,0 @@
-[tox]
-minversion = 3.3.0
-envlist = py311,py310,py39,py38,py37,lint
-isolated_build = true
-
-[testenv]
-usedevelop = True
-install_command = pip install -U {opts} {packages}
-setenv =
-  VIRTUAL_ENV={envdir}
-  QISKIT_SUPPRESS_PACKAGING_WARNINGS=Y
-deps = -r{toxinidir}/requirements-dev.txt
-passenv =
-  OMP_NUM_THREADS
-  QISKIT_PARALLEL
-  RAYON_NUM_THREADS
-  QISKIT_IBM_*
-commands = stestr run {posargs}
-
-[testenv:terra-main]
-usedevelop = True
-install_command = pip install -U {opts} {packages}
-setenv =
-  VIRTUAL_ENV={envdir}
-  QISKIT_SUPPRESS_PACKAGING_WARNINGS=Y
-deps =
-  git+https://github.com/Qiskit/qiskit-terra
-  -r{toxinidir}/requirements-dev.txt
-passenv =
-  OMP_NUM_THREADS
-  QISKIT_PARALLEL
-  RAYON_NUM_THREADS
-  QISKIT_IBM_*
-commands = stestr run {posargs}
-
-
-[testenv:lint]
-envdir = .tox/lint
-commands =
-  black --check {posargs} qiskit_experiments test tools setup.py
-  pylint -rn -j 0 --rcfile={toxinidir}/.pylintrc qiskit_experiments/ test/ tools/
-  python {toxinidir}/tools/verify_headers.py
-
-[testenv:lint-incr]
-envdir = .tox/lint
-basepython = python3
-allowlist_externals = git
-commands =
-  black --check {posargs} qiskit_experiments test tools setup.py
-  -git fetch -q https://github.com/Qiskit/qiskit-experiments :lint_incr_latest
-  python {toxinidir}/tools/pylint_incr.py -rn -j4 -sn --paths :/qiskit_experiments/*.py :/test/*.py :/tools/*.py
-  python {toxinidir}/tools/verify_headers.py qiskit_experiments test tools
-
-[testenv:black]
-envdir = .tox/lint
-commands = black {posargs} qiskit_experiments test tools setup.py
-
-[testenv:docs]
-usedevelop = False
-passenv = EXPERIMENTS_DEV_DOCS
-commands =
-  sphinx-build -T -W --keep-going -b html {posargs} docs/ docs/_build/html
-
-[testenv:docs-parallel]
-usedevelop = False
-passenv = EXPERIMENTS_DEV_DOCS
-commands =
-  sphinx-build -j auto -T -W --keep-going -b html {posargs} docs/ docs/_build/html
-
-[testenv:docs-minimal]
-usedevelop = False
-passenv = EXPERIMENTS_DEV_DOCS
-setenv = 
-  QISKIT_DOCS_SKIP_EXECUTE = 1
-commands =
-  sphinx-build -T -W --keep-going -b html {posargs} docs/ docs/_build/html
-
-[pycodestyle]
-max-line-length = 100
-
-
