### 2023-04-01 ffe502a
fix test version selection
