+++ .github/workflows/main.yml
@@ -42,10 +42,10 @@ jobs:
       - name: Install Deps
         run: python -m pip install -U nox setuptools virtualenv wheel
       - name: Install and Run Tests
-        run: nox -e test
+        run: nox -e "test-${{ matrix.python-version }}"
         if: runner.os != 'macOS'
       - name: Install and Run Tests
-        run: nox -e test
+        run: nox -e "test-${{ matrix.python-version }}"
         if: runner.os == 'macOS'
         env:
           OMP_NUM_THREADS: 1
