### 2023-04-01 e506f8f
lint `docs/conf.py`
