### 2023-04-03 73fdca3
Update docs/manuals/characterization/stark_experiment.rst
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
