+++ test/calibration/test_setup_library.py
@@ -16,7 +16,7 @@ from typing import Dict, Set
 import json
 
 from test.base import QiskitExperimentsTestCase
-import qiskit.pulse as pulse
+from qiskit import pulse
 from numpy import pi
 
 from qiskit_experiments.calibration_management.basis_gate_library import FixedFrequencyTransmon
@@ -37,7 +37,7 @@ class TestLibrary(FixedFrequencyTransmon):
         with pulse.build(name="x") as schedule:
             pulse.play(pulse.Drag(160, 0.1, 40, 0), pulse.DriveChannel(0))
 
-        schedules = dict()
+        schedules = {}
         if "x" in basis_gates:
             schedules["x"] = schedule
 
