### 2023-04-03 bc4c3eb
minor style & doc changes
