### 2023-04-03 Qiskit/qiskit-experiments#1124 @nunezco2
Improving discussion of state of the field - paper
RE: `openjournals/joss-reviews#5329`

The current JOSS paper submission would benefit from a detailed discussion of the state of the field. In particular, answers to the following questions are needed:

- How does `qiskit_experiments` differ from other approaches documented in literature [3]?
- How does `qiskit_experiments` integrate existing quantum characterization and calibration methods, software packages or artifacts, particularly those mentioned in quantum control literature --e.g., [1,2]?
-  What makes the approach documented in the paper novel and profitable for experimentalists across the quantum computing stack in contrast to existing software?
- What are the specific tasks that become less onerous or possible through `qiskit_experiments`? How is the difference measured or characterized?
- Have comparisons been drawn between code written without the package and code that makes use of it?

References:

[1] Ball, H., Biercuk, M.J., Carvalho, A.R., Chen, J., Hush, M., De Castro, L.A., Li, L., Liebermann, P.J., Slatyer, H.J., Edmunds, C. and Frey, V., 2021. Software tools for quantum control: Improving quantum computer performance through noise and error suppression. Quantum Science and Technology, 6(4), p.044011.

[2] Landig, A. and Wallraff, A., 2013. Software for arbitrary single qubit and qutrit gate calibration (Doctoral dissertation, Master’s thesis, ETH Zurich).

[3] Pasquale, A., Efthymiou, S., Ramos-Calderer, S., Wilkens, J., Roth, I. and Carrazza, S., 2023. Towards an open-source framework to perform quantum calibration and characterization. arXiv preprint arXiv:2303.10397.
