### 2023-04-04 Qiskit/qiskit-experiments#1124 @danielskatz
part of openjournals/joss-reviews#5329
