### 2023-04-04 b358c78
change round in Rabi to avoid numpy int64 data type
