### 2023-04-05 Qiskit/qiskit-experiments#1129 @TejasAvinashShetty
Improvement of the paper  submitted to JoSS
This is regarding https://github.com/openjournals/joss-reviews/issues/5329.

Dear @eggerdj @gadial @nkanazawa1989 @wshanks @coruscating @chriseclectic , @yaelbh  etc.
Apologies if I have missed any of the authors.
Congratulations on a wonderful paper. 👏
Could you please add the code to generate the plot that you have shown in the example? It would really help a new prospective researcher to try out your software. I could reproduce the example at my end. But a novice researcher with less perseverance maybe tempted to give up.

Kudos to the extensive and comprehensive documentation of the software.

Yours Sincerely,
Dr Tejas Shetty
