### 2023-04-06 #5329 @eggerdj
Dear @TejasAvinashShetty ,

In response to https://github.com/openjournals/joss-reviews/issues/5329#issuecomment-1497114965 regarding the code of the QV experiment (comment https://github.com/Qiskit/qiskit-experiments/issues/1129). I would avoid including the code in the paper itself in the unlikely event of changes to the code. You can find an example of the QV experiment here: https://qiskit.org/documentation/experiments/manuals/verification/quantum_volume.html . What I think would make sense is to amend the paper with an extra sentence that points to the experiment manuals. We could also explicitly call out the quantum volume manual.
