+++ docs/paper/paper.md
@@ -131,6 +131,8 @@ The shaded area is a $2\sigma$ confidence interval.
 # Documentation
 
 `Qiskit Experiments` documentation is available at [https://qiskit.org/documentation/experiments](https://qiskit.org/documentation/experiments).
+The documentation also includes [experiment manuals](https://qiskit.org/documentation/experiments/manuals/index.html) 
+that show how to run experiments such as the Quantum Volume presented above.
 
 # Acknowledgements
 
