### 2023-04-07 8e5f868
Update docs/paper/paper.md
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
