### 2023-04-07 984d8ab
* Text streamline
