+++ docs/paper/paper.md
@@ -91,18 +91,18 @@ maintainers to easily execute characterization and calibration experiments.
 Recently, software packages have started to emerge to fill this gap [@Pasquale2023]. 
 `Qiskit Experiments` is unique in this perspective as it provides open-source low-level 
 characterization experiments that integrate with pulse-level control [@Alexander2020].
+`Qiskit Experiments` greatly simplifies the execution of complex experiments and is 
+usable with any hardware exposed as a Qiskit backend.
+Indeed, a library provides many experiments which run multiple quantum circuits and complex fitting. 
+Crucially, each experiment only requires a few code lines to run with `Qiskit Experiments`.
+In addition, the base framework of `Qiskit Experiments` provides experimentalists a clear 
+interface to create new experiments. 
+They must (i) define how to construct the circuits, (ii) define the experiment options, 
+and optionally (iii) implement the analysis class, if not already present in the library.
 In addition, `Qiskit Experiments` provides a calibration framework to manage device calibration.
-This framework is usable with any hardware exposed as a Qiskit backend. 
-`Qiskit Experiments` greatly simplifies the execution of complex experiments. 
-Indeed, the experiments in the library run multiple quantum circuits and complex fitting 
-but only require a few code lines to run.
-In addition, the base framework of `Qiskit Experiments` provides experimentalists a clear interface
-to create new experiments. 
-They must (i) define how circuits are constructed, (ii) define the experiment
-options, and optionally (iii) implement the analysis class, if not already present in the library.
 Experiments in the `Qiskit Experiments` library and custom extensions built on top of the 
-framework have been used to explore measurements without qubit
-reset [@Tornow2022], benchmarking [@Amico2023], characterize positive operator value measures [@Fischer2022], quantum 
+framework have been used to explore measurements without qubit reset [@Tornow2022], 
+benchmarking [@Amico2023], positive operator value measures [@Fischer2022], quantum 
 states [@Hamilton2022], and time-evolutions [@Greenaway2022], as well as calibrate gates [@Vazquez2022].
 
 ![
