### 2023-04-07 6304812
Paper review (#1131)
### Summary

Small paper edits.

### Details and comments

This PR addresses
https://github.com/Qiskit/qiskit-experiments/issues/1124 and
https://github.com/Qiskit/qiskit-experiments/issues/1129. We have

* included two new references 
* outlined the value proposition of Qiskit Experiments
* included a link to the experiment manuals.

---------

Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
