+++ docs/paper/paper.bib
@@ -23,6 +23,20 @@
   copyright = {arXiv.org perpetual, non-exclusive license}
 }
 
+@article{Ball2021,
+  doi = {10.1088/2058-9565/abdca6},
+  url = {https://dx.doi.org/10.1088/2058-9565/abdca6},
+  year = {2021},
+  month = {sep},
+  publisher = {IOP Publishing},
+  volume = {6},
+  number = {4},
+  pages = {044011},
+  author = {H. Ball and M. J. Biercuk and A. R. R. Carvalho and J. C. and M. Hush and L. A. De Castro and L. Li and P. J Liebermann and H. J. Slatyer and C. Edmunds and V. Frey and C. Hempel and A. Milne},
+  title = {Software tools for quantum control: improving quantum computer performance through noise and error suppression},
+  journal = {Quantum Sci. Technol.}
+}
+
 @article{Cross2019,
   title = {Validating quantum computers using randomized model circuits},
   author = {Cross, A. W. and Bishop, L. S. and Sheldon, S. and Nation, P. D. and Gambetta, J. M.},
@@ -121,6 +135,15 @@
   issn = {1476-4687},
   doi = {10.1038/s41586-020-2649-2},
   url = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@misc{Pasquale2023,
+  title = {Towards an open-source framework to perform quantum calibration and characterization},
+  author = {A. Pasquale and S. Efthymiou and S. Ramos-Calderer and J. Wilkens and I. Roth and S. Carrazza},
+  year = {2023},
+  eprint = {2303.10397},
+  archivePrefix = {arXiv},
+  primaryClass = {quant-ph}
 }
 
 @misc{Qiskit,
