### 2023-04-07 #5329 @eggerdj
@danielskatz This PR https://github.com/Qiskit/qiskit-experiments/pull/1131 adds a link to the experiment manuals in the docs. It also addresses https://github.com/Qiskit/qiskit-experiments/issues/1124
