### 2023-04-07 Qiskit/qiskit-experiments#1124 @eggerdj
Dear @nunezco2 thank you for the review. We added citations to Ball et al. and Pasquale as you suggested to better reflect the state of the field. Thank you for pointing this out. In addition in PR #1131 we made changes to address the points that you raise.
* We better point out the gain for experimentalists. I.e., (i) complex experiments in the library are easy to run, (ii) Qiskit Experiments provides a  base framework on which one can build new experiments, and (iii) Qiskit Experiments provides a framework for calibration.
* The experiment manuals in the documentation which we now link to in the paper point out how easy it is to run a complex experiment such as Quantum Volume.
