### 2023-04-08 #5329 @danielskatz
> @danielskatz Can the reviewers discuss with each other or is it against JoSS rules? I mean only converse on Github and on this thread only.

Open discussion in this issue is fine
