### 2023-04-10 6d53507
Add dataframe support for extended equality
