### 2023-04-10 049436a
Remove `sphinx-autodoc-typehints` and minor doc fixes (#1121)
### Summary

This PR follows https://github.com/Qiskit/qiskit-terra/pull/9705 to
remove the dependency on `sphinx-autodoc-typehints` and
https://github.com/Qiskit/qiskit-metapackage/issues/1707 to use the
default Qiskit icon when the experiment manual isn't explicitly
provided.

### Details and comments

In addition, some experiment classes that didn't link to their manuals
have been updated with the link. Manual line breaks added after numbered
lists have been removed now that
https://github.com/Qiskit/qiskit_sphinx_theme/pull/218 
[…]
