+++ qiskit_experiments/library/randomized_benchmarking/interleaved_rb_experiment.py
@@ -48,6 +48,9 @@ class InterleavedRB(StandardRB):
     # section: analysis_ref
         :class:`InterleavedRBAnalysis`
 
+    # section: manual
+        :doc:`/manuals/verification/randomized_benchmarking`
+
     # section: reference
         .. ref_arxiv:: 1 1203.4550
 
