### 2023-04-11 Qiskit/qiskit-experiments#1136 @goerz
JOSS review: Reproducibility of tutorial
I'm following the [Getting Started](https://qiskit.org/documentation/experiments/tutorials/getting_started.html) tutorial from scratch, and it complains about `qiskit_aer` in the second code box of [Running your first experiment](https://qiskit.org/documentation/experiments/tutorials/getting_started.html#running-your-first-experiment).

You should either make `qiskit-aer` a dependency of `qiskit-experiments`, or put instructions on which packages the reader should install with `pip` at the beginning of the tutorial. Maybe just point to the `requirements-dev.txt` in the repo.

Later on, in the [Run options](https://qiskit.org/documentation/experiments/tutorials/getting_started.html#run-options) section, there's `'MeasLevel' is not defined`. That one is not as obvious to fix: where is `MeasLevel` imported from?

It looks to me like you're generating the output in the tutorials automatically during CI (as you should!), so I'm actually a little mystified how this can happen.

Can you review your continuous integration setup for making sure that the tutorials are fully reproducible?

Reference: https://github.com/openjournals/joss-reviews/issues/5329
