### 2023-04-11 Qiskit/qiskit-experiments#1136 @goerz
Ok, seems what's missing there is

```
from qiskit.qobj.utils import MeasLevel
```

But I'm still confused why the CI didn't catch this
