### 2023-04-11 #5329 @goerz
Ok, I finished my review. Overall, the software package and its documentation are of very high quality, and I have no fundamental objections to publication.

I came across two minor issues that should be fixed:

First, there are some small technical issues with the reproducibility of the tutorial, cf. https://github.com/Qiskit/qiskit-experiments/issues/1136. I've left the "Installation instructions" checkmark open until the issue is closed.

Second, the "Example usage" in the paper could be improved. I somewhat agree with @TejasAvinashShetty's [comment](https://github.com/openjournals/joss-reviews/issues/5329#issuecomment-1497114965). I think JOSS' *intent* for the "Example usage" is to actually show some code, and the first three code blocks of [the online QV examples](https://qiskit.org/documentation/experiments/manuals/verification/quantum_volume.html) wouldn't be much to add here. As for @eggerdj's [comment](https://github.com/openjournals/joss-reviews/issues/5329#issuecomment-1498888042) that the example code could change in future versions, you could just say "using `v0.5` of `qiskit-experiments`". That being said, everybody is probably just going to look at the documentation, not the JOSS paper, to get a usage example, so I'm not sure if I care that much ;-)

Without the code, though, the section in the paper needs some editing. The first two sentences in particular sound like they lead up to a code example. You could rephrase this as, e.g.,

> `Qiskit Experiments` can run, for example, a Quantum Volume (QV) measurement. Evaluating the QV of a quantum backend requires executing random SU(4) circuits on a noisy simulator of the backend to quantify the largest quantum circuit with equal width and depth that can be successfully run. A depth d QV circuit is successful if it has …. `Qiskit Experiments` only requires a few lines of code to run this standardized yet complex experiment, see the online documentation [url]. The analysis classes of existing experiments … Figure 2."

I've left the "Quality of writing" checkmark open for this. Please either add the code for the example or rephrase the text.
