+++ docs/manuals/characterization/tphi.rst
@@ -1,6 +1,10 @@
 Tφ Characterization
 ===================
 
+.. note::
+    This manual requires the ``qiskit-aer`` package to run simulations.
+    You can install it with ``python -m pip install qiskit-aer``.
+
 :math:`T_\varphi`, or :math:`1/\Gamma_\varphi`, is the pure dephasing time of
 depolarization in the :math:`x - y` plane of the Bloch sphere. We compute
 :math:`\Gamma_\varphi` by computing :math:`\Gamma_2`, the *transverse relaxation rate*,
@@ -19,6 +23,10 @@ superconducting qubits, :math:`T_2^*` tends to be significantly smaller than
 
 From the :math:`T_1` and :math:`T_2` estimates, we compute the results for
 :math:`T_\varphi.`
+
+.. note::
+    This manual requires the ``qiskit-aer`` package to run simulations.
+    You can install it with ``python -m pip install qiskit-aer``.
 
 .. jupyter-execute::
 
