### 2023-04-12 3c03b43
remove duplicate note
