### 2023-04-12 03de5bb
update exp option example code
