### 2023-04-12 #5329 @eggerdj
Dear @goerz , thank you for the review. I have amended the paper according to your suggestion here: https://github.com/Qiskit/qiskit-experiments/pull/1138
