### 2023-04-12 241a422
Apply suggestions from code review
Co-authored-by: Daniel J. Egger <38065505+eggerdj@users.noreply.github.com>
