### 2023-04-12 d3df73a
Add Aer install instructions to documentation (#1137)
### Summary

This PR addresses #1136 by adding Aer install instructions on tutorials
and experiment manuals that use it. The rerun analysis how-to
instructions for serializing an experiment has also been updated with
the experiment encoder and decoder.

### Details and comments

Some additional changes include:

- Simplified the simulator used in the getting started tutorial since we
want to focus on experiments
- Added an introduction for the QST manual
- Added a note to the manuals page to clarify that there are many more
experiments in t
[…]
