+++ docs/manuals/characterization/t1.rst
@@ -29,6 +29,10 @@ which we deduce from the probability estimates.
 The following code demonstrates a basic run of a :math:`T_1` experiment
 for qubit 0.
 
+.. note::
+    This manual requires the ``qiskit-aer`` package to run simulations.
+    You can install it with ``python -m pip install qiskit-aer``.
+
 .. jupyter-execute::
 
     import numpy as np
