### 2023-04-12 76aa6e3
Update docs/paper/paper.md
Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
