+++ docs/paper/paper.md
@@ -120,7 +120,7 @@ the largest quantum circuit with equal width and depth that can be successfully 
 A depth $d$ QV circuit is successful if it has mean heavy-output probability greater 
 than two-thirds with a confidence level exceeding 0.977, and at least 100 trials have been run.
 `Qiskit Experiments` only requires a few lines of code to run this standardized yet complex experiment,
-see the [online documentation](https://qiskit.org/documentation/experiments/manuals/verification/quantum_volume.html).
+as shown in the [online documentation](https://qiskit.org/documentation/experiments/manuals/verification/quantum_volume.html).
 The analysis classes of existing experiments automatically generate key figures with customizable 
 visualization options, as exemplified by the QV plot in \autoref{fig:qv}.
 
