### 2023-04-12 #5329 @eggerdj
@danielskatz and @goerz I believe the two following PRs should address the points raised by Michael. Thanks again for pointing this out!
* Tutorials: https://github.com/Qiskit/qiskit-experiments/pull/1137
* Paper: https://github.com/Qiskit/qiskit-experiments/pull/1138
