### 2023-04-12 028ab85
Review comments
- Cleanup doc
- Add handling for edge case

Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>
