### 2023-04-12 Qiskit/qiskit-experiments#1136 @danielskatz
Should this be closed now?
