### 2023-04-12 b4b519b
revert project string change
will put this in a separate PR since the corresponding theme change
isn't released yet
