### 2023-04-13 f96568d
Add strict_type and numerical_precision option
