+++ test/extended_equality.py
@@ -9,6 +9,9 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
+
+# pylint: disable=unused-argument
+
 """
 Utility for checking equality of data of Qiskit Experiments class which doesn't
 officially implement the equality dunder method.
@@ -36,10 +39,54 @@ from qiskit_experiments.framework import (
 from qiskit_experiments.visualization import BaseDrawer
 
 
-@multimethod
 def is_equivalent(
+    data1: Any,
+    data2: Any,
+    *,
+    strict_type: bool = True,
+    numerical_precision: float = 1e-8,
+) -> bool:
+    """Check if two input data are equivalent.
+
+    This function is used for custom equivalence evaluation only for unittest purpose.
+    Some third party class may not preserve equivalence after JSON round-trip with
+    Qiskit Experiments JSON Encoder/Decoder, or some Qiskit Experiments class doesn't
+    define the equality dunder method intentionally.
+
+    Args:
+        data1: First data to compare.
+        data2: Second data to compare.
+        strict_type: Set True to enforce type check before comparison. Note that serialization
+            and deserialization round-trip may not preserve data type.
+            If the data type doesn't matter and only behavioral equivalence is considered,
+            e.g. iterator with the same element; tuple vs list,
+            you can turn off this flag to relax the constraint for data type.
+        numerical_precision: Tolerance of difference between two real numbers.
+
+    Returns:
+        True when two objects are equivalent.
+    """
+    if strict_type and type(data1) is not type(data2):
+        return False
+    evaluated = _is_equivalent_dispatcher(
+        data1,
+        data2,
+        strict_type=strict_type,
+        numerical_precision=numerical_precision,
+    )
+    if not isinstance(evaluated, (bool, np.bool_)):
+        # When either one of input is numpy array type, it may broadcast equality check
+        # and return ndarray of dtype=bool. e.g. np.array([]) == 123
+        # The input values should not be equal in this case.
+        return False
+    return evaluated
+
+
+@multimethod
+def _is_equivalent_dispatcher(
     data1: object,
     data2: object,
+    **kwargs,
 ):
     """Equality check finally falls into this function."""
     if data1 is None and data2 is None:
@@ -50,33 +97,28 @@ def is_equivalent(
         return is_equivalent(
             data1.__dict__,
             data2.__dict__,
+            **kwargs,
         )
-    evaluated = data1 == data2
-    if not isinstance(evaluated, bool):
-        # When either one of input is numpy array type, it may broadcast equality check
-        # and return ndarray of dtype=bool. e.g. np.array([]) == 123
-        # The input values should not be equal in this case.
-        return False
-
-    # Return the outcome of native equivalence check.
-    return evaluated
+    return data1 == data2
 
 
-@is_equivalent.register
+@_is_equivalent_dispatcher.register
 def _check_dicts(
     data1: Union[dict, ThreadSafeOrderedDict],
     data2: Union[dict, ThreadSafeOrderedDict],
+    **kwargs,
 ):
     """Check equality of dictionary which may involve Qiskit Experiments classes."""
     if set(data1) != set(data2):
         return False
-    return all(is_equivalent(data1[k], data2[k]) for k in data1.keys())
[…]
