### 2023-04-14 f6f498d
added RoughEFFrequencyCal to library
