+++ releasenotes/notes/backend-in-rough-frequency-cal-8582610249e2327b.yaml
@@ -1,5 +1,6 @@
 ---
 fixes:
   - |
-    Added a missing `backend` parameter to :class:`~.RoughFrequencyCal`.
+    Added a missing `backend` parameter to :class:`~.RoughEFFrequencyCal` and
+    exposed it in the experiment library.
 
