### 2023-04-16 dfba3e3
updated discriminator and added test
