+++ qiskit_experiments/library/characterization/multi_state_discrimination.py
@@ -143,7 +143,7 @@ class MultiStateDiscrimination(BaseExperiment):
                         )
 
             # label the circuit
-            circuit.metadata = {"label": level}
+            circuit.metadata = {"label": str(level)}
 
             circuit.measure_all()
             circuits.append(circuit)
