### 2023-04-16 7987a6a
Update releasenotes/notes/adjust-symbolic-pulses-amp-angle-representation-f5c40007416cf938.yaml
Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>
