### 2023-04-16 ed7970b
Update releasenotes/notes/adjust-symbolic-pulses-amp-angle-representation-f5c40007416cf938.yaml
Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>
