### 2023-04-17 993fcb8
Review comments
- Docs update
- Remove schema for schedule and move qubit info to metadata
- Change function names

Co-authored-by: Daniel Egger <egger.d.j@gmail.com>
