+++ qiskit_experiments/calibration_management/save_utils.py
@@ -14,13 +14,28 @@
 
 .. warning::
 
-    This module is expected to be internal and no stable user-facing API
-    may not be guaranteed.
+    This module is expected to be internal and is not intended as a stable user-facing API.
 
 .. note::
 
-    These model version must be upgraded when :class:`.Calibrations` data structure changes
-    in future and when the current data model becames unable to support it.
+    Because a locally saved :class:`.Calibrations` instance may not conform to the
+    data model of the latest Qiskit Experiments implementation, the calibration loader
+    must be aware of the data model version.
+    CalibrationModel classes representing the data model must have
+    the version suffix and the `schema_version` field.
+    This helps the loader to raise user-friendly error rather than being crashed by
+    incompatible data, and possibly to dispatch the loader function based on the version number.
+
+    When the :class:`.Calibrations` class is refactored and starts to use new data model,
+    a developer also needs to define new CalibrationModel with new version number,
+    rather than modifying the existing CalibrationModel class to preserve backward compatibility.
+
+
+.. note::
+
+    We don't guarantee the portability of stored data across different Qiskit Experiments
+    versions. We allow the calibration loader to raise an error for non-supported
+    data models.
 
 """
 
@@ -46,38 +61,33 @@ class ParameterModelV1:
     """
 
     param_name: str
+    """Name of the parameter."""
+
     qubits: List[int]
+    """List of associated qubits."""
+
     schedule: str = ""
+    """Associated schedule name."""
+
     value: float = 0.0
-    datetime: datetime = datetime.now()
+    """Parameter value."""
+
+    datetime: datetime = None
+    """A date time at which this value is obtained."""
+
     valid: bool = True
+    """If this parameter is valid."""
+
     exp_id: str = ""
+    """Associated experiment ID which is used to obtain this value."""
+
     group: str = ""
+    """Name of calibration group in which this calibration parameter belongs to."""
 
     @classmethod
     def apply_schema(cls, data: Dict[str, Any]):
         """Consume dictionary and returns canonical data model."""
         return ParameterModelV1(**data)
-
-
-@dataclass(frozen=True)
-class ScheduleModelV1:
-    """A data schema of a single templated schedule.
-
-    .. note::
-        This is intentionally agnostic to the data structure of
-        Qiskit Experiments Calibrations for portability.
-
-    """
-
-    data: ScheduleBlock
-    qubits: List[int] = field(default_factory=list)
-    num_qubits: int = 0
-
-    @classmethod
-    def apply_schema(cls, data: Dict[str, Any]):
-        """Consume dictionary and returns canonical data model."""
-        return ScheduleModelV1(**data)
 
 
 @dataclass
@@ -91,27 +101,38 @@ class CalibrationModelV1:
     """
 
     backend_name: str
+    """Name of the backend."""
+
     backend_version: str
+    """Version of the backend."""
+
     device_coupling_graph: List[List[int]]
+    """Qubit coupling graph of the device."""
+
     control_channel_map: ControlChannelMap
-    schedules: List[ScheduleModelV1] = field(default_factory=list)
+    """Mapping of ControlChannel to qubit index."""
+
+    schedules: List[ScheduleBlock] = field(default_factory=list)
+    """Template schedules. It must contain the metadata for qubits and num_qubits."""
+
     parameters: List[ParameterModelV1] = field(default_factory=list)
+    """List of calibrated pulse parameters."""
+
     schema_version: str = "1.0"
+    """Version of this data model. This must be static."""
 
     @classmethod
     def apply_schema(cls, data: Dict[str, Any]):
         """Consume dictionary and returns canonical data model."""
         in_data = {}
[…]
