### 2023-04-17 90f3877
Use marginal_distribution() instead of marginal_counts()
This commit updates the Qiskit marginalization function used in
qiskit-experiments. Qiskit has 2 different function available for
marginalizing counts: `marginal_counts()` [1] which is designed to work
with `Counts` or `Result` objects and does **not** respect ordering of
the indices passed in. The other is `marginal_distribution()` [2] which is
designed to work with `Counts`, `Probabilities`, `QuasiProbabilities`, or
a dictionary with string keys and integer or float values that respects
the order of the indices passed in (so it can be 
[…]
