+++ qiskit_experiments/framework/composite/composite_analysis.py
@@ -15,7 +15,7 @@ Composite Experiment Analysis class.
 
 from typing import List, Dict, Union, Optional, Tuple
 import numpy as np
-from qiskit.result import marginal_counts
+from qiskit.result import marginal_distribution
 from qiskit.result.postprocess import format_counts_memory
 from qiskit_experiments.framework import BaseAnalysis, ExperimentData
 from qiskit_experiments.framework.analysis_result_data import AnalysisResultData
@@ -206,7 +206,7 @@ class CompositeAnalysis(BaseAnalysis):
                 sub_data = {"metadata": metadata["composite_metadata"][i]}
                 if "counts" in datum:
                     if composite_clbits is not None:
-                        sub_data["counts"] = marginal_counts(datum["counts"], composite_clbits[i])
+                        sub_data["counts"] = marginal_distribution(datum["counts"], composite_clbits[i])
                     else:
                         sub_data["counts"] = datum["counts"]
                 if "memory" in datum:
