### 2023-04-17 ad9f2d2
Temporarily convert np.int64 to int until Qiskit/qiskit-terra#9976 is released
