+++ releasenotes/notes/adjust-symbolic-pulses-amp-angle-representation-f5c40007416cf938.yaml
@@ -6,15 +6,22 @@ other:
     from complex amplitude representation to a duo of real (float) parameters
     (amp, angle). To avoid problems, Qiskit Experiments adopted this convention.
     
-    The :class:`.FixedFrequencyTransmon` library now includes an ``angle`` parameter
-    for every pulse, and the default values of ``amp`` are set to float instead
-    of complex.
-    
-    With the exception of :class:`.HalfAngleCal`, all experiments should continue
-    function as they did before (even with complex ``amp``).
-    ``HalfAngleCal`` was changed from updating
-    the complex amplitude of the pulse, to updating the angle. When used with
-    the ``FixedFrequencyTransmon`` library, it will continue to work seamlessly 
+    Changes were made to :class:`.FixedFrequencyTransmon` and :class:`.HalfAngleCal`
+    (see upgrade section). With the exception of :class:`.HalfAngleCal`, all 
+    library experiments should continue function as they did before (even with 
+    complex ``amp``). When used with :class:`.FixedFrequencyTransmon` library, 
+    :class:`.HalfAngleCal` will also continue working as before.
+upgrade:
+  - |
+    The representation of pulses in the :class:`.FixedFrequencyTransmon` library 
+    was changed from complex amplitude to (amp,angle) representation. All pulses
+    now include an ``angle`` parameter, and the default values of ``amp`` are set
+    to float instead
+
+  - |
+    :class:`.HalfAngleCal`, was changed from updating the complex amplitude of 
+    the pulse, to updating the angle in the (amp, angle) representation. When used with
+    the :class:`.FixedFrequencyTransmon` library, it will continue to work seamlessly 
     in the new representation. However, when the experiment is used with custom
     built pulses, which rely on the old convention of complex ``amp`` (with no
     angle parameter) - the experiment will fail. Most reasonable cases will raise
@@ -22,4 +29,5 @@ other:
     to it. Some edge cases, like a custom built pulse with an ``angle`` parameter
     which doesn't conform to the naming convention of Qiskit Terra's 
     ``ScalableSymbolicPulse`` class - will result in the updating of the wrong
-    parameter. 
+    parameter.
+    
