### 2023-04-17 8b900b7
Update qiskit_experiments/curve_analysis/visualization/mpl_drawer.py
Co-authored-by: Ian Hincks <ian.hincks@gmail.com>
