+++ releasenotes/notes/0.5/experiment_data_fixes-f69c3569a8ba1342.yaml
@@ -8,6 +8,8 @@ upgrade:
     setting of the experiment database service via passing the provider.
     3. The `end_datetime` property of :class:`.ExperimentData` is now being
     set to the latest time a successful job terminated.
+    4. :meth:`.ExperimentData.save` can now raise exceptions when saving
+    fails, unless the `suppress_errors` flag is set (on by default)
 fixes:
   - |
    1. Fixed a bug in :meth:`.ExperimentData._add_job_data` where the wrong
