### 2023-04-18 ccbd512
Fixed calibration experiments failing on backends with no coupling map (backport #1117) (#1148)
This is an automatic backport of pull request #1117 done by
[Mergify](https://mergify.com).


Co-authored-by: Will Shanks <willshanks@us.ibm.com>
