+++ releasenotes/notes/cals-no-coupling-map-5114ae9faa2f9e69.yaml
@@ -1,0 +1,6 @@
+---
+fixes:
+  - |
+    Fixed error generating circuits for :class:`.BaseCalibrationExperiment`
+    subclasses when the backend instance had no coupling map. Fixed `#1116
+    <https://github.com/Qiskit/qiskit-experiments/issues/1116>`_.
