### 2023-04-18 d12fb7b
Remove current release from previous releases list (#1140)
### Summary

The previous releases section on the documentation sidebar includes the
current version (0.5), which is unnecessary. The link is also broken
right now (will be fixed in
https://github.com/Qiskit/qiskit_sphinx_theme/pull/267), which is
another motivation to remove this from the current docs. This PR changes
the list so the current version won't be included.
