### 2023-04-18 6a12b1b
Fix `_retrieve_data` exit condition
