### 2023-04-18 3f5127b
Drop Python 3.7 support (#1112)
### Summary

Python 3.7 is nearing its end of life and is causing multiple problems
with package incompatibilities. This PR drops support for 3.7 and
implements newer NumPy code depending on the NumPy version.

### Details and comments

- Added `numpy_version()` for checking NumPy version
- Start to convert variables that can only be one of a few values to
Literal type (added in Python 3.8)
- Minor docstring fixes
