### 2023-04-18 4eb87f9
Authors affiliations (#1146)
### Summary

In response to
https://github.com/openjournals/joss-reviews/issues/5329#issuecomment-1505449868

### Details and comments

Minor changes to the affiliations and names.

---------

Co-authored-by: Helena Zhang <hadron.diaspora@gmail.com>
