### 2023-04-18 #5329 @eggerdj
Author and affiliation check done: https://github.com/Qiskit/qiskit-experiments/pull/1146
