### 2023-04-18 2a08924
Update release notes and documentation (backport #1147) (#1151)
This is an automatic backport of pull request #1147 done by
[Mergify](https://mergify.com).


---


Co-authored-by: Helena Zhang <Helena.Zhang@ibm.com>
