+++ releasenotes/notes/0.5/0_5_1_release-e445b6cc64742cc0.yaml
@@ -1,11 +1,3 @@
 ---
-fixes:
-  - |
-    Fixed a deprecated Matplotlib ``MarkerStyle`` usage in the visualization module that was causing
-    warnings in Matplotlib 3.6+.
-  - |
-    Added a missing `backend` parameter to :class:`~.RoughFrequencyCal`.
-  - |
-    Fixed error generating circuits for :class:`.BaseCalibrationExperiment`
-    subclasses when the backend instance had no coupling map. Fixed `#1116
-    <https://github.com/Qiskit/qiskit-experiments/issues/1116>`_.
+prelude: >
+  Qiskit Experiments 0.5.1 is a minor bug fix release.  
