### 2023-04-18 147db6f
Update qiskit_experiments/library/calibration/half_angle_cal.py
Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>
