### 2023-04-18 5e1e3e7
Update releasenotes/notes/adjust-symbolic-pulses-amp-angle-representation-f5c40007416cf938.yaml
Co-authored-by: Naoki Kanazawa <nkanazawa1989@gmail.com>
