+++ docs/conf.py
@@ -38,7 +38,7 @@ os.environ["QISKIT_DOCS"] = "TRUE"
 # The short X.Y version
 version = "0.5"
 # The full version, including alpha/beta/rc tags
-release = "0.5.0"
+release = "0.5.1"
 project = f"Qiskit Experiments {version}"
 copyright = f"2021-{datetime.date.today().year}, Qiskit Development Team"  # pylint: disable=redefined-builtin
 author = "Qiskit Development Team"
