+++ releasenotes/notes/0.5/0_5_1_release-e445b6cc64742cc0.yaml
@@ -1,0 +1,3 @@
+---
+prelude: >
+  Qiskit Experiments 0.5.1 is a minor bug fix release.  
