### 2023-04-19 bdfa010
Moved the setting of end_datetime to the correct location
