+++ docs/paper/paper.bib
@@ -164,7 +164,7 @@
 }
 
 @article{Magesan2011,
-  title = {Scalable and Robust {Randomized} {Benchmarking} of Quantum Processes},
+  title = {Scalable and Robust Randomized Benchmarking of Quantum Processes},
   author = {Magesan, E. and Gambetta, J. M. and Emerson, J.},
   journal = {Phys. Rev. Lett.},
   volume = {106},
@@ -191,7 +191,7 @@
 @Article{Hunter2007,
   Author    = {Hunter, J. D.},
   Title     = {Matplotlib: A 2D graphics environment},
-  Journal   = {Comput. Sci. Engineer.},
+  Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
   Pages     = {90--95},
@@ -210,7 +210,7 @@
  }
 
 @misc{Lebigot2016,
-  title = {{Uncertainties}: a Python package for calculations with uncertainties},
+  title = {{Uncertainties}: a {P}ython package for calculations with uncertainties},
   author = {Lebigot, E. O.},
   url = {https://pythonhosted.org/uncertainties/},
 }
@@ -233,6 +233,7 @@
   title={A Practical Quantum Instruction Set Architecture},
   author={Smith, R. S. and Curtis, M. J. and Zeng, W. J.},
   journal={arXiv preprint arXiv:1608.03355},
+  doi={10.48550/arXiv.1608.03355},
   year={2016}
 }
 
