+++ docs/manuals/verification/quantum_volume.rst
@@ -12,9 +12,8 @@ Textbook <https://learn.qiskit.org/course/quantum-hardware/measuring-quantum-vol
 for an explanation on the QV method, which is described in Refs. [1]_ [2]_.
 
 The Quantum Volume is determined by the largest successful circuit depth
-:math:`d_{max}`, and equals to :math:`2^{d_{max}}`. In the QV experiment
-we generate `QV
-circuits <https://qiskit.org/documentation/stubs/qiskit.circuit.library.QuantumVolume.html>`__
+:math:`d_{max}`, and equals to :math:`2^{d_{max}}`. In this experiment,
+we generate QV circuits using the :class:`qiskit.circuit.library.QuantumVolume` class
 on :math:`d` qubits, which contain :math:`d` layers, where each layer
 consists of random 2-qubit unitary gates from :math:`SU(4)`, followed by
 a random permutation on the :math:`d` qubit. Then these circuits run on
