+++ docs/manuals/verification/quantum_volume.rst
@@ -25,7 +25,7 @@ probability’ > 2/3 with confidence level > 0.977 (corresponding to
 z_value = 2), and at least 100 trials have been ran.
 
 .. note::
-    This manual requires the ``qiskit-aer`` package to run simulations.
+    This manual requires the :mod:`qiskit_aer` package to run simulations.
     You can install it with ``python -m pip install qiskit-aer``.
 
 .. jupyter-execute::
