### 2023-04-19 0d49f3e
Fix docs deployment and update links (#1153)
### Summary

This PR fixes the development docs deployment, which was uploading to
the wrong link. It also removes extraneous docs template files and
updates links to the new ecosystem URL.

### Details and comments

- Updated version logic in `conf.py` to use the defined release variable
instead of `git describe`. This allows for easier testing without having
to wait for releases
- Added intersphinx links to Aer
- Removed deployment to `documentation/` now that the redirect to
`ecosystem/` is live
