+++ qiskit_experiments/library/randomized_benchmarking/interleaved_rb_experiment.py
@@ -53,7 +53,6 @@ class InterleavedRB(StandardRB):
 
     """
 
-    @deprecate_arguments({"qubits": "physical_qubits"}, "0.5")
     def __init__(
         self,
         interleaved_element: Union[QuantumCircuit, Gate, Delay, Clifford],
