+++ qiskit_experiments/library/characterization/cr_hamiltonian.py
@@ -136,7 +136,6 @@ class CrossResonanceHamiltonian(BaseExperiment):
         def __init__(self, width: ParameterValueType):
             super().__init__("cr_gate", 2, [width])
 
-    @deprecate_arguments({"qubits": "physical_qubits"}, "0.5")
     def __init__(
         self,
         physical_qubits: Tuple[int, int],
