+++ qiskit_experiments/library/quantum_volume/qv_experiment.py
@@ -71,7 +71,6 @@ class QuantumVolume(BaseExperiment):
 
     """
 
-    @deprecate_arguments({"qubits": "physical_qubits"}, "0.5")
     def __init__(
         self,
         physical_qubits: Sequence[int],
