+++ qiskit_experiments/library/tomography/qpt_experiment.py
@@ -49,14 +49,6 @@ class ProcessTomography(TomographyExperiment):
 
     """
 
-    @deprecate_arguments(
-        {
-            "qubits": "physical_qubits",
-            "measurement_qubits": "measurement_indices",
-            "preparation_qubits": "preparation_indices",
-        },
-        "0.5",
-    )
     def __init__(
         self,
         circuit: Union[QuantumCircuit, Instruction, BaseOperator],
