### 2023-04-19 #5329 @danielskatz
Another proof, for final check
